% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R, R/NMRData1D.R
\docType{methods}
\name{acqus-set}
\alias{acqus-set}
\alias{acqus<-,NMRData-method}
\alias{acqus<-,NMRData1D-method}
\title{Replace the "acqus" slot of an NMRData object}
\usage{
acqus(object) <- value

\S4method{acqus}{NMRData}(object) <- value

\S4method{acqus}{NMRData1D}(object) <- value
}
\arguments{
\item{object}{An NMRData1D or NMRData2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "acqus" slot of an NMRData
object.
}
