% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R, R/NMRData1D.R
\docType{methods}
\name{acqus}
\alias{acqus}
\alias{acqus,NMRData-method}
\alias{acqus,NMRData1D-method}
\title{Access the "acqus" slot of an NRMData object}
\usage{
acqus(object, ...)

\S4method{acqus}{NMRData}(object)

\S4method{acqus}{NMRData1D}(object)
}
\arguments{
\item{object}{An NMRData1D or NMRData2d object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "acqus" slot of an NMRData
object.
}
