% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R, R/NMRData1D.R, R/NMRData2D.R
\name{add_baseline}
\alias{add_baseline}
\alias{add_baseline,NMRData1D-method}
\alias{add_baseline,NMRData2D-method}
\title{Add baseline to data}
\usage{
add_baseline(object, ...)

\S4method{add_baseline}{NMRData1D}(object, baseline, knots = c())

\S4method{add_baseline}{NMRData2D}(object, baseline, knots = c())
}
\arguments{
\item{object}{An NMRData object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{baseline}{Vector of baseline spline parameters, where the domain of the
baseline must match the domain of the underlying data, i.e.,
a cmplx1 or complex vector is needed for 1D data and a cmplx2
vector is needed for 2D data. If a numeric vector is
provided, the same baseline is applied to all domains. Note
that for 2D data, the baseline components represent both
dimensions on an even grid and must therefore have a length
that is a perfect square. These values themselves represent
slices of baseline along the direct dimension in order of
increasing indirect dimension chemical shift.}

\item{knots}{A vector of relative internal knots ranging from 0
(corresponding to lower direct.shift) and 1 (corresponding to
upper direct.shift). Boundary knots are added automatically and
should not be included.}
}
\value{
A new NMRData object with modified intensity values.
}
\description{
Baseline terms are defined in the same way as for NMRFit1D/NMRFit2D, so it
should be possible to apply these directly.
}
