% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R, R/NMRData1D.R, R/NMRData2D.R
\name{add_phase}
\alias{add_phase}
\alias{add_phase,NMRData1D-method}
\alias{add_phase,NMRData2D-method}
\title{Add phase to data}
\usage{
add_phase(object, ...)

add_noise(object, ...)

\S4method{add_phase}{NMRData1D}(object, phase, degrees = FALSE)

\S4method{add_phase}{NMRData2D}(object, phase, degrees = FALSE)
}
\arguments{
\item{object}{An NMRData object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{phase}{A vector of phase correction terms -- 0 order, 1st order in the
direct dimension, 1st order in the indirect.dimension.}

\item{degrees}{TRUE if phase terms are provided in degrees, FALSE by default.}

\item{sd}{Standard deviation of Gaussian noise as fraction of maximum
intensity (in the real domain).}
}
\value{
A new NMRData object with modified intensity values.

A new NMRData object with modified intensity values.
}
\description{
Manual phase correction meant to apply phase terms calculated through
NMRFit. Note that phase terms are specified in radians unless "degrees" is
set to TRUE.

Helper function to add noise for simulations. At this point, noise is
assumed to be independent of chemical shift and Gaussian, but these
assumptions will hopefully be revisited at a later date.
}
