% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\docType{methods}
\name{all_columns}
\alias{all_columns}
\alias{.all_columns,NMRScaffold1D-method}
\title{Output all columns of "peaks" data.frame}
\usage{
.all_columns(object, index = FALSE, peak.type = NULL, ...)

\S4method{.all_columns}{NMRScaffold1D}(object, index = FALSE,
  peak.type = NULL)
}
\arguments{
\item{object}{NMRScaffold1D or NMRScaffold2D object.}

\item{index}{TRUE to output column indeces, FALSE to output names}

\item{peak.type}{One of either 'lorenz', 'gauss', 'pvoigt', or 'voigt'}

\item{...}{Additional arguments passed to inheriting methods.}
}
\value{
A vector of integers or characters.
}
\description{
Internal helper function used to generalize peak processing.
}
