% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{apodize_exponential}
\alias{apodize_exponential}
\title{Apply exponential apodization function}
\usage{
apodize_exponential(direct.time, signal, lb = 1,
  output.weights = FALSE)
}
\arguments{
\item{direct.time}{A vector of time points of the signal data.}

\item{signal}{A vector of complex signal data.}

\item{lb}{Line broadening factor (in Hz), where the apodization function
is calculated as exp(-lb*pi*t).}

\item{output.weights}{Set to TRUE to output the apodization function rather
than the weighted signal.}
}
\value{
Corrected signal vector (or vector of weights if output.weights
        is TRUE).
}
\description{
Apodizes signal data according to an exponential function.
}
