% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{apodize_gaussian}
\alias{apodize_gaussian}
\title{Apply Gaussian apodization function}
\usage{
apodize_gaussian(direct.time, signal, lb = 1, gb = 0.5,
  output.weights = FALSE)
}
\arguments{
\item{direct.time}{A vector of time points of the signal data.}

\item{signal}{A vector of complex signal data.}

\item{lb}{Line broadening factor (in Hz) component of the apodization (must
be negative to conform with Topspin parameters).}

\item{gb}{Gaussian broadening factor (0 - 1), corresponding to the position of
the function maximum as a fraction of acquisition time.}

\item{output.weights}{Set to TRUE to output the apodization function rather
than the weighted signal.}
}
\value{
Corrected signal vector (or vector of weights if output.weights
        is TRUE).
}
\description{
Apodizes signal data according to a Gaussian function.
}
