% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{apodize_signal}
\alias{apodize_signal}
\title{Apply apodization function}
\usage{
apodize_signal(direct.time, fid, method = "exponential", ...,
  output.weights = FALSE)
}
\arguments{
\item{direct.time}{A vector of time points of the signal data.}

\item{method}{One of either 'exponential', 'gaussian', 'sine', 'sine2',
or 'trapezoid' (where sine2 stands for sine squared -- qsin).}

\item{...}{Shape parameters that depend on method,
lb -- Line broadening (in Hz) used for exponential and
      gaussian apodization (negative for gaussian)
gb -- Gaussian broadening factor (0 - 1)
ssb -- Sine shift bell used for sine and sine2 apodization
tm1 -- Leftmost edge of trapezoid (0 - 1)
tm2 -- Rightmost edge of trapezoid (0 - 1)}

\item{output.weights}{Set to TRUE to output the apodization function rather
than the weighted signal.}

\item{signal}{A vector of complex signal data.}
}
\value{
Corrected signal vector (or vector of weights if output.weights
        is TRUE).
}
\description{
A thin wrapper around a number of common apodization functions.
}
