% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{apodize_sine}
\alias{apodize_sine}
\title{Apply sine apodization function}
\usage{
apodize_sine(direct.time, signal, ssb = 1, output.weights = FALSE)
}
\arguments{
\item{direct.time}{A vector of time points of the signal data.}

\item{signal}{A vector of complex signal data.}

\item{ssb}{Sine bell position parameter -- 1 for pure sine, 2 for pure cosine.}

\item{output.weights}{Set to TRUE to output the apodization function rather
than the weighted signal.}
}
\value{
Corrected signal vector (or vector of weights if output.weights
        is TRUE).
}
\description{
Apodizes signal data according to a sine function.
}
