% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{apodize_trapezoid}
\alias{apodize_trapezoid}
\title{Apply trapezoid apodization function}
\usage{
apodize_trapezoid(direct.time, signal, tm1 = 0, tm2 = 0.5,
  output.weights = FALSE)
}
\arguments{
\item{direct.time}{A vector of time points of the signal data.}

\item{signal}{A vector of complex signal data.}

\item{tm1}{The leftmost edge of the trapezoid as a fraction of 
acquisition time.}

\item{tm2}{The rightmost edge of the trapezoid as a fraction of 
acquisition time.}

\item{output.weights}{Set to TRUE to output the apodization function rather
than the weighted signal.}
}
\value{
Corrected signal vector (or vector of weights if output.weights
        is TRUE).
}
\description{
Apodizes signal data according to a trapezoid function that ramps up from
0 to tm1 and ramps down from tm2 until the end of the acquisition time, where
tm1 and tm2 are relative values from 0 to 1.
}
