% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R, R/NMRData1D.R
\docType{methods}
\name{apply_phase}
\alias{apply_phase}
\alias{apply_phase,NMRData1D-method}
\title{Correct spectrum phase}
\usage{
apply_phase(object, ...)

\S4method{apply_phase}{NMRData1D}(object, phase, degrees = TRUE)
}
\arguments{
\item{object}{An NMRData object.}

\item{phase}{A vector of phase corrections, from zero-order and up. 1st order
correction for NMRData1D objects require a vector of length 2 (0
order term, then first order term) while NMRData2D objects
require a vector of length 3 (0 order term, first order in the
direct dimension, 1st order in the indirect dimension).}

\item{degrees}{TRUE if the phase angles are in degrees, FALSE if in radians.}
}
\value{
NMRData object with corrected phase.
}
\description{
This function applies phase correction to NMRData1D and NMRData2D objects.
The nature of the phase correction vector depends on the object. The 0 order
phase correction term is always taken as the phase correction at 0 ppm.
}
