% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRScaffold2D.R
\docType{methods}
\name{areas}
\alias{areas}
\alias{areas,NMRScaffold1D-method}
\alias{areas,NMRScaffold2D-method}
\title{Calculate peak areas}
\usage{
areas(object, ...)

\S4method{areas}{NMRScaffold1D}(object, sf = nmroptions$direct$sf,
  sum.peaks = TRUE, include.id = FALSE, components = "r/i")

\S4method{areas}{NMRScaffold2D}(object, sf = nmroptions$direct$sf,
  sum.peaks = TRUE, include.id = FALSE, components = "r/i")
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{sf}{Sweep frequency (in MHz) -- needed to convert peak widths from Hz
to ppm. In most cases, it is recommended to set a single default
value using nmroptions$direct$sf, but an override can be
provided here.}

\item{sum.peaks}{TRUE to add all individual peaks together and output a
single area, FALSE to output a data frame of peak area
values.}

\item{include.id}{TRUE to include id as outer column if outputting data
frame.}
}
\value{
A single overall area or a data frame of areas with columns
        "resonance" (optional), "peak", and "area".
}
\description{
Calculate total peak areas based on peak parameters.
}
