% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R, R/NMRData1D.R
\docType{methods}
\name{as.data.frame.NMRData}
\alias{as.data.frame.NMRData}
\alias{as.data.frame,NMRData1D-method}
\title{Convert NMRData object to data.frame}
\usage{
as.data.frame.NMRData(x)

\S4method{as.data.frame}{NMRData1D}(x)
}
\arguments{
\item{x}{NMRData object.}
}
\description{
Outputs the processed data.frame, dropping acqus and procs lists.
}
