% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\name{as.data.frame.NMRIntensities}
\alias{as.data.frame.NMRIntensities}
\title{Convert NMRIntensities object into a full data.frame}
\usage{
as.data.frame.NMRIntensities(x)
}
\arguments{
\item{x}{NMRIntensities object.}
}
\value{
A data.frame with columns labelled as real/imaginary components.
}
\description{
Since an NMRIntensities object is essentially a data.frame broken down by
rows so that it can used as a column inside other data.frames, it's
sometimes convenient to access the data as a proper data.frame object.
}
