% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R, R/NMRData1D.R
\docType{methods}
\name{as.list.NMRData}
\alias{as.list.NMRData}
\alias{as.list,NMRData1D-method}
\title{Convert NMRData object to list}
\usage{
as.list.NMRData(x)

\S4method{as.list}{NMRData1D}(x)
}
\arguments{
\item{x}{NMRData object.}
}
\description{
Outputs all slot values as entries in a list.
}
