% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRResonance1D.R
\docType{methods}
\name{baseline-set}
\alias{baseline-set}
\alias{baseline<-,NMRScaffold1D-method}
\title{Replace the "baseline" of an NMRScaffold object}
\usage{
baseline(object) <- value

\S4method{baseline}{NMRScaffold1D}(object) <- value

baseline(object) <- value

\S4method{baseline}{NMRScaffold1D}(object) <- value
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{value}{Replacement value.}

\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "baseline" of an NMRScaffold
object.

Generic convenience method to replace the "baseline" of an NMRScaffold
object.
}
