% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRResonance1D.R
\docType{methods}
\name{baseline_difference}
\alias{baseline_difference}
\alias{baseline_difference,NMRScaffold1D-method}
\title{Access the "baseline_difference" of an NRMScaffold object}
\usage{
baseline_difference(object, ...)

\S4method{baseline_difference}{NMRScaffold1D}(object)

baseline_difference(object, ...)

\S4method{baseline_difference}{NMRScaffold1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "baseline_difference" of an NMRScaffold
object.

Generic convenience method to access the "baseline_difference" of an NMRScaffold
object.
}
