% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\docType{methods}
\name{bounds-set}
\alias{bounds-set}
\alias{bounds<-,NMRResonance1D-method}
\title{Set object bounds}
\usage{
bounds(object) <- value

\S4method{bounds}{NMRResonance1D}(object) <- value
}
\arguments{
\item{object}{An NMRResonance1D object.}

\item{value}{A list with "lower" and "upper" elements, each containing a data
frame with "peak", "position", "width", "height", and
"fraction.gauss" columns.}
}
\description{
Generic convenience method to set the bounds of an NMRResonance1D object.
}
