% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRResonance1D.R,
%   R/NMRSpecies1D.R, R/NMRMixture1D.R, R/NMRFit1D.R, R/NMRSpecies2D.R
\name{bounds}
\alias{bounds}
\alias{bounds,NMRResonance1D-method}
\alias{bounds,NMRSpecies1D-method}
\alias{bounds,NMRMixture1D-method}
\alias{bounds,NMRFit1D-method}
\alias{bounds,NMRSpecies2D-method}
\title{Get object bounds}
\usage{
bounds(object, ...)

\S4method{bounds}{NMRResonance1D}(object, include.id = FALSE)

\S4method{bounds}{NMRSpecies1D}(object, include.id = FALSE)

\S4method{bounds}{NMRMixture1D}(object, include.id = FALSE)

\S4method{bounds}{NMRFit1D}(object)

\S4method{bounds}{NMRSpecies2D}(object, include.id = FALSE)
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{include.id}{TRUE to return a column of resonance or species ids.}
}
\description{
Generic convenience method to access the bounds of any NMRScaffold1D object.
If used on an anything more complicated than an NMRResonance1D, the function
combines the bounds data frames of component resonances.
}
