% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\docType{methods}
\name{calc_baseline}
\alias{calc_baseline}
\alias{calc_baseline,NMRScaffold1D-method}
\title{Calculate baseline}
\usage{
calc_baseline(object, ...)

\S4method{calc_baseline}{NMRScaffold1D}(object, direct.shift = NULL)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\value{
A vector of NMR intensity values corresponding to the baseline
        estimated in the fit.
}
\description{
Calculate baseline values at desired chemical shift.
}
\details{
TO DO.
}
