% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\name{$.NMRIntensity}
\alias{$.NMRIntensity}
\title{Get a specific real/imaginary component}
\usage{
\method{$}{NMRIntensity}(...)
}
\arguments{
\item{object}{NMRIntensity data.}

\item{name}{A character name for the component. This can be an exact match
such as "ri" or a partial match such as "r.".}
}
\value{
A numeric vector if only a single component is matched, a complex
        vector if two components are matched, and an NMRIntensity object if
        three or more components are matched.
}
\description{
Specify the components of interest using a series of "r" and "i" characters,
such as "ri" or "ir". Internally, the name is passed into grep() so "r.",
for example, will be matched as both "rr" and "ri".
}
