% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R
\name{check_conformity}
\alias{check_conformity}
\title{Check data conformity}
\usage{
check_conformity(object, ...)
}
\arguments{
\item{object}{An NMRScaffold object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{nmrdata}{An NMRData object.}

\item{error}{True to issue an error message, otherwise, return message as
string.}
}
\value{
TRUE if data conforms, FALSE otherwise.
}
\description{
This is intended to a primarily internal function used to check whether
provided data conforms to the scaffold. The check is currently limited to
ensuring the correct number of dimensions and that data covers the chemical
shift of specific positions
}
