% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\docType{methods}
\name{check_data}
\alias{check_data}
\alias{.check_data,NMRScaffold-method}
\alias{.check_data,NMRScaffold1D-method}
\title{Check NMRData availability}
\usage{
.check_data(object, nmrdata, ...)

\S4method{.check_data}{NMRScaffold}(object, nmrdata)

\S4method{.check_data}{NMRScaffold1D}(object, nmrdata)
}
\arguments{
\item{object}{NMRScaffold1D or NMRScaffold2D object.}

\item{nmrdata}{NMRData1D, NMRData2D, or NULL}

\item{...}{Additional arguments passed to inheriting methods.}
}
\value{
NMRData1D or NMRData2D object.
}
\description{
Validates NMRData input and issues an error if NMRData can not be found
as either a slot or an argument.
}
