% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{check_syntax}
\alias{check_syntax}
\title{A convenience function for coupling syntax error messages.}
\usage{
check_syntax(invalid.syntax, syntax, hint)
}
\description{
A convenience function for coupling syntax error messages.
}
