% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R
\name{combine_all}
\alias{combine_all}
\alias{combine_all,NMRScaffold-method}
\title{Combine all components of object}
\usage{
combine_all(object, ...)

\S4method{combine_all}{NMRScaffold}(object, getter, slot.name, ...)
}
\arguments{
\item{object}{An NMRScaffold object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{getter}{Getter function.}

\item{slot.name}{Slot name as character.}
}
\description{
This is an internal function used for all getter functions that output a
data.frame object -- stitching together the results of all components.
}
