% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R
\name{combine_childrens}
\alias{combine_childrens}
\title{Combine children if they exist}
\usage{
combine_children(object, ...)
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{getter}{Getter function.}

\item{slot.name}{Slot name as character.}
}
\description{
This is an internal function used for all getter functions that output a
data.frame object. If the object is a species or mixture that has children,
the getter is applied to all children, a children column is added and the
resulting objects are stitched together.
}
