% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R
\name{combine_dimensions}
\alias{combine_dimensions}
\alias{combine_dimensions,NMRScaffold2D-method}
\title{Combine direct and indirect dimensions}
\usage{
combine_dimensions(object, ...)

\S4method{combine_dimensions}{NMRScaffold2D}(object, getter, ...)
}
\arguments{
\item{object}{An NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{getter}{Getter function.}
}
\description{
This is an internal function used for all getter functions that output a
data.frame object. Essentially, the getter is passed on to the direct and
indirect components, a dimension column is added and the resulting objects
are stitched together.
}
