% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R
\docType{methods}
\name{combine_getter}
\alias{combine_getter}
\alias{combine_getter,NMRScaffold2D-method}
\title{Combine direct and indirect dimensions}
\usage{
combine_getter(object, ...)

\S4method{combine_getter}{NMRScaffold2D}(object, getter, ...)
}
\arguments{
\item{object}{An NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
This is an internal function used for all getter functions that output a
data.frame object. Essentially, the getter is passed on to the direct and
indirect components, a dimension column is added and the resulting objects
are stitched together.
}
