% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R
\name{components}
\alias{components}
\alias{components,NMRScaffold-method}
\title{Unpack NMR components}
\usage{
components(object, ...)

\S4method{components}{NMRScaffold}(object, level = "resonance")
}
\arguments{
\item{object}{An NMRScaffold object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{level}{One of "resonance", "species", or "mixture".}
}
\value{
A tibble with varying numbers of descriptor columns and a single
        "component" column holding NMRScaffold objects.
}
\description{
This function enables access to lower-level NMR objects (such as
NMRResonance1D) which are contained within higher-level ones (such as
NMRFit1D). Note that splitting up resonances within species drops area
constraints.
}
