% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R,
%   R/NMRResonance.R, R/NMRResonance1D.R
\docType{methods}
\name{constraints-set}
\alias{constraints-set}
\alias{constraints<-,NMRScaffold-method}
\alias{constraints<-,NMRScaffold1D-method}
\title{Replace the "constraints" of an NMRScaffold object}
\usage{
constraints(object) <- value

\S4method{constraints}{NMRScaffold}(object) <- value

\S4method{constraints}{NMRScaffold1D}(object) <- value

constraints(object) <- value

\S4method{constraints}{NMRScaffold}(object) <- value

\S4method{constraints}{NMRScaffold1D}(object) <- value
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{value}{Replacement value.}

\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "constraints" of an NMRScaffold
object.

Generic convenience method to replace the "constraints" of an NMRScaffold
object.
}
