% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{constraints}
\alias{constraints}
\alias{constraints,NMRScaffold-method}
\alias{constraints,NMRScaffold1D-method}
\alias{constraints,NMRFit1D-method}
\title{Access the "constraints" slot of an NRMScaffold object}
\usage{
constraints(object, ...)

\S4method{constraints}{NMRScaffold}(object)

\S4method{constraints}{NMRScaffold1D}(object)

\S4method{constraints}{NMRFit1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "constraints" slot of an NMRScaffold
object.
}
