% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{corr_antiecho}
\alias{corr_antiecho}
\title{Decode echo-antiecho signals}
\usage{
corr_antiecho(indirect.time, direct.time, signal)
}
\arguments{
\item{indirect.time}{A vector of delay times for the signal.}

\item{direct.time}{A vector of sampling times for the signal.}

\item{signal}{A vector of complex fid data.}
}
\value{
Corrected fid data.
}
\description{
Decodes echo-antiecho pulses by converting them into real/imaginary
sequence pairs.
}
