% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\docType{methods}
\name{couplings-set}
\alias{couplings-set}
\alias{couplings<-,NMRResonance1D-method}
\title{Set object couplings}
\usage{
couplings(object) <- value

\S4method{couplings}{NMRResonance1D}(object) <- value
}
\arguments{
\item{object}{An NMRResonance1D object.}

\item{value}{A data frame with "id.1", "id.2", "position.difference", and
"area.ratio" columns.}
}
\description{
Generic convenience method to set the coupling definitions of an
NMRResonance1D object.
}
