% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R
\name{couplings}
\alias{couplings}
\alias{couplings,NMRScaffold-method}
\title{Get object couplings}
\usage{
couplings(object, ...)

\S4method{couplings}{NMRScaffold}(object, include.id = FALSE)
}
\arguments{
\item{object}{An NMRScaffold object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{include.id}{TRUE to return a column of resonance or species ids.}
}
\description{
Generic convenience method to access the coupling definitions of any
NMRScaffold object. If used on anything more complicated than an
NMRResonance1D, the function combines the couplings data frames of component
resonances.
}
