% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R
\name{direct-set}
\alias{direct-set}
\title{Set direct dimension resonance object}
\usage{
direct(object) <- value
}
\arguments{
\item{object}{An NMRScaffold2D object.}

\item{value}{A valid NMRScaffold1D object.}
}
\description{
Generic convenience method to set the direct dimension resonance component
of an NMRScaffold2D object.
}
