% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance2D.R
\docType{methods}
\name{direct}
\alias{direct}
\alias{direct,NMRResonance2D-method}
\title{Get direct dimension resonance object}
\usage{
direct(object, ...)

\S4method{direct}{NMRResonance2D}(object)
}
\arguments{
\item{object}{An NMRResonance2D or NMRSpecies2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the direct dimension resonance
component of an NMRResonance2D or NMRSpecies2D object.
}
