% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmplxn.R
\name{domain}
\alias{domain}
\alias{domain.vctrs_cmplx1}
\alias{domain.vctrs_cmplx2}
\title{Flexible domain selection}
\usage{
domain(x, ...)

\method{domain}{vctrs_cmplx1}(x, domain = "r/i", use.cmplx1 = FALSE)

\method{domain}{vctrs_cmplx2}(x, domain = "rr/ri/ir/ii", use.cmplx1 = FALSE)
}
\arguments{
\item{x}{cmplx1 or cmplx2 object.}

\item{domain}{A character specifying which real/imaginary domains to output
for intensity data. For 1D data "r/i" outputs both real and
imaginary data, with "r" and "i" used to output one of either
real or imaginary. For 2D data, this is expanded to
"rr/ri/ir/ii". All domains are outputted by default. Warning:
this function can be used to flip real and imaginary data by
providing i/r as opposed to r/i.}

\item{use.cmplx1}{cmplx1 is the internal representation of complex() that
matches cmplx2 for 2D data. It is replaced by default to
complex() for user-facing functions. However, this behavior
can be overriden here.}
}
\value{
If one domain selected, a simple numeric vector. If two selected, a
        complex() object with real and imaginary term replaced by the
        selected variants. If 3 or 4 selected, a cmplx2 object.
}
\description{
Flexible domain selection

Flexible domain selection
}
