% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R
\name{.filter_shift}
\alias{.filter_shift}
\title{Internal filter functions used for both direct and indirect shifts}
\usage{
.filter_shift(x, lower, upper, round, align)
}
\description{
Internal filter functions used for both direct and indirect shifts
}
