% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\docType{methods}
\name{drop_bounds}
\alias{drop_bounds}
\alias{.drop_bounds,NMRScaffold-method}
\alias{.drop_bounds,NMRScaffold1D-method}
\title{Drop bounds if they don't conform to parent object}
\usage{
.drop_bounds(object, ...)

\S4method{.drop_bounds}{NMRScaffold}(object)

\S4method{.drop_bounds}{NMRScaffold1D}(object, f, ...)
}
\arguments{
\item{object}{NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\value{
Modified NMRScaffold object.
}
\description{
If changes made to a parent NMRScaffold object make current bounds
no longer applicable, they must be dropped.
}
