% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\docType{methods}
\name{f_baseline}
\alias{f_baseline}
\alias{f_baseline,NMRFit1D-method}
\title{Generate baseline function}
\usage{
f_baseline(object, components = "r/i", ...)

\S4method{f_baseline}{NMRFit1D}(object, components = "r/i")
}
\arguments{
\item{object}{An NMRFit1D object.}

\item{components}{'r/i' to output both real and imaginary data, 'r' to output
only real and 'i' to output only imaginary.}
}
\value{
A function that outputs spectral intensity data of the fit baseline
        given a vector input of chemical shifts.
}
\description{
This is primarily an internal method that outputs a function that outputs
spectral intensity data of the fit baseline given a vector input of chemical
shifts.
}
