% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRScaffold2D.R
\docType{methods}
\name{f_lineshape}
\alias{f_lineshape}
\alias{f_lineshape,NMRScaffold1D-method}
\alias{f_lineshape,NMRScaffold2D-method}
\title{Generate lineshape function}
\usage{
f_lineshape(object, ...)

\S4method{f_lineshape}{NMRScaffold1D}(object, sf = nmroptions$direct$sf,
  sum.peaks = TRUE, include.id = FALSE, components = "r/i")

\S4method{f_lineshape}{NMRScaffold2D}(object,
  sf = c(nmroptions$direct$sf, nmroptions$indirect$sf),
  sum.peaks = TRUE, include.id = FALSE, components = "rr/ii")
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2d object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{sf}{Sweep frequency (in MHz) -- needed to convert peak widths from Hz
to ppm. A single value can be used for 1D and homonuclear 2D
experiments or two values for the direct and indirect dimensions,
in that order. this can be  In most cases, it is recommended to set
a single default value using, for example, nmroptions$direct$sf =
..., but an override can be provided here.}

\item{sum.peaks}{TRUE to add all individual peaks together and output a
single function, FALSE to output a data frame of functions
that correspond to individual peaks.}

\item{include.id}{TRUE to include id outer column if outputting data frame.}

\item{components}{A string specifying the real/imaginary components to
output, e.g. 'r/i' or 'rr/ri/ir/ii'. Use the symbols 'r',
'i' for 1D experiments and 'rr', 'ri', 'ir', 'ii' for 2D
experiments. The symbols can be separated by spaces or
practically any other characters, so 'r i' is interpreted
the same as 'r/i'. The total number of components will
determine the final data type of the output, with 1
component corresponding to real values, 2 components to
cmplx1d objects (similar to the normal complex type), and 3
or 4 components to cmplx2d hypercomplex objects.}
}
\value{
A function or tbl_df data frame of functions where each function
        outputs spectral intensity data given a vector input of chemical
        shifts. In the latter case, the functions are stored in a list column
        called f.
}
\description{
This is primarily an internal method that outputs a function (or a tbl_df
data frame of functions), where each function outputs spectral intensity
data given a vector input of chemical shifts.
}
