% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{fft_signal}
\alias{fft_signal}
\title{Fourier tranform signal data}
\usage{
fft_signal(direct.time, signal, acqus.list)
}
\arguments{
\item{direct.time}{A vector of sampling times for the signal.}

\item{signal}{A vector of complex signal data.}

\item{acqus.list}{A list of acqus parameters that contains 'o1', 'sfo1', 
and 'sw' entries. This list can be 
generated using read_acqus_1d() or through other means.
These parameters can also be nested within a list
item called 'acqus' if multiple dimensions are read
at once.}
}
\value{
A data.frame made of two columns -- "direct.shift" containing the 
        chemical shift and "intensity" containing the real valued 
        spectrum data from the real component of the signal.
}
\description{
Performs fourier transform of complex signal data (i.e., in the direct
dimension).
}
