% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{fft_spectrum}
\alias{fft_spectrum}
\title{Fourier tranform intensity data}
\usage{
fft_spectrum(indirect.time, intensity, acqus.list, hypercomplex = FALSE)
}
\arguments{
\item{indirect.time}{A vector of delay times for the signal.}

\item{intensity}{A vector of complex intensity data.}

\item{acqus.list}{A list of lists containing acqus parameters with 
'o1', 'sfo1', 'sw', and 'fnmode' entries. 
This list can be generated using read_acqus() or 
through other means. These parameters can also be nested 
within a list item called 'acqus2' if multiple dimensions
are read at once.}

\item{hypercomplex}{TRUE to output full quadrature components
(rr, ri, ir, ii), FALSE to omit imaginary components
in the direct dimension (ir, ii).}
}
\value{
A data.frame made of two columns -- "indirect.shift" containing
        the indirect dimension chemical shift and "intensity" containing 
        the real component of the fourier transformed direct dimension data 
        (corresponding to rr and ri as a complex pair). If the hypercomplex 
        option is set to TRUE, the ir and ii data is added as another
        complex column labelled "dispersive".
}
\description{
Performs fourier transform of complex intensity data (i.e., in the 
indirect dimension). Although the procedure is very similar to
fft_signal, real/imaginary data processing will differ depending on
the mode of indirect data acquisition.
}
