% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R
\docType{methods}
\name{filter_direct}
\alias{filter_direct}
\alias{filter_direct,NMRData-method}
\alias{filter_indirect}
\alias{filter_indirect,NMRData-method}
\title{Filter NMRData based on chemical shift}
\usage{
filter_direct(object, ...)

\S4method{filter_direct}{NMRData}(object, lower = NA, upper = NA,
  round = FALSE, align = "middle")

filter_indirect(object, ...)

\S4method{filter_indirect}{NMRData}(object, lower = NA, upper = NA,
  round = FALSE, align = "middle")
}
\arguments{
\item{object}{An NMRData object.}

\item{lower}{A lower bound for chemical shift.}

\item{upper}{An upper bound for chemical shift.}

\item{round}{One of either "up", "down", or FALSE. Convolution is
considerably faster when performed on vector lengths that are
powers of two, and as such, it may be beneficial to round the
total number remaining points, even if it means ignoring one or
both of the bounds. Use this option to round the total number of
filtered points up or down.}

\item{align}{One of either "lower", "upper", or "middle". Used to align
points to the specified bounds, when using the "round" option
prevents the specified bounds to be matched exaxctly.}
}
\value{
An NMRData1D object with filtered processed data.
}
\description{
Filter processed data to include only those points that are contained
between a set of lower and upper bounds on chemical shift. Data can be
filtered based on direct or indirect chemical shifts.
}
