% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_1d.R
\docType{data}
\name{fit.env}
\alias{fit.env}
\title{Fit proto class}
\format{An object of class \code{proto} (inherits from \code{environment}) of length 9.}
\usage{
fit.env
}
\description{
Fit proto class.
}
\keyword{datasets}
