% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R, R/NMRFit2D.R
\name{fit}
\alias{fit}
\alias{fit,NMRFit1D-method}
\alias{fit,NMRFit2D-method}
\title{Fit NMR data to peaks.}
\usage{
fit(object, ...)

\S4method{fit}{NMRFit1D}(
  object,
  sf = nmroptions$direct$sf,
  init = nmroptions$fit$init,
  opts = nmroptions$fit$opts
)

\S4method{fit}{NMRFit2D}(
  object,
  direct.sf = nmroptions$direct$sf,
  indirect.sf = nmroptions$indirect$sf,
  init = nmroptions$fit$init,
  opts = nmroptions$fit$opts
)
}
\arguments{
\item{object}{An NMRFit1D object.}

\item{sf}{Sweep frequency (in MHz) -- needed to convert peak widths from Hz
to ppm. In most cases, it is recommended to set a single default
value using nmroptions$direct$sf = ..., but an override can be
provided here.}

\item{init}{An initialization function that takes NMRFit1D object and returns
a modified NMRFit1D object. Use the "identity" function to
override the default initialization in the
nmroptions$fit$init. Note that all arguments provided to the
fit() function are also passed on the init() function.}

\item{opts}{A list of NLOPT fit options to override the default options in
the nmroptionsd$fit$opts.}
}
\description{
Given an NMRFit1D object, this function performs a least squares fit on the
data and updates the peak parameters.
}
