% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcamp.R
\name{flatten_jcamp}
\alias{flatten_jcamp}
\title{Flatten JCAMP list file}
\usage{
flatten_jcamp(jcamp_list)
}
\arguments{
\item{jcamp_list}{The list output of read_jcamp() or process_jcamp().}
}
\value{
A flattened list containing all jcamp entries.
}
\description{
As most JCAMP files contain only one block/ntuple, it's not necessary
to use deep nesting for all the entries. This function checks the
the number of blocks/ntuples and then stores header, block, and ntuple
entries in a single flat list. Only the ntuple element remains a
list, containing the elements 'descriptors' and 'pages'. Note, tag names
are switched between processed and non-processed forms automatically.
If there are name conlicts between header and block elements, only
the block elements are kept.
}
