% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\docType{methods}
\name{format,NMRIntensities-method}
\alias{format,NMRIntensities-method}
\title{Format an NMRIntensities object}
\usage{
\S4method{format}{NMRIntensities}(x, ...)
}
\description{
Format an NMRIntensities object
}
