% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\docType{methods}
\name{format,NMRIntensity-method}
\alias{format,NMRIntensity-method}
\title{Format an NMRIntensity object}
\usage{
\S4method{format}{NMRIntensity}(x, ...)
}
\description{
Format an NMRIntensity object
}
