% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\docType{methods}
\name{gen_parameters}
\alias{gen_parameters}
\alias{.gen_parameters,NMRScaffold1D-method}
\title{Generate a combined vector from individual parameters}
\usage{
.gen_parameters(object, ...)

\S4method{.gen_parameters}{NMRScaffold1D}(object, peaks, baseline,
  baseline.diff, phase)
}
\arguments{
\item{object}{NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Individual parameters that depend on whether the object
is 1D or 2D.}
}
\value{
A combined vector of parameters.
}
\description{
Combines a series of individual parameters into a single vector based
on the Scaffold being fit.
}
