% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData1D.R
\docType{methods}
\name{get_parameter}
\alias{get_parameter}
\alias{get_parameter,NMRData1D-method}
\title{Get parameter from NMRData1D object}
\usage{
get_parameter(object, ...)

\S4method{get_parameter}{NMRData1D}(object, param.name, list.name,
  error = FALSE)
}
\arguments{
\item{object}{An NMRData1D object.}

\item{param.name}{Name of parameter entry.}

\item{list.name}{Name of slot list to acqus (basically either 'acqus' or
'procs')}

\item{error}{Issue an error if the required parameters can't be found.}
}
\description{
A safe method of getting a parameter value from either acqus or procs
list with a fallback to generic parameters list.
}
