% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\docType{methods}
\name{Im,NMRIntensities-method}
\alias{Im,NMRIntensities-method}
\title{Get the pure imaginary component of the data}
\usage{
\S4method{Im}{NMRIntensities}(z)
}
\arguments{
\item{z}{NMRIntensities data.}
}
\value{
A numeric vector.
}
\description{
Retrieves the pure imaginary component of NMRIntensities.
}
