% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\docType{methods}
\name{Im,NMRIntensity-method}
\alias{Im,NMRIntensity-method}
\title{Get the pure imaginary component of the data}
\usage{
\S4method{Im}{NMRIntensity}(z)
}
\arguments{
\item{z}{NMRIntensity data.}
}
\value{
A numeric vector.
}
\description{
Retrieves either the pure imaginary component or the component with the most
imaginary terms when dealing with segmented data.
}
