% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R
\name{indirect-set}
\alias{indirect-set}
\title{Set indirect dimension resonance object}
\usage{
indirect(object) <- value
}
\arguments{
\item{object}{An NMRScaffold2D object.}

\item{value}{A valid NMRScaffold1D object.}
}
\description{
Generic convenience method to set the indirect dimension resonance component
of an NMRScaffold2D object.
}
