% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance2D.R
\docType{methods}
\name{indirect}
\alias{indirect}
\alias{indirect,NMRResonance2D-method}
\title{Get indirect dimension resonance object}
\usage{
indirect(object, ...)

\S4method{indirect}{NMRResonance2D}(object)
}
\arguments{
\item{object}{An NMRResonance2D or NMRSpecies2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the indirect dimension resonance
component of an NMRResonance2D or NMRSpecies2D object.
}
