% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R
\docType{methods}
\name{initialize_heights}
\alias{initialize_heights}
\alias{initialize_heights,NMRScaffold1D-method}
\title{Initialize peak heights of an NMRScaffold1D object}
\usage{
initialize_heights(object, ...)

\S4method{initialize_heights}{NMRScaffold1D}(object, nmrdata,
  exclusion.level = nmroptions$exclusion$level,
  exclusion.notification = nmroptions$exclusion$notification)
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{nmrdata}{An NMRData1D object.}

\item{exclusion.level}{A string specifying what to do when peaks are found to
fall outside of the data range: either 'species' to
exclude the whole species to which the offending peak
belongs, 'resonance' to exclude the resonance to which
the offending peak belongs, or 'peak' to exclude just
the peak itself.}

\item{exclusion.notification}{A function specifying how to report when peaks
are found to be outside the data range: 'none'
to ignore, 'message' to issue a message,
'warning' to issue a warning, and 'stop' to
issue an error.}
}
\value{
A new object with modified peak heights.
}
\description{
Generates peak height estimates based on spectral data. If used on an
anything more complicated than an NMRResonance1D, the function propogates
the bounds to component resonances. Since estimates cannot be provided for
any peak outside the given data range, any such peaks are ignored. Whether
or not warning/error messages are generated when that occurs is specified by
the exclusion.notification parameter. Similarly, exclusion.level provided
options to omit the whole resonance/species if a peak is found to be outside
the data bounds.
}
\details{
At this point, there is just one approach: take peak height as the intensity
of the data at the current position of the peak. There are plans to develop
more sophisticated approaches in the future.
}
