% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\docType{methods}
\name{is_conformant}
\alias{is_conformant}
\alias{.is_conformant,NMRScaffold1D-method}
\title{Check whether NMRScaffold objects conform}
\usage{
.is_conformant(object1, object2)

\S4method{.is_conformant}{NMRScaffold1D}(object1, object2)
}
\arguments{
\item{object1}{NMRScaffold1D or NMRScaffold2D object.}

\item{object2}{NMRScaffold1D or NMRScaffold2D object.}
}
\value{
TRUE if the two objects are conformant, FALSE otherwise.
}
\description{
For their parameters to be conformant, peaks must have the same rows and
columns, while baseline, baseline_difference, and phase must all be the  
same length.
}
