% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R, R/NMRFit2D.R
\docType{methods}
\name{knots-set}
\alias{knots-set}
\alias{knots<-,NMRFit1D-method}
\alias{knots<-,NMRFit2D-method}
\title{Set object baseline knots}
\usage{
knots(object) <- value

\S4method{knots}{NMRFit1D}(object) <- value

\S4method{knots}{NMRFit2D}(object) <- value
}
\arguments{
\item{object}{An NMRFit1D object.}

\item{value}{A vector of chemical shifts used to designate the internal
baseline b-spline knots. Note that the number of internal knots
impacts the length of the baseline and if the length of knots is
changed, current baseline values may not feasible with the new
knots.}
}
\description{
Generic convenience method to set the baseline knots definition of an
NMRFit1D object.
}
