% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R, R/NMRFit2D.R
\docType{methods}
\name{knots}
\alias{knots}
\alias{knots,NMRFit1D-method}
\alias{knots,NMRFit2D-method}
\title{Get object baseline knots}
\usage{
knots(object, ...)

\S4method{knots}{NMRFit1D}(object)

\S4method{knots}{NMRFit2D}(object)
}
\arguments{
\item{object}{An NMRFit1D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the baseline knots definition of an
NMRFit1D object.
}
