% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData1D.R
\name{lines.NMRData1D}
\alias{lines.NMRData1D}
\title{Add to existing NMRData1D plot}
\usage{
lines.NMRData1D(x, p, domain = "r", legendgroup = 2, color = NULL, name = NULL)
}
\arguments{
\item{x}{An NMRData1D object.}

\item{p}{A plot_ly object.}

\item{domain}{One of either 'r' or 'i' corresponding to either real or
imaginary data. are displayed in separate subplots.}

\item{legendgroup}{Unique value for grouping legend entries.}

\item{color}{Line colour as character}

\item{name}{Data name as character.}
}
\value{
A plot_ly plot.
}
\description{
Meant to be primarily an internal function that adds a new line to an
existing plot.
}
