% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData2D.R
\name{lines.NMRData2D}
\alias{lines.NMRData2D}
\title{Add to existing NMRData2D plot}
\usage{
lines.NMRData2D(
  x,
  p,
  domain = "rr",
  legendgroup = 2,
  color = NULL,
  name = NULL
)
}
\arguments{
\item{x}{An NMRData2D object.}

\item{domain}{One of either 'rr', 'ri', 'ir', or 'ii' corresponding to
combinations of real and imaginary data from the direct and
indirect dimensions.}

\item{legendgroup}{Unique value for grouping legend entries.}

\item{color}{Intended for internal use -- line colour as character}

\item{name}{Intended for internal use -- data name as character.}
}
\value{
A plot_ly plot.
}
\description{
Meant to be primarily an internal function that adds a new line to an
existing plot.
}
