% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lineshape_1d}
\alias{lineshape_1d}
\title{Generate 1D-NMR peak lineshape}
\usage{
lineshape_1d(x, par)
}
\arguments{
\item{x}{Vector of chemical shift data in ppm.}

\item{par}{n-by-4 matrix with each row corresponding to a separate
peak to be added together and the 4 columns corresponding
to position, width, height, and fraction_gauss.}
}
\value{
complex vector
}
\description{
Calculates complex lineshapes based on chemical shift and peak parameters.
}
