% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R,
%   R/NMRResonance.R, R/NMRResonance1D.R
\docType{methods}
\name{lower_bounds}
\alias{lower_bounds}
\alias{lower_bounds,NMRScaffold-method}
\alias{lower_bounds,NMRScaffold1D-method}
\title{Access the "lower_bounds" of an NRMScaffold object}
\usage{
lower_bounds(object, ...)

\S4method{lower_bounds}{NMRScaffold}(object)

\S4method{lower_bounds}{NMRScaffold1D}(object)

lower_bounds(object, ...)

\S4method{lower_bounds}{NMRScaffold}(object)

\S4method{lower_bounds}{NMRScaffold1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "lower_bounds" of an NMRScaffold
object.

Generic convenience method to access the "lower_bounds" of an NMRScaffold
object.
}
