% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\name{methodEllipse}
\alias{methodEllipse}
\alias{methodEllipse}
\title{Inherited ellipses description}
\arguments{
\item{...}{Additional arguments passed to inheriting methods.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Inherited ellipses description

Inherited ellipses description
}
