% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmr.chol.bad}
\alias{nmr.chol.bad}
\title{An example NMRData1D object 
The data comes from analysis of a pure cholesterol sample 
using a 1D 13C INEPT pulse sequence. The spectrum has been artificially
modified to feature a 15 degree 0th and 1st order phase error as well as
a rolling baseline.}
\format{An NMRData1D object.}
\usage{
nmr.chol.bad
}
\description{
An example NMRData1D object 
The data comes from analysis of a pure cholesterol sample 
using a 1D 13C INEPT pulse sequence. The spectrum has been artificially
modified to feature a 15 degree 0th and 1st order phase error as well as
a rolling baseline.
}
\keyword{datasets}
