% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmr.chol}
\alias{nmr.chol}
\title{An example NMRData1D object 
The data comes from analysis of a pure cholesterol sample 
using a 1D 13C INEPT pulse sequence.}
\format{An NMRData1D object.}
\usage{
nmr.chol
}
\description{
An example NMRData1D object 
The data comes from analysis of a pure cholesterol sample 
using a 1D 13C INEPT pulse sequence.
}
\keyword{datasets}
