% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R
\docType{class}
\name{NMRData-class}
\alias{NMRData-class}
\alias{NMRData}
\title{Super class combining NMR data and scan parameters.}
\description{
See NMRData1D or NMRData2D.
}
\section{Slots}{

\describe{
\item{\code{processed}}{A data.frame containing chemical shift and intensity data.}

\item{\code{acqus}}{A list of acqus parameters.}

\item{\code{procs}}{A list of procs parameters.}
}}

