% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData1D.R
\docType{class}
\name{NMRData1D-class}
\alias{NMRData1D-class}
\alias{NMRData1D}
\title{A class combining NMR data and scan parameters}
\description{
Although it's likely to change in the near future, this class currently
serves to combine processed 1D data from the pdata folder of an experiment
with the acqus and procs parameter files.
}
\section{Slots}{

\describe{
\item{\code{processed}}{1r/1i data stored as a data.frame with direct.shift and
intensity as columns.}

\item{\code{acqus}}{A list of acqus parameters.}

\item{\code{procs}}{A list of procs parameters.}

\item{\code{product}}{A vector representation of an apodization function that was
used to generate the intensity data (in this context, zero-
filling is considered to be the equivalent of apodization by a
step function). To facilitate convolution, the length of this
vector is set to 2*n - 1 where n is the number of spectrum
points.}

\item{\code{convolution}}{The Fourier transform of the product slot, suitable for
convolution with the spectrum intensity data.}
}}

