% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData1D.R
\docType{class}
\name{NMRData1D-class}
\alias{NMRData1D-class}
\alias{NMRData1D}
\title{A class combining NMR data and scan parameters}
\description{
An extension of the generic NMRData class to provide 1D-specific methods
such as constructors and plotting.
}
