% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData1D.R
\name{nmrdata_1d}
\alias{nmrdata_1d}
\title{Constructor for generating an NMRData1D object}
\usage{
nmrdata_1d(path, number = NA)
}
\arguments{
\item{path}{Path to a Bruker scan directory.}

\item{number}{Processing file number. Defaults to the smallest
number in the pdata directory.}
}
\value{
An NMRData1D object containing the 1r/1i processed data as well
        as the procs and acqus parameters.
}
\description{
Loads data in Bruker folder and uses it to initialize an NMRData1D object.
}
