% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData2D.R
\name{nmrdata_2d}
\alias{nmrdata_2d}
\alias{nmrdata_2d_from_pdata}
\title{Constructors for generating an NMRData1D object}
\usage{
nmrdata_2d(path, procs.number = NA)

nmrdata_2d_from_pdata(path, procs.number = NA)
}
\arguments{
\item{path}{Path to a Bruker scan directory.}

\item{procs.number}{Specifies pdata directory to load. Defaults to the lowest
available number.}
}
\value{
An NMRData1D object.
}
\description{
\code{nmrdata_2d()} can be used as a generic constructor method that will
eventually handle different types of input (currently limited to Bruker
pdata directories). The specific constructor functions can also be called
using \code{nmrdata_2d_from_pdata()}.
}
