% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData2D.R
\name{nmrdata_2d_from_scaffold}
\alias{nmrdata_2d_from_scaffold}
\title{Convert NMRScaffold2D to NMRData2D}
\usage{
nmrdata_2d_from_scaffold(
  object,
  direct.shift = NULL,
  indirect.shift = NULL,
  direct.sf = nmroptions$direct$sf,
  indirect.sf = nmroptions$indirect$sf
)
}
\arguments{
\item{object}{An NMRScaffold2D object}

\item{direct.shift}{Vector of chemical shift data in ppm.}

\item{indirect.shift}{Similar to direct.shift but for the indirect dimension.}

\item{direct.sf}{Sweep frequency (in MHz) -- needed to convert peak widths
from Hz to ppm. In most cases, it is recommended to set a
single default value using nmroptions$direct$sf = ..., but
an override can be provided here.}

\item{indirect.sf}{Similar to direct.sf but for the indirect dimension.}
}
\value{
An NMRData2D object.
}
\description{
Generate lineshape from specified peaks.
}
