% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\name{nmrfit_1d}
\alias{nmrfit_1d}
\title{Generate an NMRFit1D object}
\usage{
nmrfit_1d(
  species,
  nmrdata,
  baseline.order = nmroptions$baseline$order,
  n.knots = nmroptions$baseline$n.knots,
  phase.order = nmroptions$phase$order,
  delay.fit = FALSE,
  sf = nmroptions$direct$sf,
  init = nmroptions$fit$init,
  opts = nmroptions$fit$opts,
  ...
)
}
\arguments{
\item{species}{A list of NMRSpecies1D objects or other objects that can be
converted to NMRSpecies1D objects. See ?nmrresonance_1d and
?nmrspecies_1d for more details about this conversion. If list
elements are named, these names will be use to replace species
ids.}

\item{nmrdata}{An NMRData1D object used to fit the supplied peaks.
automatically generated from the resonance names.}

\item{baseline.order}{An integer specifying the order of the baseline spline
function. Note the the internal B-spline implementation
requires an order of 1 or greater. You can use an order
of -1 to disable baseline correction.}

\item{n.knots}{An integer specifying the number of internal knots to use. The
specific position of these knots can be modified later using
knots() function. To modify initial values of the baseline,
use baseline() function.}

\item{phase.order}{An integer specifying the order of the phase correction
polynomial -- may be 0 or 1 (-1 to disable).
in practice, but a higher order correction is possible.}

\item{delay.fit}{FALSE to immediately run least squares optimization after
the NMRFit1D object is initialized, TRUE to skip the
optimization, enabling more customization. The fit can be
run manually using fit().}

\item{sf}{Sweep frequency (in MHz) -- needed to convert peak widths from Hz
to ppm. In most cases, it is recommended to set a single default
value using nmroptions$direct$sf, but an override can be provided
here.}

\item{init}{An initialization, function that takes an NMRFit1D object and
returns a modified NMRFit1D object. Use the "identity" function
to override the default initialization in the
nmroptions$fit$init. Note that all arguments provided to the
fit() function are also passed on the init() function.}

\item{opts}{A list of NLOPT fit options to override the default options in
the nmroptions$fit$opts.}

\item{...}{Options passed to nmrmixture_1d if conversion has to be performed.
See ?nmrmixture_1d for more details.}
}
\value{
An NMRFit1D object.
}
\description{
Generates an NMRFit1D object based on an NMRMixture1D or a list of
NMRSpecies1D objects. See ?nmrmixture_1d and ?nmrspecies_1d for more details
about this conversion. It is important to note that all NMR objects defined
in the species list must fall inside the provided data (although partial
fitting may be implemented in the future).
}
