% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\docType{class}
\name{NMRIntensities-class}
\alias{NMRIntensities-class}
\alias{NMRIntensities}
\title{Container for NMR intensity data}
\description{
Multiple NMRintensity vectors can be combined in an NMRIntensities list that
supports the same Re(), Im() and $ functions as well as conversions to
data.frames.
}
