% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\docType{class}
\name{NMRIntensity-class}
\alias{NMRIntensity-class}
\alias{NMRIntensity}
\title{Container for NMR intensity data}
\description{
Although 1D intensity can be represented by a single real/imaginary complex
pair, there is no built-in representation for 2D (or higher dimension) data
that may consist of real-real, real-imaginary, imaginary-real, and
imaginary-imaginary components. This class approximates the extension of a
complex values to multiple dimensions be defining a vector, where each
element corresponds to a specific real-imaginary component. Re() and Im()
functions can still be used to obtain pure real and pure imaginary
compenents, whereas other specific real/imaginary can be are accesible with
the $ function. For 2D data, for example, $rr will output a numeric vector
corresponding to the real-real data (equivalent to Re()).
}
