% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\name{nmrintensity}
\alias{nmrintensity}
\title{Generate an NMRIntensity or NMRIntensities object}
\usage{
nmrintensity(...)
}
\arguments{
\item{...}{A collection of vector, matrix, or data.frame objects. Vectors
must be specified with names whereas the component names for
matrix and data.frame objects are taken from their column names.}
}
\description{
Generates an NMRIntensity or plural NMRIntensities object from a series of
vector, matrix, or data.frame inputs.
}
