% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMixture1D.R
\docType{class}
\name{NMRMixture1D-class}
\alias{NMRMixture1D-class}
\alias{NMRMixture1D}
\title{Definition of a mixture of species}
\description{
In the same way that an NMRSpecies1D object is just a collection of
NMRResonance1D objects, an NMRMixture1D object is a collection of
NMRSpecies1D objects. This class serves primarily for organizational
purposes and as the basis of fitting methods.
}
\section{Slots}{

\describe{
\item{\code{species}}{A list of NMRSpecies1D objects.}
}}

