% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMixture2D.R
\docType{class}
\name{NMRMixture2D-class}
\alias{NMRMixture2D-class}
\alias{NMRMixture2D}
\title{Definition of a mixture of species}
\description{
In the same way that an NMRSpecies2D object is just a collection of
NMRResonance2D objects, an NMRMixture2D object is a collection of
NMRSpecies2D objects. This class serves primarily for organizational
purposes and as the basis of fitting methods.
}
\section{Slots}{

\describe{
\item{\code{species}}{A list of NMRSpecies2D objects.}
}}

