% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMixture1D.R
\name{nmrmixture_1d}
\alias{nmrmixture_1d}
\title{Generate an NMRMixture1D object}
\usage{
nmrmixture_1d(species, ...)
}
\arguments{
\item{species}{A list of NMRSpecies1D objects or other objects that can be
converted to NMRSpecies1D objects. See ?nmrresonance_1d and
?nmrspecies_1d for more details about this conversion. If list
elements are named, these names will be use to replace
species ids.}

\item{...}{Options passed to nmrspecies_1d if resonances are being
converted from character/numeric vectors. See ?nmrresonance_1d for
more details.}
}
\value{
An NMRMixture1D object.
}
\description{
Generates an NMRMixture1D object from a list of NMRResonance1D/NMRSpecies1D
objects or a list of character/numeric vectors that can be converted to
NMRResonance1D objects. See ?nmrresonance_1d for more details about this
conversion.
}
