% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRMixture2D.R
\name{nmrmixture_2d}
\alias{nmrmixture_2d}
\title{Generate an NMRMixture2D object}
\usage{
nmrmixture_2d(species, ...)
}
\arguments{
\item{species}{A list of NMRSpecies2D objects. If list elements are named,
these names will be use to replace species ids.}

\item{...}{Currently ignored.}
}
\value{
An NMRMixture2D object.
}
\description{
Generates an NMRMixture2D object from a list of NMRResonance2D/NMRSpecies2D
objects.
}
