% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\docType{class}
\name{NMRResonance1D-class}
\alias{NMRResonance1D-class}
\alias{NMRResonance1D}
\title{Definition of an NMR resonance.}
\description{
Essentially, this class is used to define coupling relationships to group
individual peaks into resonances.
}
\section{Slots}{

\describe{
\item{\code{peaks}}{A data.frame describing a series of singlets, with one row per
peak. All peaks are characterized by a position (in ppm), height
(in relative intensity units), width (in ppm or Hz), and
fraction.guass (in percent).}

\item{\code{couplings}}{A data.frame relating the position and area parameters of
the peaks, effectively combining singlets into multiplets.}

\item{\code{couplings.leeway}}{A list specifying how tightly enforced the coupling
constraints on peak positions, areas, and widths
should be. E.g. position = 0 specifies that the j
coupling constant is exact, whereas width = 0.1
specifies that the widths of individual peaks may
differ by +/- 10 percent.}

\item{\code{bounds}}{A list of lower and upper bounds on the peak parameters where
both bounds take the same shape as the peaks data.frame.}
}}

