% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance2D.R
\name{nmrresonance_2d}
\alias{nmrresonance_2d}
\title{Generate an NMRResonance2D object based on simplified peak list}
\usage{
nmrresonance_2d(direct.peaks, indirect.peaks,
  direct.sf = nmrsession_2d("sf")$direct,
  indirect.sf = nmrsession_2d("sf")$indirect, id = NULL, width = 1,
  fraction.gauss = 0, position.leeway = 0, area.leeway = 0,
  width.leeway = 0)
}
\arguments{
\item{direct.peaks}{An NMRResonance1D object, numeric vector of singlet
chemical shifts, or a character string specifying
multiplets of the form "3 d 1.2". See ?parse_peaks_1d for
more information.}

\item{indirect.peaks}{An NMRResonance1D object, numeric vector of singlet
chemical shifts, or a character string specifying
multiplets of the form "3 d 1.2". See ?parse_peaks_1d
for more information.}

\item{direct.sf}{Sweep frequency (MHz) in the direct dimension -- needed to
convert coupling constants from Hz to ppm. In most cases, it
is recommended to set a single default value using
nmrsession_2d$sf$direct  = ..., but an override can be
provided here.}

\item{indirect.sf}{Sweep frequency (MHz) in the indirect dimension -- needed
to convert coupling constants from Hz to ppm. In most
cases, it is recommended to set a single default value
using nmrsession_2d$sf$indirect  = ..., but an override
can be provided here.}

\item{id}{A string specifying resonance name. If left empty, a name is
automatically generated from the peaks argument.}

\item{width}{Initial estimate of peak width (in Hz). For Voigt lineshapes,
this value is taken as the Lorentzian component, with the
Gaussian component calculated from
peak.width*frac.guass/(1-frac.gauss).}

\item{fraction.gauss}{Fraction of overall peak width that corresponds to a
Gaussian lineshape. A value of 0 corresponds to a
Lorentz peak whereas a value of 1 corresponds to a
Gaussian peak. Values in between 0 and 1 are modelled
as a Voigt lineshape but the specific value of
frac.gauss does not have a physical interpretation.}

\item{position.leeway}{A fraction specifying how tightly enforced the
coupling constraints on peak positions, should be.
E.g. coupling.leeway = 0 specifies that the j coupling
constant is exact, whereas couping.leeway = 0.1
specifies that the coupling constant may differ by +/-
10 percent.}

\item{area.leeway}{Similar to position.leeway but for peak areas. Determines
how strictly the coupling area ratios are enforced.}

\item{width.leeway}{Similar to position.leeway but for peak widths.
Determines how strictly equal peak widths for all coupled
peaks are enforced.}
}
\value{
An NMRResonance2D object.
}
\description{
Generates an NMRResonance2D object by combining direct and indirect peak
definitions. These peak definitions can be provided either with a string
that will be parsed by \code{parse_peaks_1d()} or a previously generated
NMRResonance1D object.
}
