% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R
\docType{class}
\name{NMRScaffold-class}
\alias{NMRScaffold-class}
\alias{NMRScaffold}
\title{Super-class for all 1D and 2D peak descriptions.}
\description{
This class is not meant to be used directly. Instead, it provides a set of
common methods for all 1D and 2D NMR objects. As such, it has no slots and
all of its methods are meant to be inherited by the objects listed above.
}
