% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R
\name{nmrscaffold_1d}
\alias{nmrscaffold_1d}
\title{Generate an NMRScaffold1D object based on simplified peak list}
\usage{
nmrscaffold_1d(peak.list, nmrdata, peak.type = "lorenz",
  peak.width = 1, baseline.degree = 3, n.knots = 3,
  include.difference = TRUE, include.phase = TRUE, store = TRUE)
}
\arguments{
\item{peak.list}{A vector or list of character strings specifying
multiplets of the form "3 d 1.2". See details section.}

\item{nmrdata}{An NMRData1D object.}

\item{peak.type}{One of lorenz, gauss, pvoigt, or voigt.}

\item{peak.width}{Initial estimate of peak width (in Hz)}

\item{baseline.degree}{Degree of B-spline polynomials to use for baseline.
As the degree can be 0, use NULL to remove
baseline entirely (default value).}

\item{n.knots}{The number of equally spaced interior knots used for 
B-spline baseline. The knots can be changed after the
scaffold in initialized.}

\item{include.difference}{Include a baseline difference term to account
for slight difference between real and imaginary
components. By default, this difference is
initialized to be the same as the baseline
term. This can be changed later.}

\item{include.phase}{Include a phasing term in the fit.}

\item{store}{FALSE to not store the nmrdata object. Although storing
an nmrdata object is likely to be more convenient,
it may result in unnecessary data copies.}
}
\value{
An NMRScaffold1D object.
}
\description{
Generates an NMRScaffold1D object by converting multiplet definitions into
a set of singlets related by constraints on their position and area.
height and convert coupling constants from Hz to ppm. If no data is 
provided, peak heights are assumed to be 0. This is a bad initial 
guess and is unlikely to result in a good fit.
}
\details{
Multiplets are specified in the from of "3 d 1.2", which indicates
a doublet at 3 ppm, with a coupling constant of 1.2 Hz. Supported  
multiplet codes include s, d, t, q, quint, sext, sept, oct, dd, dt, 
td, and tt. The latter 4 expect two coupling values. The exact 
syntax of the specification is quite flexible, as the function  
removes all brackets and most punctuation and converts common spacers 
like - and _ into whitespace. Satellite coupling constants (in Hz) can be
added using "j" in the specification, so "3 s j 10" would indicate
a singlet at 3 ppm and 1 pair of satellites with a coupling constant of
10 Hz. It's also possible to offset the center of the satellite
doublet from the center of the main isotopic peak using a colon. In the
specification of "3 s j 10:0.01", the center of the satellite doublet is
offset 0.01 ppm upfield.
}
