% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{nmrsession_1d}
\alias{nmrsession_1d}
\title{Set global options}
\usage{
nmrsession_1d(..., RESET = FALSE, READ.ONLY = NULL, LOCAL = FALSE,
  ADD = FALSE)
}
\arguments{
\item{sf}{Sweep frequncy (MHz). Used to convert peak width and coupling
information from Hz into ppm.}
}
\description{
A number of common parameters are used across different functions. Such
global options are assumed to be valid for all (or at least most) analysis
carried out during a single session.
}
