% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{nmrsession_2d}
\alias{nmrsession_2d}
\title{Set global options for 2D peak processing and fitting}
\usage{
nmrsession_2d(..., RESET = FALSE, READ.ONLY = NULL, LOCAL = FALSE,
  ADD = FALSE)
}
\description{
A number of common parameters are used across different functions. Such
global options are assumed to be valid for all (or at least most) analysis
carried out during a single session.
}
