% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRSpecies1D.R
\docType{class}
\name{NMRSpecies1D-class}
\alias{NMRSpecies1D-class}
\alias{NMRSpecies1D}
\title{Definition of an NMR species.}
\description{
Essentially, this class is used to collect and define area relations between
multiple resonances as part of a single species.
}
\section{Slots}{

\describe{
\item{\code{resonances}}{A list of NMRSpecies1D objects.}

\item{\code{connections}}{A data.frame relating the areas of the resonances.}

\item{\code{connections.leeway}}{A value specifying how tightly enforced the
connection constraints on resonance areas should be.
E.g. a value of = 0 specifies that the area ratios
are exact, whereas 0.1 specifies that the
areas of the resonances may differ by +/- 10
percent from the specified ratios.}
}}

