% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRSpecies1D.R
\name{nmrspecies_1d}
\alias{nmrspecies_1d}
\title{Generate an NMRSpecies1D object}
\usage{
nmrspecies_1d(resonances, areas = NULL, id = NULL,
  connections.leeway = 0, ...)
}
\arguments{
\item{resonances}{A list of NMRResonance1D objects or a list of
character/numeric vectors that can be converted to
NMRResonance1D objects. See ?nmrresonance_1d for more
details about this conversion. If list elements are named,
these names will be use to replace resonance ids.}

\item{areas}{A vector of areas corresponding to the expected areas of the
resonances. Set to NULL by default, signifying no fixed
constraings.}

\item{id}{A string specifying species name. If left empty, a name is
automatically generated from the resonance names.}

\item{connections.leeway}{A value specifying how tightly enforced the
connection constraints on resonance areas should
be. E.g. a value of = 0 specifies that the area
ratios are exact, whereas 0.1 specifies that the
areas of the resonances may differ by +/- 10
percent from the specified ratios.}

\item{...}{Options passed to nmrresonance_1d if resonances are being
converted from character/numeric vectors. See ?nmrresonance_1d for
more details.}
}
\value{
An NMRSpecies1D object.
}
\description{
Generates an NMRSpecies1D object from a list of NMRResonance1D objects or a
list of character/numeric vectors that can be converted to NMRResonance1D
objects. See ?nmrresonance_1d for more details about this conversion.
}
