% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRSpecies2D.R
\name{nmrspecies_2d}
\alias{nmrspecies_2d}
\title{Generate an NMRSpecies2D object}
\usage{
nmrspecies_2d(
  resonances,
  volumes = NULL,
  id = NULL,
  connections.leeway = 0,
  ...
)
}
\arguments{
\item{resonances}{A list of NMRResonance2D objects.}

\item{volumes}{A vector of volumes corresponding to the expected volume
ratios of the resonances. Set to NULL by default, signifying
no fixed constraints.}

\item{id}{A string specifying species name. If left empty, a name is
automatically generated from the resonance names.}

\item{connections.leeway}{A value specifying how tightly enforced the
connection constraints on resonance volumes should
be. E.g. a value of = 0 specifies that the volume 
ratios are exact, whereas 0.1 specifies that the
volumes of the resonances may differ by +/- 10
percent from the specified ratios.}

\item{...}{Currently ignored.}
}
\value{
An NMRSpecies2D object.
}
\description{
Generates an NMRSpecies2D object from a list of NMRResonance2D objects.
}
