% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{normalized-set}
\alias{normalized-set}
\alias{normalized<-,NMRScaffold-method}
\alias{normalized<-,NMRScaffold1D-method}
\alias{normalized<-,NMRFit1D-method}
\title{Replace the "normalized" slot of an NMRScaffold object}
\usage{
normalized(object) <- value

\S4method{normalized}{NMRScaffold}(object) <- value

\S4method{normalized}{NMRScaffold1D}(object) <- value

\S4method{normalized}{NMRFit1D}(object) <- value
}
\description{
Generic method to replace the normalized slot of NMRScaffold1D or 
NMRScaffold2D object. This is a convenience function, see set_normalized() 
for more details.
}
