% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\name{pack_peaks}
\alias{pack_peaks}
\alias{pack_peaks,NMRMixture1D-method}
\title{Convert peak parameters into vector for fitting}
\usage{
pack_peaks(object, ...)

\S4method{pack_peaks}{NMRMixture1D}(
  object,
  sf = nmroptions$direct$sf,
  shift.range = c(0, 1),
  intensity.range = c(0, 1)
)
}
\arguments{
\item{object}{An NMRMixture1D object.}

\item{sf}{Field strength (used to convert peak width from Hz to ppm).}

\item{shift.range}{A two-element vector representing the ppm range of the fit
(used to scale peak width and position).}

\item{intensity.range}{A two-element vector representing the intensity range
of the fit (used to scale peak heights).}
}
\value{
A list of vectors -- peaks, lb, and ub.
}
\description{
This is an internal function used to convert the peaks tibble into a single
vector for the optimization algorithm.
}
