% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R
\docType{methods}
\name{parameters-set}
\alias{parameters-set}
\alias{parameters<-,NMRData-method}
\title{Set generic parameters}
\usage{
parameters(object) <- value

\S4method{parameters}{NMRData}(object) <- value
}
\description{
Set generic parameters
}
