% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R
\docType{methods}
\name{parameters}
\alias{parameters}
\alias{parameters,NMRData-method}
\title{Access generic parameters}
\usage{
parameters(object, ...)

\S4method{parameters}{NMRData}(object)
}
\description{
Access generic parameters
}
