% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R
\name{parse_constraints}
\alias{parse_constraints}
\alias{parse_constraints,NMRScaffold-method}
\title{Parse constraints for fitting}
\usage{
parse_constraints(object, ...)

\S4method{parse_constraints}{NMRScaffold}(object, direct.span = 1, indirect.span = 1)
}
\arguments{
\item{object}{An NMRScaffold object.}

\item{direct.span}{The ppm range of the fit used to scale position
differences in the direct dimension.}

\item{indirect.span}{The ppm range of the fit used to scale position
differences in the indirect dimension.}
}
\description{
This is an internal function used to parse equality and inequality
constraints into a form that can be used for lower level fit code.
Essentially, each constraint is defined as a vector with a variable number
of elements codified as follows: 1) Code of either 0 (position), 1 (width),
2 (height), 3 (fraction.gauss), 4 (area); 2) The equality or inequality
target value; 3+) Integers corresponding to peak indexes in the overall
parameter vector. Positive vs negative values are treated differently
depending on the code. For positions, parameters with negative indexes are
subtracted while for width and area, all parameters with negative indexes
are added up before dividing the positive ones.
}
