% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_coupling}
\alias{parse_coupling}
\title{Parse coupling strings}
\usage{
parse_coupling(coupling.string)
}
\arguments{
\item{coupling.string}{A character vector with elements of the form 
'3.0 d 1.0'}
}
\value{
A list of three lists -- chemical.shift, peaks, and coupling
        respectively. Both peaks and couplng lists may have one or two
        length vectors as elements. Singlets are assigned a coupling of NA.
}
\description{
Converts a coupling specification of the form '3.0 d 1.0' into the
chemical shift, thenumber of peaks involved and the coupling between them. 
Currently, the supported codes include s, d, t, q, quint, sext, sept, oct, 
dd, dt, td, and tt. The latter 4 expect two coupling values. The exact 
syntax of the specification is quite flexible, as the function tries to 
remove all brackets and punctuation and convert common spacers like - and _
into whitespace.
}
