% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\name{parse_peaks_1d}
\alias{parse_peaks_1d}
\title{Parse peak coupling strings}
\usage{
parse_peaks_1d(coupling.string)
}
\arguments{
\item{coupling.string}{A character vector with elements of the form '3.0 d
1.0' or '2 dt 1.0 1.2'}
}
\value{
A list of three vectors -- chemical.shift, peak numbers, and coupling
        constants respectively.
}
\description{
Converts a coupling specification of the form '3.0 d 1.0' into the chemical
shift, number of peaks involved and the coupling between them. Currently,
the supported codes include s, d, t, q, pentet, sextet, septet, octet, and
nonet or any combination of the above. Essentially, the function splits the
coupling into a position (number), coupling description (text), and coupling
constant (number), parsing the latter two to extract the required codes and
numbers. The actual parsing is quite flexible so "1.0 dtpnt 1.5/2/1" will be
parsed the same as "1.0 d t pentet 1.5 2 1".
}
