% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{peak_type-set}
\alias{peak_type-set}
\alias{peak_type<-,NMRScaffold-method}
\alias{peak_type<-,NMRScaffold1D-method}
\alias{peak_type<-,NMRFit1D-method}
\title{Replace the "peak_type" slot of an NMRScaffold object}
\usage{
peak_type(object) <- value

\S4method{peak_type}{NMRScaffold}(object) <- value

\S4method{peak_type}{NMRScaffold1D}(object) <- value

\S4method{peak_type}{NMRFit1D}(object) <- value
}
\description{
Generic method to replace the peak_type of NMRScaffold1D or NMRScaffold2D
object. This is a convenience function that makes some assumptions, 
see set_peak_type() for more details.
}
