% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\docType{methods}
\name{peak_units-set}
\alias{peak_units-set}
\alias{peak_units<-,NMRScaffold-method}
\alias{peak_units<-,NMRScaffold1D-method}
\title{Replace the "peak_units" slot of an NMRScaffold object}
\usage{
peak_units(object) <- value

\S4method{peak_units}{NMRScaffold}(object) <- value

\S4method{peak_units}{NMRScaffold1D}(object) <- value
}
\description{
Generic method to replace the peak_units slot of NMRScaffold1D or 
NMRScaffold2D object. This is a convenience function, see set_peak_units() 
for more details.
}
