% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R, R/NMRFit2D.R
\name{phase}
\alias{phase}
\alias{phase,NMRFit1D-method}
\alias{phase,NMRFit2D-method}
\title{Get object phase}
\usage{
phase(object, ...)

\S4method{phase}{NMRFit1D}(object)

\S4method{phase}{NMRFit2D}(object)
}
\arguments{
\item{object}{An NMRFit1D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the phase definition of an
NMRFit1D object.
}
