% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{phase_signal}
\alias{phase_signal}
\title{Correct signal phase}
\usage{
phase_signal(signal, phase, degrees = TRUE)
}
\arguments{
\item{signal}{A vector of complex signal data.}

\item{phase}{A vector of phase corrections, from zero-order and up.}

\item{degrees}{TRUE if the phase angles are in degrees, FALSE if in radians.}
}
\value{
Corrected signal data.
}
\description{
This function applies a phase correction directly to the signal in the 
time domain once the correct phase angles have been identified.
}
