% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{phase_spectrum}
\alias{phase_spectrum}
\title{Correct spectrum phase}
\usage{
phase_spectrum(spectrum, phase, degrees = TRUE)
}
\arguments{
\item{spectrum}{A vector of complex spectrum data.}

\item{phase}{A vector of phase corrections, from zero-order and up.}

\item{degrees}{TRUE if the phase angles are in degrees, FALSE if in radians.}
}
\value{
Corrected spectrum.
}
\description{
This function applies a phase correction to the fourier transformed
spectrum once the correct phase angles have been identified.
}
