% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\name{pillar_shaft.NMRIntensities}
\alias{pillar_shaft.NMRIntensities}
\title{Format tibble pillar for NMRIntensities object}
\usage{
\method{pillar_shaft}{NMRIntensities}(x, ...)
}
\description{
Format tibble pillar for NMRIntensities object
}
