% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData1D.R
\name{plot.NMRData1D}
\alias{plot.NMRData1D}
\title{Plot NMRData1D object}
\usage{
plot.NMRData1D(x, domain = "r", legendgroup = 1, color = NULL, name = NULL)
}
\arguments{
\item{x}{An NMRData1D object.}

\item{domain}{One of either 'r' or 'i' corresponding to either real or
imaginary data.}

\item{legendgroup}{Unique value for grouping legend entries.}

\item{color}{Intended for internal use -- line colour as character}

\item{name}{Intended for internal use -- data name as character.}
}
\value{
A plot_ly plot.
}
\description{
Convenience function that generates a plot of the spectral data.
}
