% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData2D.R
\name{plot.NMRData2D}
\alias{plot.NMRData2D}
\title{Plot NMRData2D object}
\usage{
plot.NMRData2D(x, components = "rr")
}
\arguments{
\item{x}{An NMRData2D object.}

\item{components}{One of 'rr', 'ii', 'ir', or 'ri' to specify various real
and imaginary components. 3D plot subplots are not
currently supported.}
}
\value{
A plot_ly plot.
}
\description{
Convenience function that generates a plot of the spectral data.
}
