% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\name{plot.NMRFit1D}
\alias{plot.NMRFit1D}
\title{Plot NMRFit1D object}
\usage{
plot.NMRFit1D(x, components = "r", apply.phase = TRUE,
  sum.level = "species", sum.baseline = TRUE)
}
\arguments{
\item{x}{An NMRFit1D object.}

\item{components}{One of either 'r', 'i', or 'r/i' to include real, imaginary
or both components. If both components are selected, they
are displayed in separate subplots.}

\item{apply.phase}{TRUE to apply the calculated phase to the data.}

\item{sum.level}{One of either 'all', 'species', 'resonance', 'peak' to
specify whether all peaks should be summed together the
peaks should be summed at a lower level.}

\item{sum.baseline}{TRUE to add the baseline to each fit.}
}
\value{
A ggplot2 plot.
}
\description{
Generates an interactive plot object using the plotly package.
}
\details{
Convenience function that generates a graphical representation of the fit.
The original data is plotted as a black line, the fit is plotted in red, the
baseline is plotted in blue, the residual in red. The fit can be plotted as
a composite of all the peaks, or individually.
}
