% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit2D.R
\name{plot.NMRFit2D}
\alias{plot.NMRFit2D}
\title{Plot NMRFit2D object}
\usage{
plot.NMRFit2D(
  x,
  components = "rr",
  apply.phase = TRUE,
  sum.level = "species",
  sum.baseline = TRUE
)
}
\arguments{
\item{x}{An NMRFit2D object.}

\item{components}{One of 'rr', 'ii', 'ir', or 'ri' to specify various real
and imaginary components. 3D plot subplots are not
currently supported.}

\item{apply.phase}{TRUE to apply the calculated phase to the data.}

\item{sum.level}{One of either 'all', 'species', 'resonance', to specify
whether all peaks should be summed together.}

\item{sum.baseline}{TRUE to add the baseline to each fit.}
}
\value{
A ggplot2 plot.
}
\description{
Generates an interactive plot object using the plotly package.
}
\details{
Convenience function that generates a graphical representation of the fit.
The original data is plotted as a black line, the fit is plotted in red, the
baseline is plotted in blue, the residual in red. The fit can be plotted as
a composite of all the peaks, or individually.
}
