% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R
\name{plot.NMRScaffold1D}
\alias{plot.NMRScaffold1D}
\title{Plot NMRScaffold1D object}
\usage{
plot.NMRScaffold1D(
  x,
  domain = "r",
  direct.shift = NULL,
  direct.sf = nmroptions$direct$sf,
  sum.level = "species",
  add.baseline = TRUE,
  add.phase = TRUE
)
}
\arguments{
\item{x}{An NMRScaffold1D object.}

\item{domain}{One of either 'r' or 'i' corresponding to either real or
imaginary data. are displayed in separate subplots.}

\item{direct.shift}{Used to override default selection of chemical shift
values.}

\item{direct.sf}{Sweep frequency (in MHz) -- needed to convert peak widths
from Hz to ppm. In most cases, it is recommended to set a
single default value using nmroptions$direct$sf = ..., but
an override can be provided here.}

\item{sum.level}{One of either 'all', 'species', 'resonance', 'peak' to
specify whether all peaks should be summed together.}

\item{add.baseline}{TRUE to add calculated baseline correction (if it exists)
to each fit.}

\item{add.phase}{TRUE to add the calculated phase correction (if it exists)
to the data.}
}
\value{
A ggplot2 plot.
}
\description{
Generates an interactive plot object using the plotly package.
}
\details{
If the input is an NMResonance1D, NMRSpecies1D, or NMRMixture1D, the peak
lines are simply drawn in red. If the input is an NMRFit1D object, then the
output features more components -- the original data is plotted as a black
line, the fit is plotted in red, the baseline is plotted in blue, the
residual in green. The fit can be plotted as a composite of all the peaks,
or individually.
}
