% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcamp.R
\name{process_jcamp}
\alias{process_jcamp}
\title{Process imported JCAMP data}
\usage{
process_jcamp(jcamp_list, tags = TRUE, entries = TRUE, ...)
}
\arguments{
\item{jcamp_list}{The output of read_jcamp().}

\item{tags}{TRUE to convert entry tags. All tags are converted to 
lowercase with dots replacing spaces. Some tags known to
take multiple forms ('JCAMPDX' vs. 'JCAMP-DX') are converted
to a single fixed form (full list to come).}

\item{entries}{TRUE to convert entries using a system of ad-hoc rules.
All entries designated with '<>' are converted to strings.
All other entries are converted to numeric if possible.
Lists and tables are converted to matrices if possible.}

\item{...}{Extra arguments passed into process_jcamp_tag().}
}
\value{
...
}
\description{
Converts string key-value pairs from parsed JCAMP file into numeric and 
matrix formats where applicable.
}
