% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcamp.R
\name{process_jcamp_entry}
\alias{process_jcamp_entry}
\title{Process JCAMP entry}
\usage{
process_jcamp_entry(jcamp.entry, sep = NULL)
}
\arguments{
\item{jcamp.entry}{Single character string from the output of read_jcamp().}

\item{sep}{A regex string used to split entries into vectors. By default,
no splitting is performed.}
}
\value{
Character string, numeric vector, or data.frame depending on
        the form of the entry
}
\description{
Converts single JCAMP character string entry into numeric vector or
data.frame types where possible. If no conversion options found, returns
character string unchanged.
}
