% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcamp.R
\name{process_jcamp_tag}
\alias{process_jcamp_tag}
\title{Process JCAMP tag}
\usage{
process_jcamp_tag(jcamp.tag, f_case = tolower, tag.space = ".",
  tag.map = NA)
}
\arguments{
\item{jcamp.tag}{Single character string corresponding to list names
from the output of read_jcamp().}

\item{f_case}{The function used to perform the case conversion, e.g.
tolower().}

\item{tag.space}{A string used to replace empty spaces and underscores.}

\item{tag.map}{A vector of strings corresponding to new tags,
where the vector names correspond to the tags that will 
be renamed. NULL avoids map conversions while NA uses
the default conversions below:}
}
\value{
A renamed character string.
}
\description{
Processes a JCAMP tag into standard form. This consists of selecting
a case conversion function, such as tolower(), as well as a map of
specfic conversions to make. Note, case conversions are applied after
the mapped conversions.
}
