% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R, R/NMRData1D.R
\docType{methods}
\name{processed-set}
\alias{processed-set}
\alias{processed<-,NMRData-method}
\alias{processed<-,NMRData1D-method}
\title{Replace the "processed" of an NMRData object}
\usage{
processed(object) <- value

\S4method{processed}{NMRData}(object) <- value

\S4method{processed}{NMRData1D}(object) <- value
}
\arguments{
\item{object}{An NMRData1D or NMRData2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "processed" of an NMRData
object.
}
