% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R
\docType{methods}
\name{processed}
\alias{processed}
\alias{processed,NMRData-method}
\title{Access Fourier-transformed data}
\usage{
processed(object, ...)

\S4method{processed}{NMRData}(object)
}
\description{
Access Fourier-transformed data
}
