% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R
\docType{methods}
\name{procs-set}
\alias{procs-set}
\alias{procs<-,NMRData-method}
\title{Set procs parameters}
\usage{
procs(object) <- value

\S4method{procs}{NMRData}(object) <- value
}
\description{
Set procs parameters
}
