% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R, R/NMRData1D.R
\docType{methods}
\name{procs}
\alias{procs}
\alias{procs,NMRData-method}
\alias{procs,NMRData1D-method}
\title{Access the "procs" slot of an NRMData object}
\usage{
procs(object, ...)

\S4method{procs}{NMRData}(object)

\S4method{procs}{NMRData1D}(object)
}
\arguments{
\item{object}{An NMRData1D or NMRData2d object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "procs" slot of an NMRData
object.
}
