% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData2D.R
\name{project}
\alias{project}
\title{Get 1D projection of dimension}
\usage{
projection(object, dimension, ...)
}
\arguments{
\item{object}{An NMRData2D or NMRScaffold2D object.}

\item{dimension}{Currently limited to either "direct" or "indirect"}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Warning: this is currently a convenience method to facilitate internal data
manipulation, so the conversion may drop important data. When applied to a
an NMRScaffold2D object, non-selected components are simply dropped, and
when applied to an NMRData2D object, the non-selected are summed together to
give a 1D projection. No normalization is performed so non-uniform sampling
may result in distorted results. The functions direct(object) and
indirect(object) are provided as shortcuts to projection(object, "direct")
and projection(object, "indirect").
}
