% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\docType{methods}
\name{propagate_to_bounds}
\alias{propagate_to_bounds}
\alias{.propagate_to_bounds,NMRScaffold-method}
\alias{.propagate_to_bounds,NMRScaffold1D-method}
\title{Propagate function to bounds}
\usage{
.propagate_to_bounds(object, ...)

\S4method{.propagate_to_bounds}{NMRScaffold}(object, f, ...)

\S4method{.propagate_to_bounds}{NMRScaffold1D}(object, f, ...)
}
\arguments{
\item{object}{NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Extra arguments to f.}

\item{f}{Function to be applied.}
}
\value{
Modified NMRScaffold object.
}
\description{
Since lower and upper bounds should be members of the same class as
the object they are bounding, function applied on the parent object
can be directly applied to the bounds.
}
