% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\docType{methods}
\name{Re,NMRIntensities-method}
\alias{Re,NMRIntensities-method}
\title{Get the pure real component of the data}
\usage{
\S4method{Re}{NMRIntensities}(z)
}
\arguments{
\item{z}{NMRIntensities data.}
}
\value{
A numeric vector.
}
\description{
Retrieves the pure real component of NMRIntensities.
}
