% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\docType{methods}
\name{Re,NMRIntensity-method}
\alias{Re,NMRIntensity-method}
\title{Get the pure real component of the data}
\usage{
\S4method{Re}{NMRIntensity}(z)
}
\arguments{
\item{z}{NMRIntensity data.}
}
\value{
A numeric vector.
}
\description{
Retrieves either the pure real component or the component with the most real
terms when dealing with segmented data.
}
