% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{read_acqus_file}
\alias{read_acqus_file}
\title{Read Bruker acqus parameters}
\usage{
read_acqus_file(path, ...)
}
\arguments{
\item{path}{Character string that points to an acqus file.}

\item{...}{Arguments passed into process_jcamp().}
}
\value{
A list made up of nested lists with the processed acqus entries.
}
\description{
Reads the acqus parameters from a single file.  Since the acqus files are 
in JCAMP format, the actual parsing is just a thin wrapper around
read_jcamp(), process_jcamp(), and flatten_jcamp() that reads Bruker 
scan parameters and puts them in a flat list.
}
