% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcamp.R
\name{read_jcamp}
\alias{read_jcamp}
\title{Import data from generic JCAMP file}
\usage{
read_jcamp(path, process.tags = TRUE, process.entries = TRUE, ...)
}
\arguments{
\item{path}{Path to file.}

\item{process.tags}{TRUE to convert entry tags to lowercase with dots
in place of spaces.}

\item{process.entries}{FALSE to keep all block elements as characters.
TRUE to attempt some basic conversion.}

\item{...}{Arguments passed into process_jcamp().}
}
\value{
A list with the following entries: 'header', 'blocks',  
        and 'comments'. Header elements are all those contained between
        the 'TITLE', 'OWNER', and 'ORIGIN', and may vary between files.
        All bon-header entries (including notes) are stored
        in the 'blocks' sublist. If no block information is specified,
        all information is stored in block 1. comments' contains a list of 
        comment blocks indexed by corresponding JCAMP element of the line 
        number on which they were found (if there is no corresponding 
        element). If either convert.tags or contvert.entries is true,
        the parsed output is also processed using the process_jcamp()
        function. See ?process_jcamp for more details.
}
\description{
Reads data from JCAMP file in the most general fashion possible. No
checks are performed on file validity. Tag and entry conversion is
done by passing output directly into process_jcamp().
}
