% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{read_processed_1d}
\alias{read_processed_1d}
\title{Read 1D Bruker 1r/1i files}
\usage{
read_processed_1d(path, procs.list, number = NA)
}
\arguments{
\item{path}{Character string that points to a Bruker scan directory.}

\item{procs.list}{A list of procs parameters that contains 'sw.p', 'si',
'sf', 'reverse', and 'offset' entries. This list can be
generated using read_procs_1d() or through other means.
These parameters can also be nested within a list
item called 'procs' if multiple dimensions are read
at once.}

\item{number}{The processing file number.Defaults to smallest number in pdata
directory.}
}
\value{
A data.frame made of two columns -- "direct.shift" and "intensity",
        corresponding to direct dimension chemical shift and the complex 
        spectrum intensity data, respectively.
}
\description{
Reads processed bruker 1D files based on specified parameters.
}
