% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{read_processed_2d}
\alias{read_processed_2d}
\title{Read 2D Bruker rr/ri/ir/ii files}
\usage{
read_processed_2d(path, procs.list, number = NA, hypercomplex = TRUE)
}
\arguments{
\item{path}{Character string that points to a Bruker scan directory.}

\item{procs.list}{A list of lists containing procs parameters with 
'sw.p', 'si', 'sf', 'reverse', and 'offset'  entries. 
This list can be generated using read_procs() or 
through other means. Unless the sublists are named
procs and proc2s, the first and second sublits are
assumed to correspond to direct and indirect dimensions.}

\item{number}{The processing file number. Defaults to the smallest number
in the pdata directory.}

\item{hypercomplex}{TRUE to output full quadrature components
(2rr, 2ri, 2ir, 2ii) if present, FALSE to omit imaginary 
components in the indirect dimension (2ir, 2ii).}
}
\value{
A data.frame made of three or four columns -- "direct.shift"
        containing the direct dimension chemical shift, "indirect.shift" 
        containing the indirect dimension chemical shift, "intensity",
        and possibly "intensity2". See description for more details.
}
\description{
Reads processed bruker 2D files based on specified parameters.
As the processed data can vary considerably based on the quadrature 
method used, the resulting data.frame will also vary quite a bit. 
If all 2rr/2ri/2ir/2ii files are present, they are stored as 2 complex 
valued columns named intensity (2rr/2ri) and intensity2 (2ir/2ii). 
If only two files are present, they are stored as a complex valued 
column intensity, regardless of which files are actually present, 
e.g. 2rr/2ii or 2rr/2ir.
}
