% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{read_procs_dir}
\alias{read_procs_dir}
\title{Read Bruker procs parameters}
\usage{
read_procs_dir(path, number = NA, ...)
}
\arguments{
\item{path}{Character string that points to a scan directory.}

\item{number}{The processing file number to use. Defaults to smallest
number in pdata folder.}

\item{...}{Arguments passed into process_jcamp().}

\item{dimension}{The dimension of the scan parameters (1 or 2). 
Defaults to loading all}
}
\value{
A list made up of nested lists with the processed procs entries.
}
\description{
Reads the procs parameters across all acquired dimensions. Data from
each dimension is stored as a separate list item in Bruker order,
i.e., list(procs = ..., proc2s = ..., etc.). Since the procs files are 
in JCAMP format, the actual parsing is just a thin wrapper around
read_jcamp(), process_jcamp(), and flatten_jcamp() that reads Bruker 
scan parameters and puts them in a flat list.
}
