% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{read_signal_1d}
\alias{read_signal_1d}
\title{Read 1D Bruker fid file}
\usage{
read_signal_1d(path, acqus.list)
}
\arguments{
\item{path}{Character string that points to either an fid file or
the broader scan directory.}

\item{acqus.list}{A list of acqus parameters that contains 'td', 'sw',
'sfo1', and 'bytorda' entries. This list can be
generated using read_acqus_1d() or through other means.
These parameters can also be nested within a list
item called 'acqus' if multiple dimensions are read
at once.}
}
\value{
A data.frame made of two columns -- "direct.time" and "signal",
        corresponding to sampling time and the complex signal,
        respectively.
}
\description{
Reads a raw bruker 1D fid file based on specified parameters.
}
