% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruker.R
\name{read_signal_2d}
\alias{read_signal_2d}
\title{Read 2D Bruker ser file}
\usage{
read_signal_2d(path, acqus.list)
}
\arguments{
\item{path}{Character string that points to either a ser file or
the broader scan directory.}

\item{acqus.list}{A list of lists containing acqus parameters with 
'td', 'sw', 'sfo1', and 'bytorda' entries. 
This list can be generated using read_acqus() or 
through other means. Unless the sublists are named
acqus and acqu2s, the first and second sublits are
assumed to correspond to direct and indirect dimensions.}
}
\value{
A data.frame made of three columns -- "direct.time",
        "indirect.timea," and "signal", corresponding to signal sampling
        time, the delay time, and the complex signal, respectively.
}
\description{
Reads a raw bruker 2D ser file based on specified parameters.
}
