% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{safe_read}
\alias{safe_read}
\title{Safely read data from file}
\usage{
safe_read(path, data, ...)
}
\arguments{
\item{path}{Path to file.}

\item{data}{Character determining read function to use: 'char' for
readChar or 'bin' for readBin.}

\item{...}{Arguments passed to underlying read function.}
}
\value{
Character or numeric vector depending on input data.
}
\description{
Reads data from file if it exists and is readable. Outputs sane
error message if either of these conditions is not met.
}
