% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\docType{methods}
\name{set_absolute_bounds}
\alias{set_absolute_bounds}
\alias{set_absolute_bounds,NMRScaffold1D-method}
\title{Set absolute bounds on an NMRScaffold1D or NMRScaffold2D object}
\usage{
set_absolute_bounds(object, position = NULL, height = NULL,
  width = NULL, baseline = NULL, phase = NULL, normalized = FALSE,
  peak.units = "hz", ...)

\S4method{set_absolute_bounds}{NMRScaffold1D}(object, position = NULL,
  height = NULL, width = NULL, baseline = NULL, phase = NULL,
  normalized = FALSE, peak.units = "hz")
}
\arguments{
\item{object}{An NMRScaffold1D or NMRscaffold2D object.}

\item{position}{A vector of two elements corresponding to a lower 
and upper bound for peak position. If the normalized
argument is true, 0 corresponds to the leftmost range
of the data and 1 to the rightmost. Otherwise, the units
are in ppm.}

\item{height}{A vector of two elements corresponding to a lower
and upper bound for peak height. If the normalized
argument is true, 0 corresponds to the lowest value 
of spectral intensity and 1 to the largest. Otherwise, 
the units correspond to arbitrary spectral intensity values.}

\item{width}{A vector of two elements corresponding to a lower
and upper bound for peak width. If the normalized
argument is true, 0 corresponds to the leftmost range
of the data and 1 to the rightmost. Otherwise, the units
depend on the peak.units argument.}

\item{baseline}{A vector of two elements corresponding to a lower 
and upper bound for baseline height. If the normalized
argument is true, 0 corresponds to the lowest value 
of spectral intensity and 1 to the largest. Otherwise, 
the units correspond to arbitrary spectral intensity 
values. The same constraints are applied to the real
baseline and the imaginary baseline difference.}

\item{phase}{A vector of two elements corresponding to a lower and upper 
bound for phase in radians.}

\item{normalized}{TRUE to set bounds in terms of the underlying data, where
the x and y values of the data are scaled between 0 and 1.
FALSE to use natural units of ppm/Hz and spectral
intensity.}

\item{peak.units}{The units of peak width -- either 'ppm' or 'hz'.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\value{
A new NMRScaffold1D or NMRScaffold2D object with modified parameters.
}
\description{
Although any NMRScaffold object can act as a boundary on another
NMRScaffold object, this function provides a convenience method
for generating such bounds using a simple set of lower and upper
constraints on basic peak parameters such as position, height,
width, etc... Note that the term "absolute" refers to the fact
that the same bounds are applied to each and every peak, regardless
of current parameter values. However, these bounds can still be
normalized to the data using the normalized argument.
}
