% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRResonance1D.R,
%   R/NMRSpecies1D.R, R/NMRMixture1D.R, R/NMRFit1D.R, R/NMRResonance2D.R,
%   R/NMRSpecies2D.R, R/NMRMixture2D.R, R/NMRFit2D.R
\name{set_conservative_bounds}
\alias{set_conservative_bounds}
\alias{set_conservative_bounds,NMRResonance1D-method}
\alias{set_conservative_bounds,NMRSpecies1D-method}
\alias{set_conservative_bounds,NMRMixture1D-method}
\alias{set_conservative_bounds,NMRFit1D-method}
\alias{set_conservative_bounds,NMRResonance2D-method}
\alias{set_conservative_bounds,NMRSpecies2D-method}
\alias{set_conservative_bounds,NMRMixture2D-method}
\alias{set_conservative_bounds,NMRFit2D-method}
\title{Set conservative bounds on an NMRScaffold1D object}
\usage{
set_conservative_bounds(object, ...)

\S4method{set_conservative_bounds}{NMRResonance1D}(
  object,
  position = TRUE,
  height = TRUE,
  width = TRUE,
  nmrdata = NULL,
  widen = FALSE
)

\S4method{set_conservative_bounds}{NMRSpecies1D}(object, ...)

\S4method{set_conservative_bounds}{NMRMixture1D}(object, ...)

\S4method{set_conservative_bounds}{NMRFit1D}(
  object,
  ...,
  nmrdata = NULL,
  widen = FALSE,
  baseline = TRUE,
  phase = TRUE
)

\S4method{set_conservative_bounds}{NMRResonance2D}(
  object,
  position = TRUE,
  height = TRUE,
  width = TRUE,
  nmrdata = NULL,
  widen = FALSE
)

\S4method{set_conservative_bounds}{NMRSpecies2D}(object, ...)

\S4method{set_conservative_bounds}{NMRMixture2D}(object, ...)

\S4method{set_conservative_bounds}{NMRFit2D}(
  object,
  ...,
  nmrdata = NULL,
  widen = FALSE,
  baseline = TRUE,
  phase = TRUE
)
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{position}{Without reference data, position is limited to plus or minus
0.1 ppm. With reference data, the position of the peaks is
also forced inside the domain of the data. FALSE to disable.}

\item{height}{Without reference data, height is set to strictly positive.
With reference data, height is also limited to no more than 150
percent of the maximum peak value. FALSE to disable.}

\item{width}{Without reference data, minimum peak width is set to almost, but
not quite 0 Hz (1e-3 Hz) and a maximum peak width of 3 Hz. With
reference data, peak width is also prevented from being more
than 20 percent of the data range. FALSE to disable.}

\item{nmrdata}{An optional NMRData1D object that can serve as a reference
point for the bounds.}

\item{widen}{FALSE to prevent new bounds from widening existing bounds.}

\item{baseline}{Without reference data, the baseline control points are left
unrestricted. With reference data, the real baseline control
points are limited to no more than 50 percent of the maximum
peak value. FALSE to disable.}

\item{phase}{With or without reference data, phase correction is limited to
-pi/2 to pi/2.}
}
\value{
A new object with modified bounds.
}
\description{
A convenience function that sets reasonable bounds on an NMRScaffold1D
object. These bounds are assumed to be widely applicable to most simple NMR
data. Each set of bounds can be turned on or off as necessary. A slightly
better set of bounds can be selected if a reference NMRData1D object is
provided. If used on an anything more complicated than an NMRResonance1D,
the function propogates the bounds to component resonances.
}
