% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData1D.R
\docType{methods}
\name{set_convolution}
\alias{set_convolution}
\alias{set_convolution,NMRData1D-method}
\title{Set convolution applied to processed data}
\usage{
set_convolution(object, trim = 0, param = list(), f = NULL)

\S4method{set_convolution}{NMRData1D}(object, trim = 0, param = list(),
  f = NULL)
}
\arguments{
\item{object}{An NMRData1D object.}

\item{trim}{Number of point to truncate from the end of the fid's 
acquisition time (not including zero fill). It's common to 
circle shift Bruker's group delay artefact from the start of 
a collected fid to the last 50-100 points. While most 
apodization methods will passively get rid of or diminish the
influence of these points, they may have a stronger impact
with minimal or no apodization.}

\item{param}{A list of lists combining apodization names and parameters,
(see apodize_signal) for a full list of apodizations and their
parameters. E.g. list('exponential'=list('lb'=1)).}

\item{f}{A function f(n) capable of generating the convolution vector
of length n (ranging from 0 to the acquisition time). If provided, 
f overrides the param list.}
}
\value{
An NMRData1D object with modified convolution vector.
}
\description{
Although it's possible to fit the processed data based on its final
appearance, it may sometimes be more useful to try and fit the underlying
lineshape with an explicit consideration of the convolution that it has
undergone. This convolution includes any apodization as well as the process 
of zero-filling, which can be considered as the convolution of a 
hypothetically longer experiment by a step function. By default, the 
applied apodization is read from the procs parameters, but these can
be overwritten by either a list of parameters or a custom function.
}
