% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R,
%   R/NMRFit1D.R
\docType{methods}
\name{set_general_bounds}
\alias{set_general_bounds}
\alias{set_general_bounds,NMRResonance1D-method}
\alias{set_general_bounds,NMRSpecies1D-method}
\alias{set_general_bounds,NMRFit1D-method}
\title{Set general bounds of an NMRResonance1D object}
\usage{
set_general_bounds(object, ...)

\S4method{set_general_bounds}{NMRResonance1D}(object, position = NULL,
  height = NULL, width = NULL, fraction.gauss = NULL,
  nmrdata = NULL, widen = FALSE)

\S4method{set_general_bounds}{NMRSpecies1D}(object, ...)

\S4method{set_general_bounds}{NMRFit1D}(object, ..., nmrdata = NULL,
  widen = FALSE, baseline = NULL, phase = NULL)
}
\arguments{
\item{object}{An NMRResonance1D, NMRSpecies1D, or NMRFit1D object.}

\item{position}{A vector of two elements corresponding to a lower and upper
bound for peak position. If nmrdata is provided, 0
corresponds to the leftmost range of the data and 1 to the
rightmost. Otherwise, the units are in ppm.}

\item{height}{A vector of two elements corresponding to a lower and upper
bound for peak height. If nmrdata is provided, 0 corresponds to
the lowest value of spectral intensity and 1 to the largest.
Otherwise, the units correspond to arbitrary spectral intensity
values.}

\item{width}{A vector of two elements corresponding to a lower and upper
bound for peak width in Hz. If nmrdata is provided, values are
taken as fraction of the general data range. So 0.1 would
correspond to a nominal peak width that covers a tenth of the
general data range.}

\item{fraction.gauss}{A vector of two elements corresponding to a lower and
upper bound for the Gaussian fraction of the peak. This
can be set to c(0, 0) to force Lorentzian peaks. Any
values smaller than 0 will be treated as 0 and any
values greater than 1 will be treated as 1.}

\item{nmrdata}{An optional NMRData1D object that can serve as a reference
point for the bounds.}

\item{widen}{FALSE to prevent new bounds from widening existing bounds.}

\item{baseline}{Only applicable to NMRFit1D objects. A complex vector of two
elements corresponding to a lower and upper bound for both
real and imaginary baseline control points (which roughly
corresponds to a baseline value). If the vector has no
imaginary component, the imaginary baseline is left
unbounded. If nmrdata is provided, 0 corresponds to the
lowest value of spectral intensity and 1 to the largest.
Otherwise, the units correspond to arbitrary spectral
intensity values.}

\item{phase}{Only applicable to NMRFit1D objects. A vector of two elements
corresponding to a lower and upper bound for the phase
correction (in radians) at any point in the chemical shift
range.}
}
\value{
A new object with modified bounds.
}
\description{
This function provides a convenience method for generating bounds using a
simple set of lower and upper constraints on basic peak parameters such as
position, height, width, fraction.gauss. The term "general" refers to the
fact that the same bounds are applied to each and every peak, regardless of
current parameter values. These bounds can be normalized to a set of data
using the optional nmrdata argument. If applied to an NMRSpecies1D or
NMRFit1D object, the same bounds are propagated to every component
resonance.
}
\details{
In practice, general bounds are primarily useful for placing a hard
constraint on peak widths and preventing negative heights. Values of 0 for
widths and height can also cause issues during optimization, so simple
general bounds can be used to prevent errors.
}
