% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R,
%   R/NMRResonance.R, R/NMRResonance1D.R
\docType{methods}
\name{set_normalized}
\alias{set_normalized}
\alias{set_normalized,NMRScaffold1D-method}
\title{Normalize or undo normalization of a NMRScaffold1D or NMRScaffold2D}
\usage{
set_normalized(object, nmrdata = NULL, normalized = TRUE,
  include.bounds = FALSE, ...)

\S4method{set_normalized}{NMRScaffold1D}(object, nmrdata = NULL,
  normalized = TRUE, include.bounds = FALSE)

set_normalized(object, nmrdata = NULL, normalized = TRUE,
  include.bounds = FALSE, ...)

\S4method{set_normalized}{NMRScaffold1D}(object, nmrdata = NULL,
  normalized = TRUE, include.bounds = FALSE)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{nmrdata}{An optional NMRData object used for normalization.}

\item{normalized}{TRUE to normalize, FALSE to undo normalization.}

\item{include.bounds}{TRUE to perform the same operation on bounds.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{nmrdata}{An optional NMRData object used for normalization.}

\item{normalized}{TRUE to normalize, FALSE to undo normalization.}

\item{include.bounds}{TRUE to perform the same operation on bounds.}
}
\value{
A new NMRScaffold1D or NMRFit1D object with modified parameters.

A new NMRScaffold1D or NMRFit1D object with modified parameters.
}
\description{
The fit process benefits from scaling x and y data to a limited 0-1
range. Naturally, this requires all parameters be normalized to a 
set of data. By default, parameters are normalized to the internally
stored nmrdata slot, however, a different NMRData1D object can also
be provided.

The fit process benefits from scaling x and y data to a limited 0-1
range. Naturally, this requires all parameters be normalized to a 
set of data. By default, parameters are normalized to the internally
stored nmrdata slot, however, a different NMRData1D object can also
be provided.
}
