% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R
\docType{methods}
\name{set_offset_bounds}
\alias{set_offset_bounds}
\alias{set_offset_bounds,NMRResonance1D-method}
\alias{set_offset_bounds,NMRSpecies1D-method}
\title{Set offset bounds of an NMRResonance1D object}
\usage{
set_offset_bounds(object, ...)

\S4method{set_offset_bounds}{NMRResonance1D}(object, position = NULL,
  height = NULL, width = NULL, relative = FALSE, widen = FALSE)

\S4method{set_offset_bounds}{NMRSpecies1D}(object, ...)
}
\arguments{
\item{object}{An NMRResonance1D or an NMRSpecies1D object.}

\item{position}{A vector of two elements to be added to current peak
positions to generate a set of lower and upper bounds. If
relative is true, the values are treated as fractions to be
multipled by the current peak position before addition.
fraction of the current position.}

\item{height}{A vector of two elements to be added to current peak heights to
generate a set of lower and upper bounds. If relative is true,
the values are treated as fractions to be multipled by the
current peak heights before addition.}

\item{width}{A vector of two elements to be added to current peak widths to
generate a set of lower and upper bounds. If relative is true,
the values are treated as fractions to be multipled by the
current peak heights before addition.}

\item{relative}{TRUE to treat values as relative fractions, FALSE to apply
them directly.}

\item{widen}{FALSE to prevent new bounds from widening existing bounds.}
}
\value{
A new NMRResonance1D object with modified bounds.
}
\description{
This function provides a convenience method for generating bounds using a
simple set of lower and upper constraints on basic peak parameters such as
position, height, width, fraction.gauss. The term "offset" refers to the
fact that bounds are applied as an offset to current values of the
parameters. These bounds can be expressed in absolute (e.g. -0.1 and +0.1
ppm) or relative (e.g. -1 percent and +1 percent) terms. If applied to an
NMRSpecies1D object, the same offset bounds are propogated to every
resonance.
}
\details{
In practice, offset bounds are primarily useful for preventing peak
positions from drifting too much from initial guesses and for fine-tuning a
fit once an initial optimization is performed. It is not recommended to use
strict offset bounds based on rough initial parameter guesses.
}
