% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRScaffold2D.R,
%   R/NMRResonance1D.R, R/NMRSpecies1D.R, R/NMRMixture1D.R, R/NMRFit1D.R
\docType{methods}
\name{set_peak_type}
\alias{set_peak_type}
\alias{set_peak_type,NMRScaffold2D-method}
\alias{set_peak_type,NMRResonance1D-method}
\alias{set_peak_type,NMRSpecies1D-method}
\alias{set_peak_type,NMRMixture1D-method}
\alias{set_peak_type,NMRFit1D-method}
\title{Set peak type of an NMRScaffold1D object}
\usage{
set_peak_type(object, ...)

\S4method{set_peak_type}{NMRScaffold2D}(object, peak.type)

\S4method{set_peak_type}{NMRResonance1D}(object, peak.type)

\S4method{set_peak_type}{NMRSpecies1D}(object, ...)

\S4method{set_peak_type}{NMRMixture1D}(object, ...)

\S4method{set_peak_type}{NMRFit1D}(object, ...)
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{peak.type}{One of either "lorentz", "voigt", "gauss", or "any" where
"any" clears existing bounds on the fraction.gauss
parameter.}
}
\value{
A new object with modified bounds and peak parameters.
}
\description{
The peak type of an NMRScaffold1D object is governed by the fraction.gauss
parameter and can fluidly go from pure Lorentz to Gauss via the combined
Voigt lineshape. However, it can be computationally efficient to force a
specific peak type. This function provides a shortcut for doing so by
changing the fraction.gauss parameter as well as lower and upper bounds.
}
