% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R,
%   R/NMRResonance.R, R/NMRResonance1D.R
\docType{methods}
\name{set_peak_units}
\alias{set_peak_units}
\alias{set_peak_units,NMRScaffold-method}
\alias{set_peak_units,NMRScaffold1D-method}
\title{Switch NMRScaffold1D or NMRScaffold2D parameter units between Hz and ppm}
\usage{
set_peak_units(object, nmrdata = NULL, peak.units = "hz",
  include.bounds = TRUE, ...)

\S4method{set_peak_units}{NMRScaffold}(object, nmrdata = NULL,
  peak.units = "hz", include.bounds = TRUE)

\S4method{set_peak_units}{NMRScaffold1D}(object, nmrdata = NULL,
  peak.units = "hz", include.bounds = TRUE)

set_peak_units(object, nmrdata = NULL, peak.units = "hz",
  include.bounds = TRUE, ...)

\S4method{set_peak_units}{NMRScaffold}(object, nmrdata = NULL,
  peak.units = "hz", include.bounds = TRUE)

\S4method{set_peak_units}{NMRScaffold1D}(object, nmrdata = NULL,
  peak.units = "hz", include.bounds = TRUE)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{nmrdata}{An optional NMRData object needed for conversion parameters.}

\item{peak.units}{One of either 'hz' or 'ppm'.}

\item{include.bounds}{TRUE to perform the same operation on bounds.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{peak.units}{One of either 'hz' or 'ppm'.}

\item{nmrdata}{An optional NMRData object needed for conversion parameters.}

\item{include.bounds}{TRUE to perform the same operation on bounds.}
}
\value{
An updated NMRScaffold1D or NMRScaffold2D object.

An updated NMRScaffold1D or NMRScaffold2D object.
}
\description{
Although it's more practical to refer to peak width in terms of Hz, 
optimization requires that the width parameters be in ppm. This conversion
should generally be handled automatically, but it can be overriden manually
if desired.

Although it's more practical to refer to peak width in terms of Hz, 
optimization requires that the width parameters be in ppm. This conversion
should generally be handled automatically, but it can be overriden manually
if desired.
}
