% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\docType{methods}
\name{set_relative_bounds}
\alias{set_relative_bounds}
\alias{set_relative_bounds,NMRScaffold1D-method}
\title{Set relative bounds on an NMRScaffold1D or NMRScaffold2D object}
\usage{
set_relative_bounds(object, nmrdata = NULL, overall = 0.1,
  position = NULL, height = NULL, width = NULL, ...)

\S4method{set_relative_bounds}{NMRScaffold1D}(object, nmrdata = NULL,
  overall = 0.1, position = NULL, height = NULL, width = NULL)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{nmrdata}{An NMRData object.}

\item{overall}{Overall fraction difference applied to all parameters.}

\item{position}{Fraction difference override applied to position parameters.}

\item{height}{Fraction difference override applied to height parameters.}

\item{width}{Fraction difference override applied to width parameters.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\value{
A new NMRScaffold1D or NMRFit1D object with modified parameters.
}
\description{
Although any NMRScaffold object can act as a boundary on another
NMRScaffold object, this function provides a convenience method
for generating bounds based on a relative difference from current
parameter values.
}
