% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData1D.R
\docType{methods}
\name{show,NMRData1D-method}
\alias{show,NMRData1D-method}
\title{Display NMRData1D object}
\usage{
\S4method{show}{NMRData1D}(object)
}
\description{
Display a quick summary of the processed data.frame.
}
