% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\docType{methods}
\name{show,NMRFit1D-method}
\alias{show,NMRFit1D-method}
\title{Display NMRFit1D object}
\usage{
\S4method{show}{NMRFit1D}(object)
}
\description{
Display a quick summary of resonance parameters.
}
