% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit2D.R
\docType{methods}
\name{show,NMRFit2D-method}
\alias{show,NMRFit2D-method}
\title{Display NMRFit2D object}
\usage{
\S4method{show}{NMRFit2D}(object)
}
\description{
Display a quick summary of resonance parameters.
}
