% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\docType{methods}
\name{show,NMRIntensities-method}
\alias{show,NMRIntensities-method}
\title{Show an NMRIntensity object}
\usage{
\S4method{show}{NMRIntensities}(object)
}
\arguments{
\item{object}{An NMRIntensity object.}
}
\description{
A concise display for NMRIntensity.
}
