% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\docType{methods}
\name{show,NMRIntensity-method}
\alias{show,NMRIntensity-method}
\title{Show an NMRIntensity object}
\usage{
\S4method{show}{NMRIntensity}(object)
}
\description{
Show an NMRIntensity object
}
