% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\docType{methods}
\name{show,NMRResonance1D-method}
\alias{show,NMRResonance1D-method}
\title{Display NMRResonance1D object}
\usage{
\S4method{show}{NMRResonance1D}(object)
}
\description{
Display a quick summary of resonance parameters.
}
