% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance2D.R
\docType{methods}
\name{show,NMRResonance2D-method}
\alias{show,NMRResonance2D-method}
\title{Display NMRResonance2D object}
\usage{
\S4method{show}{NMRResonance2D}(object)
}
\description{
Display a quick summary of resonance parameters.
}
