% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R
\docType{methods}
\name{show,NMRScaffold1D-method}
\alias{show,NMRScaffold1D-method}
\title{Display NMRScaffold1D object}
\usage{
\S4method{show}{NMRScaffold1D}(object)
}
\description{
Display a quick summary of scaffold parameters.
}
