% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R
\name{show,NMRScaffold2D-method}
\alias{show,NMRScaffold2D-method}
\title{Display NMRScaffold2D object}
\usage{
\S4method{show}{NMRScaffold2D}(object)
}
\description{
Display a quick summary of resonance parameters.
}
