% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRSpecies1D.R
\docType{methods}
\name{show,NMRSpecies1D-method}
\alias{show,NMRSpecies1D-method}
\title{Display NMRSpecies1D object}
\usage{
\S4method{show}{NMRSpecies1D}(object)
}
\description{
Display a quick summary of resonance parameters.
}
