% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R
\name{split_all}
\alias{split_all}
\alias{split_all,NMRScaffold-method}
\title{Split input across all components of object}
\usage{
split_all(object, setter, slot.name, value)

\S4method{split_all}{NMRScaffold}(object, setter, slot.name, value)
}
\arguments{
\item{object}{An NMRScaffold object.}

\item{setter}{Setter function.}

\item{slot.name}{Slot name as character.}

\item{value}{Value stored in specified slot.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
This is an internal function used for all setter functions that input a
data.frame object -- splitting apart input across all components.
}
