% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R
\name{split_children}
\alias{split_children}
\alias{split_children,NMRScaffold1D-method}
\title{Split children if they exist}
\usage{
split_children(object, setter, slot.name, value)

\S4method{split_children}{NMRScaffold1D}(object, setter, slot.name, value)
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{slot.name}{Slot name as character.}

\item{value}{Value stored in specified slot.}

\item{getter}{Setter function.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
This is an internal function used for all setter functions that input a
data.frame object. Essentially, the input value is first split based on a
"resonance" or "species" column, with the setter being passed on to the 
specified children.
}
