% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{split_coupling}
\alias{split_coupling}
\title{Generate splitting pattern based on coupling numbers and constants.}
\usage{
split_coupling(peaks, coupling)
}
\arguments{
\item{peaks}{A vector of peaks per coupling constant, e.g., c(2, 2) for dd.}

\item{coupling}{A vector of coupling constants.

        relation to the center.}
}
\description{
Converts a vector of peak numbers and coupling constants into a pattern
of single peaks with accompanying area/height ratios as well as chemical
shift offsets. An input of c(2, 2) and c(1, 2) would correspond to a
double doublet with coupling constants of 1 and 2 respectively. Although
units aren't kept track of, it makes most sense to convert coupling
constants to ppm.
}
