% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R
\name{split_dimensions}
\alias{split_dimensions}
\alias{split_dimensions,NMRScaffold2D-method}
\title{Split direct and indirect dimensions}
\usage{
split_dimensions(object, setter, value)

\S4method{split_dimensions}{NMRScaffold2D}(object, setter, value)
}
\arguments{
\item{object}{An NMRScaffold2D object.}

\item{setter}{Setter function.}

\item{value}{Value stored in specified slot.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
This is an internal function used for all setter functions that input a
data.frame object. Essentially, the input value is first split based on a
"dimension" column, with the setter being passed on to the direct and
indirect components.
}
