% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\name{split_peaks_1d}
\alias{split_peaks_1d}
\title{Split peaks data.frame according to specified coupling.}
\usage{
split_peaks_1d(peaks, number, constant)
}
\arguments{
\item{peaks}{A peaks data.frame from NMRResonance1D object.}

\item{number}{The number of output peaks per input peak.}

\item{constant}{The coupling constant of the split.}
}
\value{
A modified data.frame suitable for NMRResonance1D object.
}
\description{
Split peaks data.frame according to specified coupling.
}
