% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R
\docType{methods}
\name{split_setter}
\alias{split_setter}
\alias{split_setter,NMRScaffold2D-method}
\title{Split direct and indirect dimensions}
\usage{
split_setter(object, setter, value)

\S4method{split_setter}{NMRScaffold2D}(object, setter, value)
}
\arguments{
\item{object}{An NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
This is an internal function used for all setter functions that input a
data.frame object. Essentially, the input value is first split based on a
"dimension" column, with the setter being passed on to the direct and
indirect components.
}
