% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R,
%   R/NMRResonance.R, R/NMRResonance1D.R
\docType{methods}
\name{spread_parameters}
\alias{spread_parameters}
\alias{.spread_parameters,NMRScaffold1D-method}
\title{Copy combined vector to individual parameters}
\usage{
.spread_parameters(object, ...)

\S4method{.spread_parameters}{NMRScaffold1D}(object)

.spread_parameters(object, ...)

\S4method{.spread_parameters}{NMRScaffold1D}(object)
}
\arguments{
\item{object}{NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{object}{NMRScaffold1D or NMRScaffold2D object.}
}
\value{
Modified NMRScaffold object.

Modified NMRScaffold object.
}
\description{
Updates individual parameters based on the combined vector values
(after optimization).

Updates individual parameters based on the combined vector values
(after optimization).
}
