% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\docType{methods}
\name{status}
\alias{status}
\alias{status,NMRFit1D-method}
\title{Access the "status" slot of an NRMScaffold object}
\usage{
status(object, ...)

\S4method{status}{NMRFit1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "status" slot of an NMRScaffold
object.
}
