% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\docType{methods}
\name{status_message}
\alias{status_message}
\alias{status_message,NMRFit1D-method}
\title{Access the "status_message" of an NRMScaffold object}
\usage{
status_message(object, ...)

\S4method{status_message}{NMRFit1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "status_message" of an NMRScaffold
object.
}
