% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmplxn.R
\name{unpack}
\alias{unpack}
\alias{pack}
\title{Pack cmplx1 or cmplx2 column}
\usage{
unpack(x, column.name)

unpack(x, column.name)

pack(x, column.name)
}
\arguments{
\item{x}{data.frame-like object.}

\item{column.name}{character specifying cmplx1 column in x.}

\item{x}{data.frame-like object containing cmplx1 or cmplx2 data.}

\item{column.name}{character specifying cmplx1 or cmplx2 column in x.}
}
\value{
Modified x with new "r" and "i" columns.

export

Modified tidyverse tibble. \code{unpack()} generates new "r" and "i"
        columns for cmplx1 data and "rr", "ri", "ir", "ii" columns for cmplx2
        data. \code{pack()} attempts to find the above columns and generates
        a new cmplx1 or cmplx2 column with the name specified by column.name.
}
\description{
This function undoes unpack()

Despite the conveniences offered by defining complex numbers as vctrs
objects, in some cases, it can still be more convenient to treat
real/imaginary values as separate values of a tibble data frame. One example
is when using the \code{group_by()} function from dplyr, as
\code{bind_rows()} has trouble with vctrs objects. \code{unpack()} and
\code{pack()} function provide a quick method to unpack/pack real/imginary
values into and out of separate columns.
}
