% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\name{unpack_peaks}
\alias{unpack_peaks}
\alias{unpack_peaks,NMRMixture1D-method}
\title{Convert peak parameter vector into tibble}
\usage{
unpack_peaks(object, ...)

\S4method{unpack_peaks}{NMRMixture1D}(
  object,
  par,
  sf = nmroptions$direct$sf,
  shift.range = c(0, 1),
  intensity.range = c(0, 1)
)
}
\arguments{
\item{object}{An NMRMixture1D object.}

\item{par}{Parameter vector.}

\item{sf}{Field strength (used to convert peak width from Hz to ppm).}

\item{shift.range}{A two-element vector representing the ppm range of the fit
(used to scale peak width and position).}

\item{intensity.range}{A two-element vector representing the intensity range
of the fit (used to scale peak heights).}
}
\value{
A tibble with 4 columns -- position, width, height, and
        fraction.gauss.
}
\description{
This is an internal function used to convert the fitted peak vector into a
tibble that can be used to update the peak parameters.
}
