% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R
\name{update_bounds}
\alias{update_bounds}
\alias{update_bounds,NMRScaffold-method}
\title{Update bounds}
\usage{
update_bounds(object, ...)

\S4method{update_bounds}{NMRScaffold}(object, lower.bounds, upper.bounds, widen)
}
\arguments{
\item{object}{An NMRScaffold object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{lower.bounds}{data.frame of lower bounds.}

\item{upper.bounds}{data.frame of upper bounds.}

\item{widen}{FALSE to prevent new bounds from widening existing bounds.}
}
\description{
This is an internal function used to update all the bounds of NMRScaffold object at once while ensuring that they are not widened if that is not desired. Do not use this function directly.
}
