% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R,
%   R/NMRFit1D.R
\docType{methods}
\name{update_peaks}
\alias{update_peaks}
\alias{update_peaks,NMRResonance1D-method}
\alias{update_peaks,NMRSpecies1D-method}
\alias{update_peaks,NMRFit1D-method}
\title{Update NMRResonance1D peak parameters}
\usage{
update_peaks(object, ...)

\S4method{update_peaks}{NMRResonance1D}(object, peaks,
  exclusion.level = nmrsession_1d$exclusion$level,
  exclusion.notification = nmrsession_1d$exclusion$notification)

\S4method{update_peaks}{NMRSpecies1D}(object, peaks,
  exclusion.level = nmrsession_1d$exclusion$level,
  exclusion.notification = nmrsession_1d$exclusion$notification)

\S4method{update_peaks}{NMRFit1D}(object, peaks,
  exclusion.level = nmrsession_1d$exclusion$level,
  exclusion.notification = nmrsession_1d$exclusion$notification)
}
\arguments{
\item{object}{An NMRResonance1D, NMRSpecies1D, or NMRFit1D object.}

\item{peaks}{A data frame with "position", "width", "height", and
"fraction.gauss" columns. Peaks may be defined by one to three
columns of "peak", "resonance", and "species" depending on the
nature of the original object.}

\item{exclusion.level}{A string specifying what to do when peaks are found to
fall outside of the data range: either 'species' to
exclude the whole species to which the offending peak
belongs, 'resonance' to exclude the resonance to which
the offending peak belongs, or 'peak' to exclude just
the peak itself.}

\item{exclusion.notification}{A function specifying how to report when peaks
are found to be outside the data range: 'none'
to ignore, 'message' to issue a message,
'warning' to issue a warning, and 'stop' to
issue an error.}
}
\value{
A new object with modified peak parameters.
}
\description{
This is an internal function whose role is to update existing peak
parameters, while accounting for exclusion criteria and generating relevant
errors/messages. Any peak not updated is excluded.
}
