% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{upper_bounds-set}
\alias{upper_bounds-set}
\alias{upper_bounds<-,NMRScaffold-method}
\alias{upper_bounds<-,NMRScaffold1D-method}
\alias{upper_bounds<-,NMRFit1D-method}
\title{Replace the "upper_bounds" of an NMRScaffold object}
\usage{
upper_bounds(object) <- value

\S4method{upper_bounds}{NMRScaffold}(object) <- value

\S4method{upper_bounds}{NMRScaffold1D}(object) <- value

\S4method{upper_bounds}{NMRFit1D}(object) <- value
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "upper_bounds" of an NMRScaffold
object.
}
