% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R
\name{validNMRScaffold}
\alias{validNMRScaffold}
\title{Generic NRMScaffold validity test}
\usage{
validNMRScaffold(object)
}
\description{
Combination of validity tests that are common to Scaffold1D and Scaffold2,
namely, everything except baseline and phase.
}
