% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R
\name{validNMRScaffold1D}
\alias{validNMRScaffold1D}
\title{NMRScaffold1D validity test}
\usage{
validNMRScaffold1D(object)
}
\description{
Appends baseline, phase, and parameter checks on top of the checks
handled by the NMRScaffold super class
}
