% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R
\name{validNMRScaffoldBaseline}
\alias{validNMRScaffoldBaseline}
\title{Validation function for baseline component of NMRScaffold}
\usage{
validNMRScaffoldBaseline(object, ...)
}
\arguments{
\item{object}{NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\value{
A vector of integers or characters.
}
\description{
Separate baseline validation functions due to the different structures
of 1D and 2D data.
}
