% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R, R/NMRData1D.R, R/NMRData2D.R,
%   R/NMRScaffold1D.R, R/NMRScaffold2D.R
\name{values}
\alias{values}
\alias{values,NMRData1D-method}
\alias{values,NMRData2D-method}
\alias{values,NMRScaffold1D-method}
\alias{values,NMRScaffold2D-method}
\title{Extract or calculate peak lineshape values}
\usage{
values(object, ...)

\S4method{values}{NMRData1D}(object, domain = "r/i", use.cmplx1 = FALSE)

\S4method{values}{NMRData2D}(object, domain = "rr/ri/ir/ii", use.cmplx1 = FALSE)

\S4method{values}{NMRScaffold1D}(
  object,
  direct.shift,
  direct.sf = nmroptions$direct$sf,
  sum.level = "all",
  domain = "r/i",
  use.cmplx1 = FALSE
)

\S4method{values}{NMRScaffold2D}(
  object,
  direct.shift,
  indirect.shift,
  direct.sf = nmroptions$direct$sf,
  indirect.sf = nmroptions$indirect$sf,
  sum.level = "all",
  domain = "rr/ri/ir/ii",
  use.cmplx1 = FALSE
)
}
\arguments{
\item{object}{An NMRScaffold or NMRData object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{domain}{A character specifying which real/imaginary domains to output
for intensity data. For 1D data "r/i" outputs both real and
imaginary data, with "r" and "i" used to output one of either
real or imaginary. For 2D data, this is expanded to
"rr/ri/ir/ii". All domains are outputted by default.}

\item{use.cmplx1}{cmplx1 is the internal representation of complex() that
matches cmplx2 for 2D data. It is replaced by default to
complex() for user-facing functions. However, this behavior
can be overriden here.}

\item{direct.shift}{Vector of chemical shift data in ppm.}

\item{direct.sf}{Sweep frequency (in MHz) -- needed to convert peak widths
from Hz to ppm. In most cases, it is recommended to set a
single default value using nmroptions$direct$sf = ..., but
an override can be provided here.}

\item{sum.level}{The level at which values should be calculated, with "all"
resulting in a single set of values from the sum of all
underlying components. Alternatively, separate values can be
generated at the "mixture", "species", and "resonance"
levels.}

\item{indirect.shift}{Similar to direct.shift, but in the indirect dimension.}

\item{indirect.sf}{Similar to direct.sf, but in the indirect dimension.}
}
\value{
A tibble with "direct.shift", and "intensity" columns, as well as
        "indirect.shift" for 2D data and any combination of "resonance",
        "species", and "mixture" depending on the sum.level parameter and the
        input data.
}
\description{
This is a core function used to convert an abstract NMRScaffold or NMRData
object into a tibble of chemical shifts and intensity values. The method for
this depends on the nature of the object. If the input is NMRData, the
output just extracts internal data for output. However, if the input is an
NMRScaffold object, new lineshape is data based on the chemical shifts and
sweep frequency provided. For 2D scaffolds, direct and indirect chemical
shifts will automatically generate a grid if unique.
}
