% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R
\docType{methods}
\name{volumes}
\alias{volumes}
\alias{volumes,NMRScaffold2D-method}
\title{Calculate peak volume}
\usage{
volumes(object, ...)

\S4method{volumes}{NMRScaffold2D}(object, sum.peaks = TRUE,
  include.id = FALSE)
}
\arguments{
\item{object}{An NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{sum.peaks}{TRUE to add all individual resonances together and output a
single volume, FALSE to output a data frame of peak area
values.}

\item{include.id}{TRUE to include id as outer column if outputting data
frame.}
}
\value{
A single overall volume or a data frame of volumes with resaonces
        identifier and volume columns.
}
\description{
Calculate total peak valumes based on peak parameters. Note that the 2D
representation consists of multiple "peaks", with at least one in the direct
and indirect dimensions. As such, volumes are not calculated per peak, but
per resonance.
}
