% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{which_approx_2d}
\alias{which_approx_2d}
\title{Find the indeces of closest matching values within a matrix}
\usage{
which_approx_2d(x1, x2, y1, y2)
}
\description{
Find the indeces of closest matching values within a matrix
}
