% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R,
%   R/NMRResonance.R, R/NMRResonance1D.R
\docType{methods}
\name{width_columns}
\alias{width_columns}
\alias{.width_columns,NMRScaffold1D-method}
\title{Output peak width related columns of "peaks" data.frame}
\usage{
.width_columns(object, index = FALSE, peak.type = NULL, ...)

\S4method{.width_columns}{NMRScaffold1D}(object, index = FALSE,
  peak.type = NULL)

.width_columns(object, index = FALSE, peak.type = NULL, ...)

\S4method{.width_columns}{NMRScaffold1D}(object, index = FALSE,
  peak.type = NULL)
}
\arguments{
\item{object}{NMRScaffold1D or NMRScaffold2D object.}

\item{index}{TRUE to output column indeces, FALSE to output names}

\item{peak.type}{One of either 'lorenz', 'gauss', 'pvoigt', or 'voigt'}

\item{...}{Additional arguments passed to inheriting methods.}

\item{object}{NMRScaffold1D or NMRScaffold2D object.}

\item{index}{TRUE to output column indeces, FALSE to output names}

\item{peak.type}{One of either 'lorenz', 'gauss', 'pvoigt', or 'voigt'}
}
\value{
A vector of integers or characters.

A vector of integers or characters.
}
\description{
Internal helper function used to generalize peak processing.

Internal helper function used to generalize peak processing.
}
