% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{zero_fill}
\alias{zero_fill}
\title{Zero-fill signal}
\usage{
zero_fill(direct.time, fid, times = 1)
}
\arguments{
\item{direct.time}{A vector of time points of the signal data.}

\item{times}{The number of times to multiply original data length.}

\item{signal}{A vector of complex signal data.}
}
\value{
Corrected signal vector.
}
\description{
Extends the signal by a specified number of time by adding zeros at
the end.
}
