% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R
\docType{methods}
\name{acqus-set}
\alias{acqus-set}
\alias{acqus<-,NMRData-method}
\title{Set acqus parameters}
\usage{
acqus(object) <- value

\S4method{acqus}{NMRData}(object) <- value
}
\description{
Set acqus parameters
}
