% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\docType{methods}
\name{areas}
\alias{areas}
\alias{areas,NMRResonance1D-method}
\title{Calculate peak areas}
\usage{
areas(object, sf = nmrsession_1d("sf"), sum.peaks = TRUE,
  components = "r/i", ...)

\S4method{areas}{NMRResonance1D}(object, sf = nmrsession_1d("sf"),
  sum.peaks = TRUE, components = "r/i")
}
\arguments{
\item{object}{An NMRResonance1D object.}

\item{sf}{Sweep frequency (in MHz) -- needed to convert peak widths from Hz
to ppm. In most cases, it is recommended to set a single default
value using nmrsession_1d(sf = ...), but an override can be
provided here.}

\item{sum.peaks}{TRUE to add all individual peaks together and output a
single area, FALSE to output a data frame of peak area
values.}
}
\value{
A single overall area or a data frame of areas with columns "id",
        "peak", and "area".
}
\description{
Calculate total peak areas based on peak parameters.
}
