% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\docType{methods}
\name{baseline-set}
\alias{baseline-set}
\alias{baseline<-,NMRFit1D-method}
\title{Set object baseline}
\usage{
baseline(object) <- value

\S4method{baseline}{NMRFit1D}(object) <- value
}
\arguments{
\item{object}{An NMRFit1D object.}

\item{value}{A vector of spectra intensity values used to construct the
baseline b-spline. The order of the baseline spline function is
calculated automatically based on the length of the "knots" slot
and the length of the baseline vector, where order =
legnth(baseline) - length(knots).}
}
\description{
Generic convenience method to set the baseline definition of an NMRFit1D
object.
}
