% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\docType{methods}
\name{baseline}
\alias{baseline}
\alias{baseline,NMRFit1D-method}
\title{Get object baseline}
\usage{
baseline(object, ...)

\S4method{baseline}{NMRFit1D}(object)
}
\arguments{
\item{object}{An NMRFit1D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the baseline definition of an
NMRFit1D object.
}
