% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{baseline_difference-set}
\alias{baseline_difference-set}
\alias{baseline_difference<-,NMRScaffold1D-method}
\alias{baseline_difference<-,NMRFit1D-method}
\title{Replace the "baseline_difference" of an NMRScaffold object}
\usage{
baseline_difference(object) <- value

\S4method{baseline_difference}{NMRScaffold1D}(object) <- value

\S4method{baseline_difference}{NMRFit1D}(object) <- value
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "baseline_difference" of an NMRScaffold
object.
}
