% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{baseline_knots-set}
\alias{baseline_knots-set}
\alias{baseline_knots<-,NMRScaffold1D-method}
\alias{baseline_knots<-,NMRFit1D-method}
\title{Replace the "knots" of an NMRScaffold object}
\usage{
baseline_knots(object) <- value

\S4method{baseline_knots}{NMRScaffold1D}(object) <- value

\S4method{baseline_knots}{NMRFit1D}(object) <- value
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "knots" of an NMRScaffold
object.
}
