% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{baseline_knots}
\alias{baseline_knots}
\alias{baseline_knots,NMRScaffold1D-method}
\alias{baseline_knots,NMRFit1D-method}
\title{Access the "knots" slot of an NRMScaffold object}
\usage{
baseline_knots(object, ...)

\S4method{baseline_knots}{NMRScaffold1D}(object)

\S4method{baseline_knots}{NMRFit1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "knots" slot of an NMRScaffold
object.
}
