% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{bounds-set}
\alias{bounds-set}
\alias{bounds<-,NMRScaffold-method}
\alias{bounds<-,NMRScaffold1D-method}
\alias{bounds<-,NMRFit1D-method}
\title{Replace the "bounds" slot of an NMRScaffold object}
\usage{
bounds(object) <- value

\S4method{bounds}{NMRScaffold}(object) <- value

\S4method{bounds}{NMRScaffold1D}(object) <- value

\S4method{bounds}{NMRFit1D}(object) <- value
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "bounds" slot of an NMRScaffold
object.
}
