% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\docType{methods}
\name{bounds-set}
\alias{bounds-set}
\alias{bounds<-,NMRResonance1D-method}
\title{Replace the "bounds" of an NMRResonance1D object}
\usage{
bounds(object) <- value

\S4method{bounds}{NMRResonance1D}(object) <- value
}
\arguments{
\item{object}{An NMRResonance1D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "bounds" of an NMRResonance1D
object.
}
