% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRFit1D.R, R/NMRFit2D.R
\name{bounds}
\alias{bounds}
\alias{lower_bounds}
\alias{upper_bounds}
\alias{bounds,NMRScaffold-method}
\alias{lower_bounds,NMRScaffold-method}
\alias{upper_bounds,NMRScaffold-method}
\alias{bounds,NMRFit1D-method}
\alias{bounds,NMRFit2D-method}
\title{Get object bounds}
\usage{
bounds(object, ...)

lower_bounds(object, ...)

upper_bounds(object, ...)

\S4method{bounds}{NMRScaffold}(object, include.id = FALSE)

\S4method{lower_bounds}{NMRScaffold}(object, include.id = FALSE)

\S4method{upper_bounds}{NMRScaffold}(object, include.id = FALSE)

\S4method{bounds}{NMRFit1D}(object)

\S4method{bounds}{NMRFit2D}(object)
}
\arguments{
\item{object}{An NMRScaffold object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{include.id}{TRUE to return a column of resonance or species ids.}
}
\description{
Generic convenience method to access the bounds of any NMRScaffold object.
If used on anything more complicated than an NMRResonance1D, the function
combines the bounds data frames of component resonances. Whereas
lower_bounds() and upper_bounds() return a data.frame to match peaks(),
bounds() groups the lower and upper bounds into a list.
}
