% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R
\docType{methods}
\name{bounds}
\alias{bounds}
\alias{bounds,NMRResonance1D-method}
\alias{bounds,NMRSpecies1D-method}
\title{Access the "bounds" of an NMRResonance1D object}
\usage{
bounds(object, ...)

\S4method{bounds}{NMRResonance1D}(object)

\S4method{bounds}{NMRSpecies1D}(object)
}
\arguments{
\item{object}{An NMRResonance1D.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "bounds" of an NMRResonance1D
object.
}
