% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R
\docType{methods}
\name{bounds}
\alias{bounds}
\alias{bounds,NMRResonance1D-method}
\alias{bounds,NMRSpecies1D-method}
\title{Get object bounds}
\usage{
bounds(object, include.id = FALSE, ...)

\S4method{bounds}{NMRResonance1D}(object, include.id = FALSE)

\S4method{bounds}{NMRSpecies1D}(object, include.id = FALSE)
}
\arguments{
\item{object}{An NMRResonance1D or NMRSpecies1D object.}

\item{include.id}{TRUE to return a column of resonance or species ids.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the bounds of an NMRResonance1D object.
If used on an NMRSpecies1D object, the function combines the bounds data
frames of component resonances.
}
