% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R,
%   R/NMRResonance.R, R/NMRResonance1D.R
\docType{methods}
\name{bounds}
\alias{bounds}
\alias{bounds,NMRScaffold-method}
\alias{bounds,NMRScaffold1D-method}
\title{Access the "bounds" of an NRMScaffold object}
\usage{
bounds(object, ...)

\S4method{bounds}{NMRScaffold}(object)

\S4method{bounds}{NMRScaffold1D}(object)

bounds(object, ...)

\S4method{bounds}{NMRScaffold}(object)

\S4method{bounds}{NMRScaffold1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "bounds" of an NMRScaffold
object.

Generic convenience method to access the "bounds" of an NMRScaffold
object.
}
