% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\docType{methods}
\name{calc_area}
\alias{calc_area}
\alias{calc_area,NMRScaffold1D-method}
\title{Calculate the areas of each peak}
\usage{
calc_area(object, type = "analytical", ...)

\S4method{calc_area}{NMRScaffold1D}(object, type = "analytical")
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{type}{One of either 'analytical' or 'numerical'. Analytical
calculation is the default, but numerical calculation using
integrate() is left for debugging purposes.}
}
\value{
A data.frame with three columns -- id, peak, and area.
}
\description{
Calculate the analytical area of each peak based on the peak_type.
}
