% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\docType{methods}
\name{calc_lineshape}
\alias{calc_lineshape}
\alias{calc_lineshape,NMRScaffold1D-method}
\title{Calculate lineshape}
\usage{
calc_lineshape(object, ...)

\S4method{calc_lineshape}{NMRScaffold1D}(object, direct.shift = NULL,
  include.convolution = TRUE)
}
\arguments{
\item{object}{An NMRScaffold1D on NMRFit1D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\value{
To DO.
}
\description{
Calculate the lineshape associated with each defined peak.
}
\details{
TO DO.
}
