% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{constraints-set}
\alias{constraints-set}
\alias{constraints<-,NMRScaffold-method}
\alias{constraints<-,NMRScaffold1D-method}
\alias{constraints<-,NMRFit1D-method}
\title{Replace the "constraints" slot of an NMRScaffold object}
\usage{
constraints(object) <- value

\S4method{constraints}{NMRScaffold}(object) <- value

\S4method{constraints}{NMRScaffold1D}(object) <- value

\S4method{constraints}{NMRFit1D}(object) <- value
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "constraints" slot of an NMRScaffold
object.
}
