% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R
\docType{methods}
\name{couplings-set}
\alias{couplings-set}
\alias{couplings<-,NMRResonance1D-method}
\alias{couplings<-,NMRSpecies1D-method}
\title{Replace the "couplings" of an NMRResonance1D object}
\usage{
couplings(object) <- value

\S4method{couplings}{NMRResonance1D}(object) <- value

\S4method{couplings}{NMRSpecies1D}(object) <- value
}
\arguments{
\item{object}{An NMRResonance1D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "couplings" of an NMRResonance1D
object.
}
