% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRScaffold2D.R,
%   R/NMRResonance1D.R, R/NMRSpecies1D.R, R/NMRMixture1D.R
\docType{methods}
\name{couplings}
\alias{couplings}
\alias{couplings,NMRScaffold2D-method}
\alias{couplings,NMRResonance1D-method}
\alias{couplings,NMRSpecies1D-method}
\alias{couplings,NMRMixture1D-method}
\title{Get object couplings}
\usage{
couplings(object, ...)

\S4method{couplings}{NMRScaffold2D}(object, include.id = FALSE)

\S4method{couplings}{NMRResonance1D}(object, include.id = FALSE)

\S4method{couplings}{NMRSpecies1D}(object, include.id = FALSE)

\S4method{couplings}{NMRMixture1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{include.id}{TRUE to return a column of resonance or species ids.}
}
\description{
Generic convenience method to access the coupling definitions of any
NMRScaffold1D object. If used on anything more complicated than an
NMRResonance1D, the function combines the couplings data frames of component
resonances.
}
