% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R
\docType{methods}
\name{couplings}
\alias{couplings}
\alias{couplings,NMRResonance1D-method}
\alias{couplings,NMRSpecies1D-method}
\title{Access the "couplings" of an NMRResonance1D object}
\usage{
couplings(object, ...)

\S4method{couplings}{NMRResonance1D}(object)

\S4method{couplings}{NMRSpecies1D}(object)
}
\arguments{
\item{object}{An NMRResonance1D.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "couplings" of an NMRResonance1D
object.
}
