% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance2D.R
\docType{methods}
\name{direct-set}
\alias{direct-set}
\alias{direct<-,NMRResonance2D-method}
\title{Set direct dimension resonance object}
\usage{
direct(object) <- value

\S4method{direct}{NMRResonance2D}(object) <- value
}
\arguments{
\item{object}{An NMRResonance2D or NMRSpecies2D object.}

\item{value}{A valid NMRResonance1D object.}
}
\description{
Generic convenience method to set the direct dimension resonance component
of an NMRResonance2D or NMRSpecies2D object.
}
