% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R, R/NMRResonance2D.R,
%   R/NMRSpecies2D.R, R/NMRMixture2D.R
\docType{methods}
\name{direct}
\alias{direct}
\alias{direct,NMRResonance2D-method}
\alias{direct,NMRSpecies2D-method}
\alias{direct,NMRMixture2D-method}
\title{Get 1D projection of direct dimension}
\usage{
direct(object)

\S4method{direct}{NMRResonance2D}(object)

\S4method{direct}{NMRSpecies2D}(object)

\S4method{direct}{NMRMixture2D}(object)
}
\arguments{
\item{object}{An NMRScaffold2D object.}
}
\description{
Generic convenience method to extract all direct dimension components.
}
