% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R
\name{direct}
\alias{direct}
\title{Get 1D projection of direct dimension}
\usage{
direct(object)

indirect(object)
}
\arguments{
\item{object}{An NMRScaffold2D object.}
}
\description{
Generic convenience method to extract all direct dimension components.

Generic convenience method to extract all indirect dimension components.
}
