% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R
\docType{methods}
\name{direct}
\alias{direct}
\alias{direct,NMRScaffold2D-method}
\title{Get direct dimension resonance object}
\usage{
direct(object, ...)

\S4method{direct}{NMRScaffold2D}(object)
}
\arguments{
\item{object}{An NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the direct dimension resonance
component of an NMRScaffold2D object.
}
