% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\name{enforce_couplings_1d}
\alias{enforce_couplings_1d}
\title{Enforce coupling relations between peaks.}
\usage{
enforce_couplings_1d(nmrresonance, peaks = NULL)
}
\arguments{
\item{nmrresonance}{An NMRResonance1D object to be modified.}

\item{peaks}{Vector of peak numbers to enforce coupling. The default is to
select all peaks.}
}
\value{
An NMRResonance1D object with a new set of coupling constraints.
}
\description{
Adds a set of constraints between specified peaks that fixes the differences
between peak positions and the ratios between peak widths/heights to
whatever the current values are. These constraints can be relaxed by setting
leeway parameters that convert hard equality constraints to soft inequality
constraints around a fraction of the fixed values.
}
