% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R,
%   R/NMRResonance.R, R/NMRResonance1D.R
\docType{methods}
\name{f_baseline}
\alias{f_baseline}
\alias{f_baseline,NMRScaffold1D-method}
\title{Generate baseline function}
\usage{
f_baseline(object, ...)

\S4method{f_baseline}{NMRScaffold1D}(object)

f_baseline(object, ...)

\S4method{f_baseline}{NMRScaffold1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{object}{An NMRScaffold1D or NMRScaffold2D object.}
}
\value{
TO DO.

TO DO.
}
\description{
This is primarily an internal method that generates a function which
can be used to calculate the baseline at a desired chemical shift.
Any values outside the original fit domain will be set to 0.

This is primarily an internal method that generates a function which
can be used to calculate the baseline at a desired chemical shift.
Any values outside the original fit domain will be set to 0.
}
\details{
TO DO.

TO DO.
}
