% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R,
%   R/NMRResonance.R, R/NMRResonance1D.R
\docType{methods}
\name{f_lineshape}
\alias{f_lineshape}
\alias{f_lineshape,NMRScaffold1D-method}
\title{Generate lineshape function}
\usage{
f_lineshape(object, ...)

\S4method{f_lineshape}{NMRScaffold1D}(object)

f_lineshape(object, ...)

\S4method{f_lineshape}{NMRScaffold1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{object}{An NMRScaffold1D or NMRScaffold2D object.}
}
\value{
TO DO.

TO DO.
}
\description{
This is primarily an internal method that generates a lineshape
function that depends on the object input.

This is primarily an internal method that generates a lineshape
function that depends on the object input.
}
\details{
TO DO.

Warning, a lineshape will be generated whether the parameters are
normalized or not, and the domain/range will vary as a result.

TO DO.

Warning, a lineshape will be generated whether the parameters are
normalized or not, and the domain/range will vary as a result.
}
