% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,NMRFit1D-method}
\title{Fit NMR data to peaks.}
\usage{
fit(object, ...)

\S4method{fit}{NMRFit1D}(object, sf = nmrsession_1d("sf"),
  init = nmrsession_1d$fit$init, opts = nmrsession_1d$fit$opts,
  exclusion.level = nmrsession_1d$exclusion$level,
  exclusion.notification = nmrsession_1d$exclusion$notification)
}
\arguments{
\item{object}{An NMRFit1D object.}

\item{sf}{Sweep frequency (in MHz) -- needed to convert peak widths from Hz
to ppm. In most cases, it is recommended to set a single default
value using nmrsession_1d(sf = ...), but an override can be
provided here.}

\item{init}{An initialization function that takes NMRFit1D object and returns
a modified NMRFit1D object. Use the "identity" function to
override the default initialization in the
nmrsession_1d$fit$init. Note that all arguments provided to the
fit() function are also passed on the init() function.}

\item{opts}{A list of NLOPT fit options to override the default options in
the nmrsession_1d$fit$opts.}

\item{exclusion.level}{A string specifying what to do when peaks are found to
fall outside of the data range: either 'species' to
exclude the whole species to which the offending peak
belongs, 'resonance' to exclude the resonance to which
the offending peak belongs, or 'peak' to exclude just
the peak itself.}

\item{exclusion.notification}{A function specifying how to report when peaks
are found to be outside the data range: 'none'
to ignore, 'message' to issue a message,
'warning' to issue a warning, or 'stop' to
issue an error.}
}
\description{
Given an NMRFit1D object, this function performs a least squares fit on the
data and updates the peak parameters. One of the most important things to
consider is that any peak found outside the data range of the NMRData1D
objects is automatically dropped from the resulting object. Exactly how this
happens can be modified using the exclusion.level and exclusion.notification
arguments. The exclusion.level parameter determines which part of the
overall species to exclude if any of its peaks fall outside the data range:
either 'species' for whole species, 'resonance' for just a subset of the
species and 'peak' to ignore resonance/species blocks and exclude by
specific peak alone. The exclusion.notification parameter determines how to
report when peaks are found to be outside the data range: either 'none' to
give no notice, 'message' to issue a message, 'warning' to issue a warning,
or 'stop' to issue an error.
}
