% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R
\docType{methods}
\name{get_parameter}
\alias{get_parameter}
\alias{get_parameter,NMRData-method}
\title{Access specific parameter}
\usage{
get_parameter(object, ...)

\S4method{get_parameter}{NMRData}(object, parameter, location = "acqus",
  dimension = "direct", error = FALSE)
}
\arguments{
\item{parameter}{Name of parameter.}

\item{location}{One of either "acqus" or "procs".}

\item{dimension}{One of either "direct" or "indirect".}

\item{error}{TRUE to issue an error if parameter not found.}

\item{NMRData}{object.}
}
\description{
Looks for named parameter in one of either acqus or procs before defaulting
to the generic parameters list.
}
