% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRResonance1D.R,
%   R/NMRSpecies1D.R, R/NMRResonance2D.R, R/NMRSpecies2D.R
\name{id-set}
\alias{id-set}
\alias{id<-,NMRResonance1D-method}
\alias{id<-,NMRSpecies1D-method}
\alias{id<-,NMRResonance2D-method}
\alias{id<-,NMRSpecies2D-method}
\title{Set object id}
\usage{
id(object) <- value

\S4method{id}{NMRResonance1D}(object) <- value

\S4method{id}{NMRSpecies1D}(object) <- value

\S4method{id}{NMRResonance2D}(object) <- value

\S4method{id}{NMRResonance2D}(object) <- value

\S4method{id}{NMRSpecies2D}(object) <- value
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{value}{New id (converted to character).}
}
\description{
Generic convenience method to set the id of any NMRScaffold1D object.
}
