% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\docType{methods}
\name{id-set}
\alias{id-set}
\alias{id<-,NMRResonance1D-method}
\title{Replace the "id" of an NMRResonance1D object}
\usage{
id(object) <- value

\S4method{id}{NMRResonance1D}(object) <- value
}
\arguments{
\item{object}{An NMRResonance1D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "id" of an NMRResonance1D
object.
}
