% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R
\docType{methods}
\name{id-set}
\alias{id-set}
\alias{id<-,NMRResonance1D-method}
\alias{id<-,NMRSpecies1D-method}
\title{Set object id}
\usage{
id(object) <- value

\S4method{id}{NMRResonance1D}(object) <- value

\S4method{id}{NMRSpecies1D}(object) <- value
}
\arguments{
\item{object}{An NMRResonance1D or NMRSpecies1D object.}

\item{value}{New id (converted to character).}
}
\description{
Generic convenience method to set the id of an NMRResonance1D or
NMRSpecies1D object.
}
