% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRSpecies2D.R
\name{id}
\alias{id}
\alias{id,NMRScaffold-method}
\alias{id,NMRSpecies2D-method}
\title{Get object id}
\usage{
id(object, ...)

\S4method{id}{NMRScaffold}(object)

\S4method{id}{NMRSpecies2D}(object)
}
\arguments{
\item{object}{An NMRScaffold object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the id of any NMRScaffold object.
}
