% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRResonance2D.R,
%   R/NMRSpecies2D.R
\name{id}
\alias{id}
\alias{id,NMRScaffold1D-method}
\alias{id,NMRResonance2D-method}
\alias{id,NMRSpecies2D-method}
\title{Get object id}
\usage{
id(object, ...)

\S4method{id}{NMRScaffold1D}(object)

\S4method{id}{NMRResonance2D}(object)

\S4method{id}{NMRResonance2D}(object)

\S4method{id}{NMRSpecies2D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the id of any NMRScaffold1D object.
}
