% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R
\docType{methods}
\name{id}
\alias{id}
\alias{id,NMRResonance1D-method}
\alias{id,NMRSpecies1D-method}
\title{Get object id}
\usage{
id(object, ...)

\S4method{id}{NMRResonance1D}(object)

\S4method{id}{NMRSpecies1D}(object)
}
\arguments{
\item{object}{An NMRResonance1D or NMRSpecies1D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the id of an NMRResonance1D or
NMRSpecies1D object.
}
