% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\docType{methods}
\name{id}
\alias{id}
\alias{id,NMRResonance1D-method}
\title{Access the "id" of an NMRResonance1D object}
\usage{
id(object, ...)

\S4method{id}{NMRResonance1D}(object)
}
\arguments{
\item{object}{An NMRResonance1D.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "id" of an NMRResonance1D
object.
}
