% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance2D.R
\docType{methods}
\name{indirect-set}
\alias{indirect-set}
\alias{indirect<-,NMRResonance2D-method}
\title{Set indirect dimension resonance object}
\usage{
indirect(object) <- value

\S4method{indirect}{NMRResonance2D}(object) <- value
}
\arguments{
\item{object}{An NMRResonance2D or NMRSpecies2D object.}

\item{value}{A valid NMRResonance1D object.}
}
\description{
Generic convenience method to set the indirect dimension resonance component
of an NMRResonance2D or NMRSpecies2D object.
}
