% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R
\docType{methods}
\name{indirect-set}
\alias{indirect-set}
\alias{indirect<-,NMRScaffold2D-method}
\title{Set indirect dimension resonance object}
\usage{
indirect(object) <- value

\S4method{indirect}{NMRScaffold2D}(object) <- value
}
\arguments{
\item{object}{An NMRScaffold2D object.}

\item{value}{A valid NMRScaffold1D object.}
}
\description{
Generic convenience method to set the indirect dimension resonance component
of an NMRScaffold2D object.
}
