% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R
\docType{methods}
\name{indirect}
\alias{indirect}
\alias{indirect,NMRScaffold2D-method}
\title{Get indirect dimension resonance object}
\usage{
indirect(object, ...)

\S4method{indirect}{NMRScaffold2D}(object)
}
\arguments{
\item{object}{An NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the indirect dimension resonance
component of an NMRScaffold2D object.
}
