% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData2D.R, R/NMRResonance2D.R,
%   R/NMRSpecies2D.R, R/NMRMixture2D.R
\name{indirect}
\alias{indirect}
\alias{indirect,NMRData2D-method}
\alias{indirect,NMRResonance2D-method}
\alias{indirect,NMRSpecies2D-method}
\alias{indirect,NMRMixture2D-method}
\title{Get 1D projection of indirect dimension}
\usage{
indirect(object)

\S4method{indirect}{NMRData2D}(object)

\S4method{indirect}{NMRResonance2D}(object)

\S4method{indirect}{NMRSpecies2D}(object)

\S4method{indirect}{NMRMixture2D}(object)
}
\arguments{
\item{object}{An NMRData2D or NMRScaffold2D object.}
}
\description{
Warning: this is currently a convenience method to facilitate internal data
manipulation, so the conversion may drop important data. When applied to a
an NMRScaffold2D object, the direct components are simply dropped, and
when applied to an NMRData2D object, the direct components are summed
together to give a 1D projection. No normalization is performed so non-
uniform sampling may result in distorted results.
}
