% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R, R/NMRResonance2D.R
\name{indirect}
\alias{indirect}
\alias{indirect,NMRResonance2D-method}
\title{Get 1D projection of indirect dimension}
\usage{
indirect(object)

\S4method{indirect}{NMRResonance2D}(object)
}
\arguments{
\item{object}{An NMRScaffold2D object.}
}
\description{
Generic convenience method to extract all indirect dimension components.
}
