% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\docType{methods}
\name{initialize_bounds}
\alias{initialize_bounds}
\alias{.initialize_bounds,NMRResonance1D-method}
\title{Initialize empty set of bounds}
\usage{
.initialize_bounds(object, overwrite = FALSE, ...)

\S4method{.initialize_bounds}{NMRResonance1D}(object, overwrite = FALSE)
}
\arguments{
\item{object}{NMRResonance1D object}

\item{overwrite}{TRUE to overwrite existing bounds (to reset them), FALSE to
quietly ignore any existing bounds.}
}
\value{
Modified NMRResonacen1D object.
}
\description{
Initalize lower and upper bounds with the correct dimensions, but set to
-Inf and +Inf respectively
}
