% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R
\name{initialize_bounds}
\alias{initialize_bounds}
\title{Initialize empty set of bounds}
\usage{
.initialize_bounds(object, overwrite, ...)
}
\arguments{
\item{object}{NMRScaffold1D or NMRScaffold2D object.}

\item{overwrite}{TRUE to overwrite existing bounds, FALSE to quietly ignore
any existing bounds.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\value{
Modified NMRScaffold object.
}
\description{
To place bounds on a subset of all the parameters, it is necessary to
initialize lower and upper bound objects that mirror their parent, but
filled with NA values.
}
