% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRScaffold2D.R
\name{initialize_heights}
\alias{initialize_heights}
\alias{initialize_heights,NMRScaffold1D-method}
\alias{initialize_heights,NMRScaffold2D-method}
\title{Initialize peak heights of an NMRScaffold1D object}
\usage{
initialize_heights(object, ...)

\S4method{initialize_heights}{NMRScaffold1D}(object, nmrdata)

\S4method{initialize_heights}{NMRScaffold2D}(object, nmrdata)
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{nmrdata}{An NMRData1D object.}
}
\value{
A new object with modified peak heights.
}
\description{
Generates peak height estimates based on spectral data. At this point, there
is just one approach: take peak height as the intensity of the data at the
current position of the peak. There are plans to develop more sophisticated
approaches in the future.
}
