% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\docType{methods}
\name{initialize_heights}
\alias{initialize_heights}
\alias{initialize_heights,NMRResonance1D-method}
\title{Initialize peak heights of an NMRResonance1D object}
\usage{
initialize_heights(object, ...)

\S4method{initialize_heights}{NMRResonance1D}(object, nmrdata)
}
\arguments{
\item{object}{An NMRResonance1D, NMRSpecies1D, or NMRFit1D object.}

\item{nmrdata}{An NMRData1D object.}
}
\value{
A new object with modified peak heights.
}
\description{
Generates peak height estimates based on spectral data. If applied to an
NMRSpecies1D or NMRFit1D object, the same initialization is propagated to
every component resonance.
}
\details{
At this point, there is just one approach: take peak height as the intensity
of the data at the current position of the peak. There are plans to develop
more sophisticated approaches in the future.
}
