% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R,
%   R/NMRFit1D.R
\docType{methods}
\name{initialize_heights}
\alias{initialize_heights}
\alias{initialize_heights,NMRResonance1D-method}
\alias{initialize_heights,NMRSpecies1D-method}
\alias{initialize_heights,NMRFit1D-method}
\title{Initialize peak heights of an NMRResonance1D object}
\usage{
initialize_heights(object, ...)

\S4method{initialize_heights}{NMRResonance1D}(object, nmrdata,
  exclusion.level = nmrsession_1d$exclusion$level,
  exclusion.notification = nmrsession_1d$exclusion$notification)

\S4method{initialize_heights}{NMRSpecies1D}(object, nmrdata,
  exclusion.level = nmrsession_1d$exclusion$level,
  exclusion.notification = nmrsession_1d$exclusion$notification)

\S4method{initialize_heights}{NMRFit1D}(object, nmrdata,
  exclusion.level = nmrsession_1d$exclusion$level,
  exclusion.notification = nmrsession_1d$exclusion$notification)
}
\arguments{
\item{object}{An NMRResonance1D, NMRSpecies1D, or NMRFit1D object.}

\item{nmrdata}{An NMRData1D object.}

\item{exclusion.level}{A string specifying what to do when peaks are found to
fall outside of the data range: either 'species' to
exclude the whole species to which the offending peak
belongs, 'resonance' to exclude the resonance to which
the offending peak belongs, or 'peak' to exclude just
the peak itself.}

\item{exclusion.notification}{A function specifying how to report when peaks
are found to be outside the data range: 'none'
to ignore, 'message' to issue a message,
'warning' to issue a warning, and 'stop' to
issue an error.}
}
\value{
A new object with modified peak heights.
}
\description{
Generates peak height estimates based on spectral data. If applied to an
NMRSpecies1D or NMRFit1D object, the same initialization is propagated to
every component resonance. Since estimates cannot be provided for any peak
outside the given data range, any such peaks are ignored. Whether or not
warning/error messages are generated when that occurs is specified by the
exclusion.notification parameter. Similarly, exclusion.level provided
options to omit the whole resonance/species if a peak is found to be outside
the data bounds.
}
\details{
At this point, there is just one approach: take peak height as the intensity
of the data at the current position of the peak. There are plans to develop
more sophisticated approaches in the future.
}
