% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{lower_bounds-set}
\alias{lower_bounds-set}
\alias{lower_bounds<-,NMRScaffold-method}
\alias{lower_bounds<-,NMRScaffold1D-method}
\alias{lower_bounds<-,NMRFit1D-method}
\title{Replace the "lower_bounds" of an NMRScaffold object}
\usage{
lower_bounds(object) <- value

\S4method{lower_bounds}{NMRScaffold}(object) <- value

\S4method{lower_bounds}{NMRScaffold1D}(object) <- value

\S4method{lower_bounds}{NMRFit1D}(object) <- value
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "lower_bounds" of an NMRScaffold
object.
}
