% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R,
%   R/NMRResonance.R, R/NMRResonance1D.R
\docType{methods}
\name{merge_parameters}
\alias{merge_parameters}
\alias{.merge_parameters,NMRScaffold1D-method}
\title{Copy individual parameters to combined vector}
\usage{
.merge_parameters(object, return.object = TRUE, ...)

\S4method{.merge_parameters}{NMRScaffold1D}(object, return.object)

.merge_parameters(object, return.object = TRUE, ...)

\S4method{.merge_parameters}{NMRScaffold1D}(object, return.object)
}
\arguments{
\item{object}{NMRScaffold1D or NMRScaffold2D object.}

\item{return.object}{TRUE to return an updated vector, TRUE to return
the combined vector itself.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{object}{NMRScaffold1D or NMRScaffold2D object.}

\item{return.object}{TRUE to return an updated vector, TRUE to return
the combined vector itself.}
}
\value{
Modified NMRScaffold object or numeric vector.

Modified NMRScaffold object or numeric vector.
}
\description{
Updates the combined vector of parameters based on the individual values
(prior to optimization).

Updates the combined vector of parameters based on the individual values
(prior to optimization).
}
