% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{nmrdata-set}
\alias{nmrdata-set}
\alias{nmrdata<-,NMRScaffold-method}
\alias{nmrdata<-,NMRScaffold1D-method}
\alias{nmrdata<-,NMRFit1D-method}
\title{Replace the "nmrdata" slot of an NMRScaffold object}
\usage{
nmrdata(object) <- value

\S4method{nmrdata}{NMRScaffold}(object) <- value

\S4method{nmrdata}{NMRScaffold1D}(object) <- value

\S4method{nmrdata}{NMRFit1D}(object) <- value
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "nmrdata" slot of an NMRScaffold
object.
}
