% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R,
%   R/NMRResonance.R, R/NMRResonance1D.R
\docType{methods}
\name{nmrdata}
\alias{nmrdata}
\alias{nmrdata,NMRScaffold-method}
\alias{nmrdata,NMRScaffold1D-method}
\title{Access the "nmrdata" of an NRMScaffold object}
\usage{
nmrdata(object, ...)

\S4method{nmrdata}{NMRScaffold}(object)

\S4method{nmrdata}{NMRScaffold1D}(object)

nmrdata(object, ...)

\S4method{nmrdata}{NMRScaffold}(object)

\S4method{nmrdata}{NMRScaffold1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "nmrdata" of an NMRScaffold
object.

Generic convenience method to access the "nmrdata" of an NMRScaffold
object.
}
