% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData1D.R
\name{nmrdata_1d_from_scaffold}
\alias{nmrdata_1d_from_scaffold}
\title{Convert NMRScaffold1D to NMRData1D}
\usage{
nmrdata_1d_from_scaffold(
  object,
  direct.shift = NULL,
  direct.sf = nmroptions$direct$sf
)
}
\arguments{
\item{object}{An NMRScaffold1D object}

\item{direct.shift}{Vector of chemical shift data in ppm.}

\item{direct.sf}{Sweep frequency (in MHz) -- needed to convert peak widths
from Hz to ppm. In most cases, it is recommended to set a
single default value using nmroptions$direct$sf = ..., but
an override can be provided here.}
}
\value{
An NMRData1D object.
}
\description{
Generate lineshape from specified peaks.
}
