% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\docType{class}
\name{NMRFit1D-class}
\alias{NMRFit1D-class}
\alias{NMRFit1D}
\title{Definition of an NMR fit.}
\description{
Essentially, this class is used to combine an NMRData1D object with multiple
NMRSpecies1D objects while also defining baseline and phase corrections
terms. There are two primary methods associated with this class: fit() and
update(). fit() takes input data and peak definitions, applies a nonlinear
least squares fit, and outputs best-fit curves into a slot called results.
update() overwrites initial peak definitions with the results of the fit to
generate a new set of initial values for further rounds of fitting.
}
\section{Slots}{

\describe{
\item{\code{species}}{A list of NMRSpecies1D objects.}

\item{\code{nmrdata}}{An NMRData1D object used to fit the peaks.}

\item{\code{knots}}{A vector of chemical shifts corresponds to the internal knots of
the baseline term (two boundary knots are always included).}

\item{\code{baseline}}{A vectors of complex numeric values representing the baseline.
The actual baseline is generated from a basis spline so these
values roughly represent the intensity values of the baseline
based on the chemical shift locations of the knots. Baseline
degree and default number of internal knots can be set using
nmrsession_1d(baseline = list(degree = 3, n.knots = 0)), with
the baseline vector length equal to the degree + n.knots.}

\item{\code{phase}}{A number vectors of phase correction terms, starting from 0 order
and up. The default number of phase terms can be set using
nmrsession_1d(n.phase = 1).}

\item{\code{results}}{A list of results that contains update species list as well as
baseline and phase vectors.}
}}

