% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\docType{class}
\name{NMRFit1D-class}
\alias{NMRFit1D-class}
\alias{NMRFit1D}
\title{Definition of an NMR fit.}
\description{
Essentially, this class is used to combine an NMRData1D object with multiple
NMRSpecies1D objects while also defining baseline and phase correction
terms. There is just one primary method associated with this class: fit().
fit() takes input data and peak definitions, applies a nonlinear least
squares fit, and generates best-fit peak parameters, overwriting the initial
values. Since the fit process is destructive, the nmrfit_1d() function used
to initialize a fit object has an option to delay the fit, allowing pre-fit
and post-fit objects to be saved as different variables.
}
\section{Slots}{

\describe{
\item{\code{species}}{A list of NMRSpecies1D objects.}

\item{\code{nmrdata}}{An NMRData1D object used to fit the peaks.}

\item{\code{knots}}{A vector of chemical shifts corresponds to the internal knots of
the baseline term (two boundary knots are always included).}

\item{\code{baseline}}{A vectors of complex numeric values representing the baseline.
The actual baseline is generated from a basis spline so these
values roughly represent the intensity values of the baseline
based on the chemical shift locations of the knots. Baseline
degree and default number of internal knots can be set using
nmrsession_1d(baseline = list(degree = 3, n.knots = 0)), with
the baseline vector length equal to the degree + n.knots.}

\item{\code{phase}}{A number vectors of phase correction terms, starting from 0 order
and up. The default number of phase terms can be set using
nmrsession_1d(n.phase = 1).}

\item{\code{bounds}}{A lower and upper bounds on baseline and phase terms. Both lower
and upper bounds are lists containing "baseline", and "phase"
elements. A "peaks" element is also generated dynamically from
the species list when using the bounds() getter function. Unlike
peaks, baseline and phase only have a single lower and upper
bound, representing the overall minimum/maximum baseline/phase
value at any point in the spectrum.}
}}

