% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\name{nmrfit_1d}
\alias{nmrfit_1d}
\title{Fit an NMRScaffold1d object to a set of data}
\usage{
nmrfit_1d(object, nmrdata = NULL, normalized = TRUE, bounds = TRUE,
  coupling.leeway = 0, area.leeway = 0, satellite.leeway = 0,
  ordered.peaks = character(), include.convolution = TRUE,
  components = "r/i", opts = NULL)
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{nmrdata}{An NMRData1D object. Optional if the NMRScaffold1D object
has the nmrdata slot set.}

\item{normalized}{TRUE to use a normalized fit. If the parameters or bounds
are not already normalized, they will be normalized to
the data. If FALSE, previously normalized data will be
rescaled before fit.}

\item{bounds}{TRUE to use bounds (setting conservative bounds if none exist),
FALSE to ignore any attached bounds.}

\item{coupling.leeway}{Fraction of possible error in specified coupling
constants. Switching to a non-zero value results
in the use of inequality constraints.}

\item{area.leeway}{Fraction of possible error in specified area ratios of
coupled peaks. Switching to a non-zero value results
in the use of inequality constraints.}

\item{satellite.leeway}{Fraction of possible error in the symmetry of
satellite positions as a function of the distance
between a satellite and the main peak. Switching
to a non-zero value results in the use of 
inequality constraints.}

\item{ordered.peaks}{A vector of peak ids (names) whose order relative to
each other is not allowed to change. This option can
be useful for preventing an arbitrary collection of 
singlets used to fit a more complex peak from drifting
past or on top of each other, leading to ill-defined 
optimization.}

\item{include.convolution}{TRUE to include convolution vector of nmrdata
into fit (if present), FALSE to ignore it.}

\item{components}{String specifying the complex data componets to use for 
the fit (one of either "r/i", "r", or "i").}

\item{opts}{NLOPT option list used to overried default values.}
}
\value{
An NMRFit1D object.
}
\description{
To do.
}
