% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\name{nmrfit_1d}
\alias{nmrfit_1d}
\title{Generate an NMRFit1D object}
\usage{
nmrfit_1d(species, nmrdata,
  baseline.order = nmrsession_1d$baseline$order,
  n.knots = nmrsession_1d$baseline$n.knots,
  phase.order = nmrsession_1d$phase$order, delay.fit = FALSE,
  exclusion.level = "resonance", exclusion.notification = "warning",
  ...)
}
\arguments{
\item{species}{A list of NMRSpecies1D objects or other objects that can be
converted to NMRSpecies1D objects. See ?nmrresonance_1d and
?nmrspecies_1d for more details about this conversion. If list
elements are named, these names will be use to replace
resonance ids.}

\item{nmrdata}{An NMRData1D object used to fit the supplied peaks.
automatically generated from the resonance names.}

\item{baseline.order}{An integer specifying the order of the baseline spline
function. Note the the internal B-spline implementation
requires an order of 1 or greater. You can use an order
of -1 to disable baseline correction.}

\item{n.knots}{An integer specifying the number of internal knots to use. The
specific position of these knots can be modified later using
knots() function. To modify initial values of the baseline,
use baseline() function.}

\item{phase.order}{An integer specifying the order of the phase correction
polynomial. Only 0 and 1st order terms are typically used
in practice, but a higher order correction is possible.}

\item{delay.fit}{FALSE to immediately run least squares optimization after
the NMRFit1D object is initialized, TRUE to skip the
optimization, enabling more customization. The fit can be
run manually using fit().}

\item{exclusion.level}{A string specifying what to do when peaks are found to
fall outside of the data range: either 'species' to
exclude the whole species to which the offending peak
belongs, 'resonance' to exclude the resonance to which
the offending peak belongs, or 'peak' to exclude just
the peak itself. resonance/species blocks and exclude
by specific peak alone.}

\item{exclusion.notification}{A string specifying how to notify the user when
peaks are found to be outside the data range:
either 'none' to give no notice, 'warning' to
issue a warning, and 'error' to issue an error.
A function can also be provided with a single
input, which corresponds to the message text,
which can then be parsed or saved to a log file
as desired.}

\item{...}{Options passed to nmrspecies_1d if conversion has to be performed.
See ?nmrspecies_1d for more details.}
}
\value{
An NMRFit1D object.
}
\description{
Generates an NMRFit1D object based on a list of NMRSpecies1D objects or
other objects that can be converted to NMRSpecies1D objects. See
?nmrresonance_1d and ?nmrspecies_1d for more details about this conversion.
Apart from providing peak definitions (via species argument) and data (via
nmrdata argument), the main fit decisions are related to baseline and phase
correction, as well as how to deal with peaks that are defined outside the
range of the data. This latter decision is broken down into two arguments:
exclusion.level and exclusion.notification. The exclusion.level parameter
determines which part of the overall species to exclude if any of its peaks
fall outside the data range: either 'species' for whole species, 'resonance'
for just a subset of the species and 'peak' to ignore resonance/species
blocks and exclude by specific peak alone. The exclusion.notification
parameter determines how to respond when peaks are found to be outside the
data range: either 'nothing' to give no notice, 'warning' to issue a
warning, and 'error' to issue an error. A function can also be provided with
a single input, which corresponds to the message text, which can then be
parsed or saved to a log file as desired.
}
