% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\docType{class}
\name{NMRIntensity-class}
\alias{NMRIntensity-class}
\alias{NMRIntensity}
\title{Container for NMR intensity data}
\description{
Although 1D intensity can be represented by a single real/imaginary complex
pair, there is no built-in representation for 2D (or higher dimension) data
that may consist of real-real, real-imaginary, imaginary-real, and
imaginary-imaginary components. This class serves as a simple container list
that can be used to retrieve specific sets of values as required. Re() and
Im() can be used to obtain pure real and pure imaginary compenents, whereas
other specific real/imaginary can be are accesible via indexing. For 2D
data, for example, $rr will output a numeric vector corresponding to the
real-real data (equivalent to Re()).
}
\section{Slots}{

\describe{
\item{\code{columns}}{Defines the real/imaginary composition of each elemnt as a
series of "r" and "i" values.}
}}

