% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance2D.R
\docType{class}
\name{NMRResonance2D-class}
\alias{NMRResonance2D-class}
\alias{NMRResonance2D}
\title{Definition of an NMR resonance.}
\description{
An NMRResonance2D object is nothing but a container for 2 NMRResonance1D
objects, one for the direct and one for the indirect dimension. All methods
are merely wrappers applied to both of these objects. To modify an
NMRResonance 1D object associated with a specific dimension, use
\code{direct()} and \code{indirect()} methods.
}
\section{Slots}{

\describe{
\item{\code{dimensions}}{A list containing "direct" and "indirect" elements
corresponding to NMRResonance1D object describing peaks in
the direct and indirect dimensions.}
}}

