% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R
\docType{class}
\name{NMRScaffold-class}
\alias{NMRScaffold-class}
\alias{NMRScaffold}
\title{Super class defining an NMR lineshape.}
\description{
See NMRScaffold1D or NMRScaffold2D.
}
\section{Slots}{

\describe{
\item{\code{peaks}}{A data.frame describing a series of singlets, with one row
per peak. Specific peak parameters depend peak_type, but
all peaks are characterized by a position (in ppm), height
(in relative intensity units), and width (in ppm or Hz).}

\item{\code{baseline}}{A vector of baseline B-spline coefficients, corresponding 
to the B-spline knots.}

\item{\code{baseline_difference}}{A vector of baseline B-spline coefficients, 
corresponding to the difference between real
and imaginary baselines. Although the two
are expected to be the same under most conditions,
this equality may break down in some cases.}

\item{\code{knots}}{A vector of baseline B-spline knots, corresponding 
to the B-spline coefficients.}

\item{\code{phase}}{A vector of phase terms in radians.}

\item{\code{parameters}}{A single vector combining peak, baseline, baseline 
difference, and phase terms. This slot is meant primarily 
for internal use and should generally be avoided.}

\item{\code{constraints}}{A data.frame relating the position and width parameters
of the peaks, effectively combining singlets into
multiplets.}

\item{\code{normalized}}{A logical variable indicating whether parameters have
been normalized with respect to a set of data. Position
and width are influenced by the chemical shift range of
the data, while peak height, baseline, and baseline 
difference are influenced by the maximum intensity of the 
data.}

\item{\code{peak_type}}{The mathematical description of a singlet -- 
one of either 'lorenz', 'gauss', 'pvoigt', or 'voigt'.}

\item{\code{peak_units}}{The units of peak width -- either 'ppm' or 'hz'. Although
Hz units are easier to interpret, the peak fitting
procedure uses ppm.}

\item{\code{lower_bounds}}{An NMRScaffold object that sets lower feasible bounds
on all parameters during peak fitting. Can be set to NULL 
to leave unbounded}

\item{\code{upper_bounds}}{An NMRScaffold object that sets upper feasible bounds
on all parameters during peak fitting. Can be set to NULL 
to leave unbounded}

\item{\code{nmrdata}}{An optional NMRData object that serves as a reference for
for normalization and peak_unit conversion. However, this 
can also be provided to the individual functions as needed.}
}}

