% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold2D.R
\docType{class}
\name{NMRScaffold2D-class}
\alias{NMRScaffold2D-class}
\alias{NMRScaffold2D}
\title{Super-class for all 2D peak descriptions.}
\description{
This class is not meant to be used directly. Instead, it provides a set of
common methods for NMRResonance2D, NMRSpecies2D, NMRMixture2D, and NMRFit2D
objects. Essentially, all of the 2D methods merely wrap around their 1D
counterparts, meaning that the same basic approach can be used for all of
them.
}
