% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{normalized}
\alias{normalized}
\alias{normalized,NMRScaffold-method}
\alias{normalized,NMRScaffold1D-method}
\alias{normalized,NMRFit1D-method}
\title{Access the "normalized" of an NRMScaffold object}
\usage{
normalized(object, ...)

\S4method{normalized}{NMRScaffold}(object)

\S4method{normalized}{NMRScaffold1D}(object)

\S4method{normalized}{NMRFit1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "normalized" of an NMRScaffold
object.
}
