% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R, R/NMRData1D.R
\docType{methods}
\name{parameters-set}
\alias{parameters-set}
\alias{parameters<-,NMRData-method}
\alias{parameters<-,NMRData1D-method}
\title{Replace the "parameters" of an NMRData object}
\usage{
parameters(object) <- value

\S4method{parameters}{NMRData}(object) <- value

\S4method{parameters}{NMRData1D}(object) <- value
}
\arguments{
\item{object}{An NMRData1D or NMRData2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "parameters" of an NMRData
object.
}
