% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R, R/NMRData1D.R
\docType{methods}
\name{parameters}
\alias{parameters}
\alias{parameters,NMRData-method}
\alias{parameters,NMRData1D-method}
\title{Access the "parameters" slot of an NRMData object}
\usage{
parameters(object, ...)

\S4method{parameters}{NMRData}(object)

\S4method{parameters}{NMRData1D}(object)
}
\arguments{
\item{object}{An NMRData1D or NMRData2d object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "parameters" slot of an NMRData
object.
}
