% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\docType{methods}
\name{parse_constraints}
\alias{parse_constraints}
\alias{parse_constraints,NMRMixture1D-method}
\title{Parse constraints for fitting}
\usage{
parse_constraints(object, ...)

\S4method{parse_constraints}{NMRMixture1D}(object, x.span = 1,
  offset = 0)
}
\arguments{
\item{object}{An NMRMixture1D object.}

\item{x.span}{The ppm range of the fit used to scale position differences.}

\item{offset}{An integer offset to the indexes used to handle direct/indirect
dimensions for 2D fitting.}
}
\description{
This is an internal function used to parse equality and inequality
constraints into a form that can be used for lower level fit code.
Essentially, each constraint is defined as a vector with a variable number
of elements codified as follows: 1) Code of either 0 (position), 1 (width),
2 (height), 3 (fraction.gauss), 4 (area); 2) The equality or inequality
target value; 3+) Integers corresponding to peak indexes in the overall
parameter vector. Positive vs negative values are treated differently
depending on the code. For positions, parameters with negative indexes are
subtracted while for width and area, all parameters with negative indexes
are added up before dividing the positive ones.
}
