% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{peak_type}
\alias{peak_type}
\alias{peak_type,NMRScaffold-method}
\alias{peak_type,NMRScaffold1D-method}
\alias{peak_type,NMRFit1D-method}
\title{Access the "peak_type" slot of an NRMScaffold object}
\usage{
peak_type(object, ...)

\S4method{peak_type}{NMRScaffold}(object)

\S4method{peak_type}{NMRScaffold1D}(object)

\S4method{peak_type}{NMRFit1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "peak_type" slot of an NMRScaffold
object.
}
