% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R,
%   R/NMRResonance.R, R/NMRResonance1D.R
\docType{methods}
\name{peak_units}
\alias{peak_units}
\alias{peak_units,NMRScaffold-method}
\alias{peak_units,NMRScaffold1D-method}
\title{Access the "peak_units" of an NRMScaffold object}
\usage{
peak_units(object, ...)

\S4method{peak_units}{NMRScaffold}(object)

\S4method{peak_units}{NMRScaffold1D}(object)

peak_units(object, ...)

\S4method{peak_units}{NMRScaffold}(object)

\S4method{peak_units}{NMRScaffold1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "peak_units" of an NMRScaffold
object.

Generic convenience method to access the "peak_units" of an NMRScaffold
object.
}
