% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R
\name{peaks-set}
\alias{peaks-set}
\alias{peaks<-,NMRScaffold1D-method}
\title{Set object peaks}
\usage{
peaks(object) <- value

\S4method{peaks}{NMRScaffold1D}(object) <- value
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{value}{A data frame with a minimum of "position", "width", "height",
and "fraction.gauss" columns. Peaks may be defined by one to
three columns of "peak", "resonance", and "species" depending on
the nature of the original object.}
}
\description{
Generic convenience method to set the peak definitions of any NMRScaffold1D
object. This is primarily intended as an internal method, so use with
caution. Changing peak definitions after an object has been defined may have
unpredictable consequences.
}
