% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{peaks-set}
\alias{peaks-set}
\alias{peaks<-,NMRScaffold-method}
\alias{peaks<-,NMRScaffold1D-method}
\alias{peaks<-,NMRFit1D-method}
\title{Replace the "peaks" slot of an NMRScaffold object}
\usage{
peaks(object) <- value

\S4method{peaks}{NMRScaffold}(object) <- value

\S4method{peaks}{NMRScaffold1D}(object) <- value

\S4method{peaks}{NMRFit1D}(object) <- value
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "peaks" slot of an NMRScaffold
object.
}
