% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\docType{methods}
\name{peaks-set}
\alias{peaks-set}
\alias{peaks<-,NMRResonance1D-method}
\title{Set object peaks}
\usage{
peaks(object) <- value

\S4method{peaks}{NMRResonance1D}(object) <- value
}
\arguments{
\item{object}{An NMRResonance1D object.}

\item{value}{A data frame with "peak", "position", "width", "height", and
"fraction.gauss" columns.}
}
\description{
Generic convenience method to set the peak definitions of an NMRResonance1D
object.
}
