% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R
\docType{methods}
\name{peaks-set}
\alias{peaks-set}
\alias{peaks<-,NMRResonance1D-method}
\alias{peaks<-,NMRSpecies1D-method}
\title{Replace the "peaks" of an NMRResonance1D object}
\usage{
peaks(object) <- value

\S4method{peaks}{NMRResonance1D}(object) <- value

\S4method{peaks}{NMRSpecies1D}(object) <- value
}
\arguments{
\item{object}{An NMRResonance1D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "peaks" of an NMRResonance1D
object.
}
