% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R,
%   R/NMRFit1D.R
\docType{methods}
\name{peaks-set}
\alias{peaks-set}
\alias{peaks<-,NMRResonance1D-method}
\alias{peaks<-,NMRSpecies1D-method}
\alias{peaks<-,NMRFit1D-method}
\title{Set object peaks}
\usage{
peaks(object) <- value

\S4method{peaks}{NMRResonance1D}(object) <- value

\S4method{peaks}{NMRSpecies1D}(object) <- value

\S4method{peaks}{NMRFit1D}(object) <- value
}
\arguments{
\item{object}{An NMRResonance1D, NMRSpecies1D, or NMRFit1D object.}

\item{value}{A data frame with "position", "width", "height", and
"fraction.gauss" columns. Peaks may be defined by one to three
columns of "peak", "resonance", and "species" depending on the
nature of the original object.}
}
\description{
Generic convenience method to set the peak definitions of an NMRResonance1D
NMRSpecies1D, or NMRFit1D object. This is primarily intended as an internal
method, so use with caution. Changing peak definitions after an object has
been defined may have unpredictable consequences.
}
