% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R
\name{peaks}
\alias{peaks}
\alias{peaks,NMRScaffold-method}
\title{Get object peaks}
\usage{
peaks(object, ...)

\S4method{peaks}{NMRScaffold}(object, include.id = FALSE)
}
\arguments{
\item{object}{An NMRScaffold object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{include.id}{TRUE to return a column of component ids.}
}
\description{
Generic convenience method to access the peak definitions of any
NMRScaffod object. If used on anything more complicated than an
NMRResonance1D, the function combines the peaks data frames of component
resonances.
}
