% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRResonance1D.R,
%   R/NMRSpecies1D.R, R/NMRMixture1D.R, R/NMRResonance2D.R
\name{peaks}
\alias{peaks}
\alias{peaks,NMRResonance1D-method}
\alias{peaks,NMRSpecies1D-method}
\alias{peaks,NMRMixture1D-method}
\alias{peaks,NMRResonance2D-method}
\title{Get object peaks}
\usage{
peaks(object, ...)

\S4method{peaks}{NMRResonance1D}(object, include.id = FALSE)

\S4method{peaks}{NMRSpecies1D}(object, include.id = FALSE)

\S4method{peaks}{NMRMixture1D}(object)

\S4method{peaks}{NMRResonance2D}(object, include.id = FALSE)
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{include.id}{TRUE to return a column of component ids.}
}
\description{
Generic convenience method to access the peak definitions of any
NMRScaffod1D object. If used on anything more complicated than an
NMRResonance1D, the function combines the peaks data frames of component
resonances.
}
