% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R
\docType{methods}
\name{peaks}
\alias{peaks}
\alias{peaks,NMRResonance1D-method}
\title{Access the "peaks" of an NMRResonance1D object}
\usage{
peaks(object, ...)

\S4method{peaks}{NMRResonance1D}(object)
}
\arguments{
\item{object}{An NMRResonance1D.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "peaks" of an NMRResonance1D
object.
}
