% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{peaks}
\alias{peaks}
\alias{peaks,NMRScaffold-method}
\alias{peaks,NMRScaffold1D-method}
\alias{peaks,NMRFit1D-method}
\title{Access the "peaks" of an NRMScaffold object}
\usage{
peaks(object, ...)

\S4method{peaks}{NMRScaffold}(object)

\S4method{peaks}{NMRScaffold1D}(object)

\S4method{peaks}{NMRFit1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "peaks" of an NMRScaffold
object.
}
