% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\docType{methods}
\name{phase-set}
\alias{phase-set}
\alias{phase<-,NMRResonance1D-method}
\title{Set object phase}
\usage{
phase(object) <- value

\S4method{phase}{NMRResonance1D}(object) <- value
}
\arguments{
\item{object}{An NMRFit1D object.}

\item{value}{A vector of phase polynomial values in increasing order (in
radians). So c(pi/4, pi/8) represents a 0 order phase correction
of 45 degrees and a 1st order correction of 22.5 degrees
(referenced to 0 ppm).}
}
\description{
Generic convenience method to set the phase definition of an NMRFit1D
object.
}
