% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{phase-set}
\alias{phase-set}
\alias{phase<-,NMRScaffold-method}
\alias{phase<-,NMRScaffold1D-method}
\alias{phase<-,NMRFit1D-method}
\title{Replace the "phase" of an NMRScaffold object}
\usage{
phase(object) <- value

\S4method{phase}{NMRScaffold}(object) <- value

\S4method{phase}{NMRScaffold1D}(object) <- value

\S4method{phase}{NMRFit1D}(object) <- value
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "phase" of an NMRScaffold
object.
}
