% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R, R/NMRFit1D.R
\docType{methods}
\name{phase}
\alias{phase}
\alias{phase,NMRScaffold-method}
\alias{phase,NMRScaffold1D-method}
\alias{phase,NMRFit1D-method}
\title{Access the "phase" slot of an NRMScaffold object}
\usage{
phase(object, ...)

\S4method{phase}{NMRScaffold}(object)

\S4method{phase}{NMRScaffold1D}(object)

\S4method{phase}{NMRFit1D}(object)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "phase" slot of an NMRScaffold
object.
}
