% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData1D.R
\name{plot.NMRData1D}
\alias{plot.NMRData1D}
\title{Plot NMRData1D object}
\usage{
plot.NMRData1D(x, components = "r", nrows = 2)
}
\arguments{
\item{x}{NMRData1D object.}

\item{components}{One of either 'r', 'i', or 'r/i' to include real,
imaginary or both components. If both components
are selected, they are displayed in separate subplots.}

\item{nrows}{Max number of rows to display subplots.}
}
\description{
Generates an interactive plot object using the plotly package.
}
