% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData1D.R
\name{plot.NMRData1D}
\alias{plot.NMRData1D}
\title{Plot NMRData1D object}
\usage{
plot.NMRData1D(x, components = "r")
}
\arguments{
\item{x}{An NMRData1D object.}

\item{components}{One of either 'r', 'i', or 'r/i' to include real, imaginary
or both components. If both components are selected, they
are displayed in separate subplots.}
}
\value{
A plot_ly plot.
}
\description{
Convenience function that generates a plot of the spectral data.
}
