% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData2D.R
\name{plot.NMRData2D}
\alias{plot.NMRData2D}
\title{Plot NMRData2D object}
\usage{
plot.NMRData2D(x, components = "rr")
}
\arguments{
\item{x}{An NMRData2D object.}

\item{components}{Some combination of 'rr', 'ii', 'ir', and 'ri' in a single
string such as 'rr/ii' to include various real and
imaginary components. If more than one component is
selected, they are displayed in separate subplots.}
}
\value{
A plot_ly plot.
}
\description{
Convenience function that generates a plot of the spectral data.
}
