% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData2D.R
\name{plot.NMRData2D}
\alias{plot.NMRData2D}
\title{Plot NMRData1D object}
\usage{
plot.NMRData2D(x, components = "r")
}
\arguments{
\item{x}{An NMRFit1D object.}

\item{components}{One of either 'r', 'i', or 'r/i' to include real, imaginary
or both components. If both components are selected, they
are displayed in separate subplots.}

\item{sum.level}{One of either 'all', 'species', 'resonance', 'peak' to
specify whether all peaks should be summed together the
peaks should be summed at a lower level.}

\item{sum.baseline}{TRUE to add the baseline to each fit.}

\item{apply.phase}{TRUE to apply the calculated phase to the data.}
}
\value{
A ggplot2 plot.
}
\description{
Convenience function that generates a plot of the spectral data.
}
