% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRFit1D.R
\name{plot.NMRFit1D}
\alias{plot.NMRFit1D}
\title{Plot NMRFit1D object}
\usage{
plot.NMRFit1D(x, components = "r", apply.phase = TRUE,
  sum.lineshapes = TRUE, sum.baseline = TRUE, include.convolution = TRUE,
  deconvolve.residual = FALSE, nrows = 2, legend.position = "bottom",
  fit.legend = NULL)
}
\arguments{
\item{x}{An NMRFit1D object.}

\item{components}{One of either 'r', 'i', or 'r/i' to include real,
imaginary or both components. If both components
are selected, they are displayed in separate subplots.}

\item{apply.phase}{TRUE to apply the calculated phase to the data.}

\item{sum.lineshapes}{TRUE to sum together individual peaks.}

\item{sum.baseline}{TRUE to add the baseline to each fit.}

\item{include.convolution}{TRUE to apply convolution to the fit, if
available.}

\item{deconvolve.residual}{TRUE to remove the effect of convolution
from the residual, if convolution is not being
included in the fit.}

\item{nrows}{Max number of rows to display subplots.}

\item{legend.position}{One of either 'bottom' or 'side'.}

\item{fit.legend}{By default, the fit is shown on the legend if 
sum.lineshapes is TRUE and omitted if sum.lineshapes is
FALSE. Set fit.legend to TRUE or FALSE to override this
behaviour.}
}
\value{
A ggplot2 plot.
}
\description{
Generates an interactive plot object using the plotly package.
}
\details{
Convenience function that generates a graphical representation of the fit.
The original data is plotted as a black line, the fit is plotted in red,
the baseline is plotted in blue, the residual in red. The fit can be plotted
as a composite of all the peaks, or individually.
}
