% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R
\docType{methods}
\name{processed-set}
\alias{processed-set}
\alias{processed<-,NMRData-method}
\title{Set Fourier-transformed data}
\usage{
processed(object) <- value

\S4method{processed}{NMRData}(object) <- value
}
\description{
Set Fourier-transformed data
}
