% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R, R/NMRData1D.R
\docType{methods}
\name{processed}
\alias{processed}
\alias{processed,NMRData-method}
\alias{processed,NMRData1D-method}
\title{Access the "processed" slot of an NRMData object}
\usage{
processed(object, ...)

\S4method{processed}{NMRData}(object)

\S4method{processed}{NMRData1D}(object)
}
\arguments{
\item{object}{An NMRData1D or NMRData2d object.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\description{
Generic convenience method to access the "processed" slot of an NMRData
object.
}
