% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData.R, R/NMRData1D.R
\docType{methods}
\name{procs-set}
\alias{procs-set}
\alias{procs<-,NMRData-method}
\alias{procs<-,NMRData1D-method}
\title{Replace the "procs" of an NMRData object}
\usage{
procs(object) <- value

\S4method{procs}{NMRData}(object) <- value

\S4method{procs}{NMRData1D}(object) <- value
}
\arguments{
\item{object}{An NMRData1D or NMRData2D object.}

\item{value}{Replacement value.}
}
\description{
Generic convenience method to replace the "procs" of an NMRData
object.
}
