% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData1D.R
\name{read_processed_1d}
\alias{read_processed_1d}
\title{Read 1D Bruker 1r/1i files}
\usage{
read_processed_1d(path, procs.list, number = NA)
}
\arguments{
\item{path}{Character string that points to a Bruker scan directory.}

\item{procs.list}{A list of procs parameters that contains 'sw.p', 'si',
'sf', 'reverse', and 'offset' entries. This list can be
generated using read_procs() or through other means.}

\item{number}{The processing file number. Defaults to smallest number in
pdata directory.}
}
\value{
A tidyverse data_frame made of two columns -- "direct.shift" and
        "intensity", corresponding to direct dimension chemical shift and the
        complex spectrum intensity data stored in a cmplx1 vector,
        respectively.
}
\description{
Reads processed bruker 1D files based on specified parameters.
}
