% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRData2D.R
\name{read_processed_2d}
\alias{read_processed_2d}
\title{Read 2D Bruker rr/ri/ir/ii files}
\usage{
read_processed_2d(path, procs.list, number = NA)
}
\arguments{
\item{path}{Character string that points to a Bruker scan directory.}

\item{procs.list}{A list of lists containing procs parameters with 'sw.p',
'si', 'sf', 'reverse', and 'offset'  entries for each of
the 'direct' and 'indirect' sublists. This list can be
generated using read_procs().}

\item{number}{The processing file number. Defaults to the smallest number in
the pdata directory.}
}
\value{
A tidyverse data_frame made up of three columns -- "direct.shift"
        containing the direct dimension chemical shift, "indirect.shift"
        containing the indirect dimension chemical shift, and "intensity"
        containing either a cmplx1 or cmplx2 vector.
}
\description{
Reads processed bruker 2D files based on specified parameters. As the
processed data can vary considerably based on the quadrature method used,
the resulting output will also vary a bit. If all 2rr/2ri/2ir/2ii files are
present, they are stored as a cmplx2 object that encodes all 4 domains. If
only two files are present, they are stored as a cmplx1 object with familiar
r/i domains. If no imaginary data is provided at all, intensity is left as a
cmplx1 object with an imaginary component set to zero.
}
