% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\docType{methods}
\name{set_conservative_bounds}
\alias{set_conservative_bounds}
\alias{set_conservative_bounds,NMRScaffold1D-method}
\title{Set conservative bounds on an NMRScaffold1D or NMRScaffold2D object}
\usage{
set_conservative_bounds(object, nmrdata = NULL, position = NULL,
  height = NULL, width = NULL, ...)

\S4method{set_conservative_bounds}{NMRScaffold1D}(object, nmrdata = NULL,
  position = NULL, height = NULL, width = NULL)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRscaffold2D object.}

\item{nmrdata}{An NMRData object.}

\item{position}{Fraction difference override applied to position parameters.}

\item{height}{Fraction difference override applied to height parameters.}

\item{width}{Fraction difference override applied to width parameters.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\value{
A new NMRScaffold1D or NMRScaffold2D object with modified parameters.
}
\description{
Although any NMRScaffold object can act as a boundary on another
NMRScaffold object, this function provides a convenience method
for generating conservative normalized bounds. Essentially, most parameters
are effectively bound to an approximately 0-1 range following normalization.
}
\details{
Warning, these bounds assume that the chemical shift domain is relatively
small (< 0.1 ppm)
}
