% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R,
%   R/NMRFit1D.R
\docType{methods}
\name{set_conservative_bounds}
\alias{set_conservative_bounds}
\alias{set_conservative_bounds,NMRResonance1D-method}
\alias{set_conservative_bounds,NMRSpecies1D-method}
\alias{set_conservative_bounds,NMRFit1D-method}
\title{Set conservative bounds on an NMRResonance1D object}
\usage{
set_conservative_bounds(object, ...)

\S4method{set_conservative_bounds}{NMRResonance1D}(object,
  position = TRUE, height = TRUE, width = TRUE, nmrdata = NULL,
  widen = FALSE)

\S4method{set_conservative_bounds}{NMRSpecies1D}(object, ...)

\S4method{set_conservative_bounds}{NMRFit1D}(object, ...)
}
\arguments{
\item{object}{An NMRResonance1D, NMRSpecies1D, or NMRFit1D object.}

\item{position}{Without reference data, position is limited to plus or minus
0.1 ppm. With reference data, the position of the peaks is
forced inside the domain of the data. FALSE to disable.}

\item{height}{Without reference data, height is set to strictly positive.
With reference data, height is also limited to no more than 150
percent of the maximum peak value. FALSE to disable.}

\item{width}{Without reference data, minimum peak width is set to almost, but
not quite 0 Hz (1e-3 Hz) and a maximum peak width of 3 Hz. With
reference data, peak width is prevented from being more than 20
percent of the data range.  FALSE to disable.}

\item{nmrdata}{An optional NMRData1D object that can serve as a reference
point for the bounds.}

\item{widen}{FALSE to prevent new bounds from widening existing bounds.}
}
\value{
A new NMRResonance1D object with modified bounds.
}
\description{
A convenience function that sets reasonable bounds on the fit. These bounds
are assumed to be widely applicable to most simple NMR data. Each set of
bounds can be turned on or off as necessary. A slightly better set of bounds
can be selected if a reference NMRData1D object is provided. If applied to
an NMRSpecies1D or NMRFit1D object, the same conservated bounds are
propogated to every component resonance.
}
