% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\docType{methods}
\name{set_conservative_bounds}
\alias{set_conservative_bounds}
\alias{set_conservative_bounds,NMRScaffold1D-method}
\title{Set conservative bounds on an NMRScaffold1D or NMRScaffold2D object}
\usage{
set_conservative_bounds(object, position = TRUE, height = TRUE,
  width = TRUE, baseline = TRUE, phase = TRUE, widen = FALSE, ...)

\S4method{set_conservative_bounds}{NMRScaffold1D}(object,
  position = TRUE, height = TRUE, width = TRUE, baseline = TRUE,
  phase = TRUE, widen = FALSE)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRscaffold2D object with NMR data
attached.}

\item{position}{TRUE to prevent peaks from leaving the peak area and
to keep initial positions to within 20 percent of the 
initial values. FALSE to disable.}

\item{height}{TRUE to prevent negative peaks and limit peak height to
50 percent of the maximum peak value. FALSE to disable.}

\item{width}{TRUE to set a minimum peak width of 0.3 Hz and a maximum peak
width of 3 Hz. FALSE to disable.}

\item{baseline}{TRUE to prevent the baseline from exceeding 25 percent of
maximum peak height in the original data. FALSE to disable.}

\item{phase}{TRUE to limit phase correction within 45 degrees. FALSE to
disable.}

\item{widen}{FALSE to prevent new bounds from widening existing bounds.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\value{
A new NMRScaffold1D or NMRScaffold2D object with modified parameters.
}
\description{
A convenience function that sets reasonable bounds on the fit. These
bounds are assumed to be widely applicable to most simple NMR data.
Each set of bounds can be turned on or off as necessary.
}
