% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R,
%   R/NMRFit1D.R
\docType{methods}
\name{set_peak_type}
\alias{set_peak_type}
\alias{set_peak_type,NMRSpecies1D-method}
\alias{set_peak_type,NMRFit1D-method}
\title{Set peak type of an NMRResonance1D object}
\usage{
set_peak_type(object, ...)

\S4method{set_peak_type}{NMRSpecies1D}(object, ...)

\S4method{set_peak_type}{NMRFit1D}(object, ...)
}
\arguments{
\item{object}{An NMRResonance1D, NMRSpecies1D, or NMRFit1D object.}

\item{peak.type}{One of either "lorentz", "voigt", "gauss", or "any" where
"any" clears existing bounds on the fraction.gauss
parameter.}
}
\value{
A new object with modified bounds and peak parameters.
}
\description{
The peak type of an NMRResonance1D object is governed by the fraction.gauss
parameter and can fluidly go from pure Lorentz to Gauss via the combined
Voigt lineshape. However, it can be computationally efficient to force a
specific peak type. This function provided a shortcut for doing so by
changing the fraction.gauss parameter as well as lower and upper bounds.
}
