% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R,
%   R/NMRResonance.R, R/NMRResonance1D.R
\docType{methods}
\name{set_peak_type}
\alias{set_peak_type}
\alias{set_peak_type,NMRScaffold1D-method}
\title{Convert an NMRScaffold1D or NMRScaffold2D object to a different peak.type}
\usage{
set_peak_type(object, peak.type, frac.lorenz = 0.9, ...)

\S4method{set_peak_type}{NMRScaffold1D}(object, peak.type,
  frac.lorenz = 0.9)

set_peak_type(object, peak.type, frac.lorenz = 0.9, ...)

\S4method{set_peak_type}{NMRScaffold1D}(object, peak.type,
  frac.lorenz = 0.9)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{peak.type}{Target peak.type -- one of lorenz, gauss, pvoigt, or voigt.}

\item{frac.lorenz}{If converting from lorenz or gauss to pvoigt or voigt,
the fraction of the output peak that is expected to be
lorenz. Ignored for the reverse conversion.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{peak.type}{Target peak.type -- one of lorenz, gauss, pvoigt, or voigt.}

\item{frac.lorenz}{If converting from lorenz or gauss to pvoigt or voigt,
the fraction of the output peak that is expected to be
lorenz. Ignored for the reverse conversion.}
}
\value{
An NMRScaffold1D or NMRScaffold2D object with the new peak_type and
        modified parameters.

An NMRScaffold1D or NMRScaffold2D object with the new peak_type and
        modified parameters.
}
\description{
As the various peak representations have different parameters, Converting
between different peak types is not always trivial. This function ensures
that conversions are reversible and path independent (switching directly
from lorenz to voigt will be the same as switching from lorenz to gauss
to voigt). Each conversion preserves the area of the peak and modifies
either the width or height. Switching from pvoigt to lorenz or gauss
preserves overall peak height, while switching from voigt to lorenz or
gauss preserves overall peak width. To go in the reverse direction,
it's necessary to specify what fraction of the original height or width
should remain attributed to the lorenz line shape.

As the various peak representations have different parameters, Converting
between different peak types is not always trivial. This function ensures
that conversions are reversible and path independent (switching directly
from lorenz to voigt will be the same as switching from lorenz to gauss
to voigt). Each conversion preserves the area of the peak and modifies
either the width or height. Switching from pvoigt to lorenz or gauss
preserves overall peak height, while switching from voigt to lorenz or
gauss preserves overall peak width. To go in the reverse direction,
it's necessary to specify what fraction of the original height or width
should remain attributed to the lorenz line shape.
}
