% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRScaffold1D.R
\docType{methods}
\name{set_relative_bounds}
\alias{set_relative_bounds}
\alias{set_relative_bounds,NMRScaffold1D-method}
\title{Set relative bounds on an NMRScaffold1D or NMRScaffold2D object}
\usage{
set_relative_bounds(object, position = NULL, height = NULL,
  width = NULL, baseline = NULL, phase = NULL, normalized = FALSE,
  widen = FALSE, ...)

\S4method{set_relative_bounds}{NMRScaffold1D}(object, position = NULL,
  height = NULL, width = NULL, baseline = NULL, phase = NULL,
  normalized = TRUE, widen = FALSE)
}
\arguments{
\item{object}{An NMRScaffold1D or NMRscaffold2D object. NMR data must
be attached to object to enable normalization.}

\item{position}{A vector of two elements corresponding to a lower 
and upper bound for peak position, where the value
is taken as a fraction of the current position.
Note that whether or not the normalized argument
is set to TRUE has a big impact. If normalized is FALSE,
the fraction is applied to the peak position in ppm
(where even a small fraction has a very big impact). If
normalized is TRUE, the fraction is effectively applied
to a subset of the overall ppm range.}

\item{height}{A vector of two elements corresponding to a lower
and upper bound for peak height, where the value
  is taken as a fraction of the current height.}

\item{width}{A vector of two elements corresponding to a lower
and upper bound for peak width, where the value
   is taken as a fraction of the current width.}

\item{baseline}{A vector of two elements corresponding to a lower 
and upper bound for baseline height, where the value
is taken as a fraction of the current baseline height.}

\item{phase}{A vector of two elements corresponding to a lower and upper 
bound for phase in radians, where the value
   is taken as a fraction of the current phase.}

\item{normalized}{TRUE to set bounds in terms of the underlying data, where
the x and y values of the data are scaled between 0 and 1.
FALSE to use natural units of ppm/Hz and spectral
intensity.}

\item{widen}{FALSE to prevent new bounds from widening existing bounds.}

\item{...}{Additional arguments passed to inheriting methods.}
}
\value{
A new NMRScaffold1D or NMRScaffold2D object with modified parameters.
}
\description{
Although any NMRScaffold object can act as a boundary on another
NMRScaffold object, this function provides a convenience method
for generating such bounds using a simple set of lower and upper
constraints on basic peak parameters such as position, height,
width, etc... Note that the term "relative" refers to the fact
that bounds are applied relative to the current values of the
parameters.
}
\details{
In practice, relative bounds are primarily useful for preventing 
peak positions from drifting and for fine-tuning a fit once an initial
optimization is performed. It is not recommended to use strict relative
bounds based on rough initial parameter guesses.
}
