% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRIntensity.R
\docType{methods}
\name{show,NMRIntensities-method}
\alias{show,NMRIntensities-method}
\title{Show an NMRIntensities object}
\usage{
\S4method{show}{NMRIntensities}(object)
}
\description{
Show an NMRIntensities object
}
