% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold.R, R/NMRResonance.R
\name{validNMRScaffoldPhase}
\alias{validNMRScaffoldPhase}
\title{Validation function for phase component of NMRScaffold}
\usage{
validNMRScaffoldPhase(object, ...)

validNMRScaffoldPhase(object, ...)
}
\arguments{
\item{object}{NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{object}{NMRScaffold1D or NMRScaffold2D object.}
}
\value{
A vector of integers or characters.

A vector of integers or characters.
}
\description{
Separate phase validation functions due to the different structures
of 1D and 2D data.

Separate phase validation functions due to the different structures
of 1D and 2D data.
}
