% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRScaffold2D.R
\docType{methods}
\name{values}
\alias{values}
\alias{values,NMRScaffold1D-method}
\alias{values,NMRScaffold2D-method}
\title{Calculate peak lineshape values}
\usage{
values(object, ...)

\S4method{values}{NMRScaffold1D}(object, direct.shift,
  sf = nmroptions$direct$sf, sum.peaks = TRUE, sum.baseline = FALSE,
  include.id = FALSE, components = "r/i")

\S4method{values}{NMRScaffold2D}(object, direct.shift, indirect.shift,
  sf = c(nmroptions$direct$sf, nmroptions$indirect$sf),
  sum.peaks = TRUE, sum.baseline = FALSE, include.id = FALSE,
  components = "rr/ii")
}
\arguments{
\item{object}{An NMRScaffold1D or NMRScaffold2D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{direct.shift}{Vector of chemical shift data in ppm.}

\item{sf}{Sweep frequency (in MHz) -- needed to convert peak widths from Hz
to ppm. A single value can be used for 1D and homonuclear 2D
experiments or two values for the direct and indirect dimensions,
in that order. this can be  In most cases, it is recommended to set
a single default value using, for example, nmroptions$direct$sf =
..., but an override can be provided here.}

\item{sum.peaks}{TRUE to add all individual peaks together and output a
single set of values, FALSE to output a data frame of values
that correspond to individual peaks.}

\item{sum.baseline}{TRUE to add baseline to every peak, if one is defined.
FALSE to exclude baseline.}

\item{include.id}{TRUE to include id as outer column if outputting data
frame.}

\item{components}{A string specifying the real/imaginary components to
output, e.g. 'r/i' or 'rr/ri/ir/ii'. Use the symbols 'r',
'i' for 1D experiments and 'rr', 'ri', 'ir', 'ii' for 2D
experiments. The symbols can be separated by spaces or
practically any other characters, so 'r i' is interpreted
the same as 'r/i'. The total number of components will
determine the final data type of the output, with 1
component corresponding to real values, 2 components to
cmplx1d objects (similar to the normal complex type), and 3
or 4 components to cmplx2d hypercomplex objects.}

\item{indirect.shift}{Vector of chemical shift data in ppm.}
}
\value{
A vector of spectral intensity data or a tibble with peak
        identifiers.
}
\description{
Calculated peak intensity values over a set of chemical shifts.
}
