% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRScaffold1D.R, R/NMRResonance2D.R
\docType{methods}
\name{values}
\alias{values}
\alias{values,NMRScaffold1D-method}
\alias{values,NMRResonance2D-method}
\title{Calculate peak lineshape values}
\usage{
values(object, ...)

\S4method{values}{NMRScaffold1D}(object, direct.shift,
  sf = nmroptions$direct$sf, sum.peaks = TRUE, sum.baseline = FALSE,
  include.id = FALSE, components = "r/i")

\S4method{values}{NMRResonance2D}(object, direct.shift, indirect.shift,
  sf = c(nmroptions$direct$sf, nmroptions$indirect$sf),
  sum.peaks = TRUE, sum.baseline = FALSE, include.id = FALSE,
  components = "rr/ii")
}
\arguments{
\item{object}{An NMRScaffold1D object.}

\item{...}{Additional arguments passed to inheriting methods.}

\item{direct.shift}{Vector of chemical shift data in ppm.}

\item{sf}{Sweep frequency (in MHz) -- needed to convert peak widths from Hz
to ppm. In most cases, it is recommended to set a single default
value using nmroptions$direct$sf = ..., but an override can be
provided here.}

\item{sum.peaks}{TRUE to add all individual peaks together and output a
single set of values, FALSE to output a data frame of values
that correspond to individual peaks.}

\item{sum.baseline}{TRUE to add baseline to every peak, if one is defined.
FALSE to exclude baseline. that correspond to individual
peaks.}

\item{include.id}{TRUE to include id as outer column if outputting data
frame.}

\item{components}{'r/i' to output both real and imaginary data, 'r' to output
only real and 'i' to output only imaginary.}
}
\value{
A vector of spectral intensity data or a tibble with columns
        "resonance" (optional), "peak", "direct.shift", and "intensity".
}
\description{
Calculated peak intensity values over a set of chemical shifts.
}
