% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMRResonance1D.R, R/NMRSpecies1D.R
\docType{methods}
\name{values}
\alias{values}
\alias{values,NMRResonance1D-method}
\alias{values,NMRSpecies1D-method}
\title{Calculate peak lineshape values}
\usage{
values(object, direct.shift, sf = nmrsession_1d("sf"),
  sum.peaks = TRUE, include.id, components = "r/i", ...)

\S4method{values}{NMRResonance1D}(object, direct.shift,
  sf = nmrsession_1d("sf"), sum.peaks = TRUE, components = "r/i")

\S4method{values}{NMRSpecies1D}(object, direct.shift,
  sf = nmrsession_1d("sf"), sum.peaks = TRUE, components = "r/i")
}
\arguments{
\item{object}{An NMRResonance1D or an NMRSpecies1D object.}

\item{direct.shift}{Vector of chemical shift data in ppm.}

\item{sf}{Sweep frequency (in MHz) -- needed to convert peak widths from Hz
to ppm. In most cases, it is recommended to set a single default
value using nmrsession_1d(sf = ...), but an override can be
provided here.}

\item{sum.peaks}{TRUE to add all individual peaks together and output a
single set of values, FALSE to output a data frame of values
that correspond to individual peaks.}

\item{include.id}{TRUE to include id as "resonance" column if outputting data
frame.}

\item{components}{'r/i' to output both real and imaginary data, 'r' to output
only real and 'i' to output only imaginary.}
}
\value{
A vector of spectral intensity data or a data frame with columns
        "resonance" (optional), "peak", "direct.shift", and "intensity".
}
\description{
Calculated peak intensity values over a set of chemical shifts.
}
